% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{AbusedMilk}
\alias{AbusedMilk}
\title{Bacterial counts in abused milk}
\format{
A dataframe of 80 observations on four variables.
\describe{
\item{\code{sample}}{A factor indicating the sample used.}
\item{\code{abuse}}{A factor indicating the abuse level (time) used.}
\item{\code{lab}}{A factor indicating the lab making the count.}
\item{\code{count}}{The count of bacteria per ul.}
}
}
\usage{
AbusedMilk
}
\description{
Eight 1-gallon containers of raw milk are obtained from a dairy
and are assigned at random to four abuse treatments, two containers
per treatment.  Abuse consists of keeping the milk at 25C for
a period of time; the four abuse treatments are
four randomly selected durations between 1 and 18 hours.
After abuse, each gallon is split into five equal portions and frozen.

We have selected five contract laboratories at random from those
available in the state.  For each gallon, the five portions are
randomly assigned to the five laboratories.  The eight portions
for a given laboratory are then placed in an insulated shipping
container cooled with dry ice and shipped.  Each
laboratory is asked to provide duplicate counts of bacteria in
each milk portion.  Response is bacteria counts per ul

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
