% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{AirCells}
\alias{AirCells}
\title{Area of air cells in expanded/extruded wheat}
\format{
A dataframe of 30 observations on five variables.
\describe{
\item{\code{var}}{Factor indicating the variety of wheat.}
\item{\code{loc}}{Factor indicating the location where the wheat was grown.}
\item{\code{temp.z}}{The temperature of the extrusion.}
\item{\code{temp}}{A factor version of the temperature.}
\item{\code{area}}{The area of the air cells.}
}
}
\usage{
AirCells
}
\description{
Expanded/extruded wheat flours (think breakfast cereals or
cheese puffs) have air cells that vary
in size, and the size may depend on the variety of wheat
used to make the flour, the location where the wheat was
grown, and the temperature at which the flour was extruded.
An experiment has been conducted to assess these
factors.  The first factor is the variety of wheat used
(Butte 86, 2371,  or Grandin).  The second factor is the
growth location (MN or ND).  The third factor is the
temperature of the extrusion (120 C or 180 C).  The response
is the area in mm^2 of the air cells

Source: Oehlert (2022), originally from Sutheerawattananonda (1994).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Sutheerawattananonda, M. (1994) "Variation in physical properties and
microstructure of extruded wheat flours," Master's thesis, University of
Minnesota, St. Paul, MN.
}
\keyword{datasets}
