% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Albizia}
\alias{Albizia}
\title{Leaf bending angles}
\format{
A dataframe of 15 observations on three variables.
\describe{
\item{\code{delay.z}}{Delay before measuring, in minutes.}
\item{\code{angle}}{Angle of the leaves, in degrees.}
\item{\code{delay}}{A factor version of \code{delay.z}.}
}
}
\usage{
Albizia
}
\description{
Leaves of plants in the genus Albizia will fold and unfold
depending on light conditions. Fifteen different leaves were exposed to
red light for three minutes. The leaves were randomly assigned to one of
three delays (30, 45, or 60 minutes). After the delay, the angle of the
leaves was measured (in degrees).

Source: Oehlert (2022), originally from W. Hughes (pers com).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
