% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{AmylaseActivity}
\alias{AmylaseActivity}
\title{Amylase in sprouting maize}
\format{
A dataframe of 96 observations on six variables.
\describe{
\item{\code{aTemp.z}}{The analysis temperature.}
\item{\code{aTemp}}{A factor version of the analysis temperature.}
\item{\code{gTemp.z}}{The growth temperature.}
\item{\code{gTemp}}{A factor version of the growth temperature.}
\item{\code{variety}}{The variety of the maize.}
\item{\code{amylase}}{The amylase activity.}
}
}
\usage{
AmylaseActivity
}
\description{
Two varieties of sprouting maize (B73 or Oh43) are grown under one of two
different temperature (25 or 13 degrees C) and are analyzed for
amylase activity under one of eight different temperatures
(40, 35, 30, 25, 20, 15, 13, and 10 degrees C). Ninety-six
different runs are randomized to these 32 combinations, three runs
per combination. The response is the amylase activity in
International Units.

Source: Oehlert (2022), originally from Orman (1986).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Orman, B. A. (1986) "Maize germination and seedling growth at suboptimal temperatures,"
Master's Thesis, University of Minnesota, St. Paul.
}
\keyword{datasets}
