% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Antibiotics}
\alias{Antibiotics}
\title{Cells Surviving after Antibiotics}
\format{
A data frame for 27 observations on six variables.
\describe{
\item{\code{plate}}{A factor indicating the plate (replication).}
\item{\code{Hygro}}{A factor indicating the level of Hygromycin.}
\item{\code{Hygro.z}}{A numeric variable indicating the level of Hygromycin (ug/mL).}
\item{\code{Puro}}{A factor indicating the level of Puromycin.}
\item{\code{Puro.z}}{A numeric variable indicating the level of Puroomycin (ug/mL).}
\item{\code{Density}}{The number of live cells (cells/mL).}
}
}
\usage{
Antibiotics
}
\description{
Nine wells on a plate are seeded with HEK293 cells
and then randomly assigned to the factor/level combinations of puromycin at 
	0, 2, and 4 ug/mL (suggested concentration is 2) and hygromycin at 0, 150,
and 300 ug/mL (suggested concentration is 200). The plate is incubated and
the surviving cells are counted. This is repeated two more times on
two more plates.

Source: Oehlert (2022), from Min Lu, pers. comm.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
