% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Anxiety}
\alias{Anxiety}
\title{Memory errors under anxiety and tension}
\format{
A dataframe of 48 observations on five variables.
\describe{
\item{\code{anxiety}}{A factor indicating the anxiety level.}
\item{\code{tension}}{A factor indicating the tension level.}
\item{\code{type}}{A factor indicating the memory type.}
\item{\code{subject}}{A factor indicating the subject (labelled 1-12).}
\item{\code{errors}}{The number of memory errors.}
}
}
\usage{
Anxiety
}
\description{
Twelve subjects are assigned to one of four anxiety-tension combinations
at random.  The low-anxiety group is told that they will be awarded $5 for
participation and $10 if they remember sufficiently accurately, and the
high-anxiety group is told that they will be awarded $5 for participation
and $100 if they remember sufficiently accurately.  Everyone must squeeze
a spring-loaded grip to keep a buzzer from sounding during the testing
period.  The high-tension group
must squeeze against a stronger spring than the low-tension group.  All
subjects then perform four memory trials in random order,
testing four different types of memory.
The response is the number of errors on each memory trial.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
