% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Autoclaving}
\alias{Autoclaving}
\title{Elastic modulus of latex tubing after autoclaving}
\format{
A dataframe of 15 observations on two variables.
\describe{
\item{\code{treatment}}{A factor indicating the treatment used.}
\item{\code{modulus}}{The modulus of elasticity (kPa).}
}
}
\usage{
Autoclaving
}
\description{
This experiment examines the effects of autoclaving
on the elastic modulus (in kPa) of the latex.  The adjustable
factors for the autoclave are the temperature (121 C or 135 C)
and the time (10 minutes or 20 minutes).  We look at five
treatments:
Control (no autoclaving), low temperature/low time, low temperature/high time,
high temperature/low time, and high temperature/high time. The
response is the modulus of elasticity.

Source: Oehlert (2022), data from Z. Syedain (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
