% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BankExaminers}
\alias{BankExaminers}
\title{Scores on bank examiner exams.}
\format{
A dataframe of 50 observations on four variables.
\describe{
\item{\code{student}}{Factor indicating the student taking the exam.}
\item{\code{grader}}{Factor indicating the grader of the exam.}
\item{\code{exam}}{Factor indicating the exam form (the workbasket).}
\item{\code{score}}{The grade the student was awarded for the exam.}
}
}
\usage{
BankExaminers
}
\description{
Part of the
certification process for bank examiners involves a "work basket" of
tasks that the examinee must complete in a satisfactory fashion in a
fixed time period.  New work baskets must be constructed for each round
of examinations, and much effort is expended to make the workbaskets
comparable (in terms of average score) from exam to exam.  This year,
two new work baskets (A and B) are being evaluated.  We have three old
work baskets (C, D, and E) to form a basis for comparison.  We have ten
paid examinees (1 through 6 are certified bank examiners, 7 through 9
are noncertified bank examiners nearing the end of their training, and 10
is a public accountant with no bank examining experience or training) who
will each take all five tests.
There are five graders who will each grade ten exams.
Randomization was done so that each
examinee took each exam and each grader grades two of each exam.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
