% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BasicSkills}
\alias{BasicSkills}
\title{Grading the writing basic skills test.}
\format{
A dataframe of 150 observations on three variables.
\describe{
\item{\code{exam}}{Factor indicating the exam being graded.}
\item{\code{grader}}{Factor indicating the grader of the exam.}
\item{\code{score}}{The grade the student was awarded for the exam.}
}
}
\usage{
BasicSkills
}
\description{
The State Board of Education has adopted basic skills tests
for high school graduation; one of these is a writing test.
The student writing samples are graded by professional graders,
and the board is taking some care to be sure that the graders
are grading to the same standard.  We examine grader differences
with the following experiment.  There are 25 graders.  We
select 30 writing samples at random; each writing sample will be
graded by five graders.  Thus each grader will grade six samples,
and each pair of graders will have a test in common. The response
is the grade, or score, given each exam by each grader.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
