% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Beer}
\alias{Beer}
\title{Beer retained in mouth after expectoration.}
\format{
A dataframe of 40 observations on five variables.
\describe{
\item{\code{ml.z}}{The ml of the sample.}
\item{\code{ml}}{A factor version of ml.z.}
\item{\code{subject}}{A factor indicating the subject.}
\item{\code{day}}{A factor indicating the day.}
\item{\code{retained}}{The amount of beer retained in the mouth (in ml).}
}
}
\usage{
Beer
}
\description{
As part of a larger experiment, researchers are looking
at the amount of beer that remains in the mouth after expectoration.
Ten subjects will repeat the experiment on two separate days.
Each subject will place 10 ml or 20 ml of beer in his or her mouth
for five seconds, and then expectorate the beer.  The beer has
a dye, so the amount of expectorated beer can be determined,
and thus the amount of beer retained in the mouth (in ml).

Source: Oehlert (2022), originally from Brefort, Buinard, and Lewis (1989).
}
\references{
Brefort, H., Guinard, J. X., and Lewis, M. J. (1989). "The contribution of dextrins
of beer sensory properties, part II Aftertaste." \emph{Journal of the
Institute of Brewing} \strong{95}, 431-435.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
