% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BigSagebrush}
\alias{BigSagebrush}
\title{Viability of big sagebrush seeds}
\format{
A dataframe of 63 observations on six variables.
\describe{
\item{\code{storage.z}}{The days of storage.}
\item{\code{storage}}{A factor version of storage.}
\item{\code{humidity.z}}{The relative humidity of the storage.}
\item{\code{humidity}}{A factor version of humidity.}
\item{\code{viability}}{The percentage of viable seeds.}
\item{\code{viable}}{The number of viable seeds (out of 300).}
}
}
\usage{
BigSagebrush
}
\description{
Big sagebrush is often planted in range restoration
projects.  An experiment is performed to determine
the effects of storage length and relative humidity
on the viability of seeds.  Sixty-three batches of
300 seeds each are randomly divided into 21 groups
of three.  These 21 groups each receive a different
treatment, namely the combinations of storage
length (0, 60, 120, 180, 240, 300, or 360 days) and
storage relative humidity (0, 32, or 45\%).  After
the storage time, the seeds are planted, and the
response is the percentage of seeds that sprout

Source: Oehlert (2022), originally from Welch (1996).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Welch, B. (1996) "Effects of humidity on storing big sagebrush seed." Research Paper
INT-RP-493, USDA Forest Service, Intermountain Research Station.
}
\keyword{datasets}
