% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BioequivalenceFull}
\alias{BioequivalenceFull}
\title{Bioequivalence of drugs (complete data set)}
\format{
A dataframe of 36 observations on four variables.
\describe{
\item{\code{period}}{A factor giving the time period for the observation.}
\item{\code{subject}}{A factor giving the subject for the observation.}
\item{\code{treatment}}{A factor giving the treatment for the observation.}
 \item{\code{area}}{The area under the time-concentration curve.}
}
}
\usage{
BioequivalenceFull
}
\description{
Bioequivalence studies
seek to determine if
different drug delivery systems have similar biological
effects.  One variable to compare is the area under
the time-concentration curve.  This area is proportional to
the average concentration of the drug over time.

This experiment compares three methods for delivering a drug:
a solution, a tablet, and a capsule.  The response is
the area under the time-concentration curve.  We anticipate
Large subject to subject differences were expected, so subject
was used as a block.
Each subject will be given the drug
three times, once with each of the three methods.  Because
the body may adapt to the drug in some way, each drug will
be used once in the first period, once in the second
period, and once in the third period.

Source: Oehlert (2022), originally from Selwyn and Hall (1984).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Selwyn, M. R. and N. R. Hall (1984). "On Bayesian methods for bioequivalence."
\emph{Biometrics} \strong{40}, 1103-1108.
}
\keyword{datasets}
