% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BirchSeedlings}
\alias{BirchSeedlings}
\title{Weights of birch seedlings exposed to acid rain}
\format{
A dataframe of 235 observations on three variables.
\describe{
\item{\code{weight}}{Weight of the seedling.}
\item{\code{pH.z}}{The pH treatment (as numerical).}
\item{\code{pH}}{The pH treatment (as a factor).}
}
}
\usage{
BirchSeedlings
}
\description{
Wood and Bormann (1974) studied the effect of acid rain 
on trees. "Clean" precipitation has a pH in the 5.0 to 5.5 range, 
but observed precipitation pH in northern New Hampshire is often 
in the 3.0 to 4.0 range. 

One of their experiments used 240 six-week-old yellow birch seedlings. 
These seedlings were divided into five groups of 48 at random, 
and the seedlings within each group received an acid mist 
treatment 6 hours a week for 17 weeks. The five treatments 
differed by mist pH: 4.7, 4.0, 3.3, 3.0, and 2.3; otherwise, the 
seedlings were treated identically. After the 17 weeks, the 
seedlings were weighed, and total plant (dry) weight in grams was 
taken as response.  

This data set is computer generated to match the treatment
means and mean square for error reported by Wood and Bormann (1974).

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Wood, T. and F. H. Bormann (1974). Effects of an artificial acid 
mist upon the growth of Betula alleghaninsis britt. 
\emph{Environmental Pollution} 7, 259–268.
}
\keyword{datasets}
