% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{BirdBones}
\alias{BirdBones}
\title{Calcium in bird bones.}
\format{
A dataframe of 32 observations on three variables.
\describe{
\item{\code{treatment}}{A factor indicating the treatment group.}
\item{\code{diameter}}{The diameter of the chick bone.}
\item{\code{calcium}}{The calcium the chick bone.}
}
}
\usage{
BirdBones
}
\description{
Pollutants may reduce the strength of bird bones.  We believe
that the strength reduction, if present, is due to a change in the
bone itself, and not a change in the size of the bone.  One measure of
bone strength is calcium content.  We have an instrument which can
measure the total amount of calcium in a 1cm length of bone.  Bird
bones are essentially thin tubes in shape, so the total amount of
calcium will also depend on the diameter of the bone.

Thirty-two chicks are divided at random into four groups.  Group 1
is a control group and receives a normal diet.  Each other group receives
a diet including a different toxin  (pesticides related to DDT).  At
6 weeks, the chicks are sacrificed and the calcium content (in mg)
and diameter (in mm) of the right femur is measured for each chick.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
