% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Bottles}
\alias{Bottles}
\title{Net weights of filled bottles.}
\format{
A dataframe of 16 observations on two variables.
\describe{
\item{\code{head}}{A factor indicating the filling head.}
\item{\code{weight}}{The weight of oil in the bottle.}
}
}
\usage{
Bottles
}
\description{
A 24-head machine fills bottles with vegetable oil.
Five of the heads are chosen at random, and several consecutive
bottles from these heads were taken from the line.  The data values are the actual net
weight of oil in these selected bottles (in ounces).

Source: Oehlert (2022), originally from Swallow and Searle (1978).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Swallow, William H. and Searle, S. R. (1978) "Minimum Variance Quadratic Unbiased Estimation (MIVQUE)
of variance components." \emph{Technometrics} \strong{20}, 265-272.
}
\keyword{datasets}
