% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Breakage}
\alias{Breakage}
\title{Breakage rates during shipping}
\format{
A dataframe of 20 observations on two variables.
\describe{
\item{\code{company}}{A factor indicating the company used.}
\item{\code{breakage}}{The percentage of broken items.}
}
}
\usage{
Breakage
}
\description{
Synthetic data. Twenty batches of fragile items are randomly assigned to one of
four shipping companies, five batches per company. The response is the percentage of items
broken in each shipment.

Source: Oehlert (2022); synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
