% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brownforsythe.R
\name{brown.forsythe.test}
\alias{brown.forsythe.test}
\title{Brown-Forsythe modified F-test}
\usage{
brown.forsythe.test(formula, data)
}
\arguments{
\item{formula}{A two-sided model formula with a single (factor) predictor.}

\item{data}{An optional data frame.}
}
\value{
A list of class "htest" with components statistic (the modified F-test),
parameter (the numerator and modified denominator degrees of freedom),
p.value, method (a character string mentioning Brown-Forsythe),
and data.name (a character string naming the response and predictor
variables).
}
\description{
Brown-Forsythe alternative to ANOVA with heteroskedastic errors
}
\details{
The Brown-Forsythe test is a modified F-test for testing
equality of means in a one-way classification when there may be
unequal variances.  It takes the form of a modifed F-statistic,
which is the usual sum of squares between treatments divided by
a weighted linear combination of within group variances.  The
degrees of freedom are the usual numerator degrees of freedom and
a modified denominator degrees of freedom.

\code{\link[stats:oneway.test]{oneway.test()}} serves a similar purpose,
but uses a different adjustment for non-constant variance.
}
\examples{
data(ResinLifetimes)
brown.forsythe.test(Time~temp,data=ResinLifetimes)
oneway.test(Time~temp,data=ResinLifetimes) # alternative in R base

}
\seealso{
\code{\link[stats:oneway.test]{oneway.test()}}
}
