% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CaffeineAdenine}
\alias{CaffeineAdenine}
\title{Adenine brain concentrations post caffeine treatment}
\format{
A dataframe of 43 observations on three variables:
\describe{
\item{\code{caffeine.z}}{Caffeine concentration.}
\item{\code{caffeine}}{A factor version of \code{caffeine.z}, for convenience.}
\item{\code{adenine}}{Adenine concentration.}
}
}
\usage{
CaffeineAdenine
}
\description{
Caffeine may affect the ability of other compounds to cross the blood
brain barrier. Forty-three rats are randomly assigned to one of eight different
treatments. Each treatment consists of an arterial injection of C^14-labeled
adenine together with one of eight different concentrations of caffeine (0, .1,
.5, 1, 5, 10, 25, 50 mM). The response is concentration of labeled adenine in
the rats' brain post injection.

Source: Oehlert (2022), originally from McCall, Millington, and Wurtman (1982)
}
\references{
Mccall, A. L, Millington, W. R., and Wurtman, R. J. (1982) "Blood-brain
barrier transport of caffeine: dose-related restriction of adenine transport."
\emph{Life Sciences} \strong{31}, 2709-2715.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
