% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CakeBaking2}
\alias{CakeBaking2}
\title{Cake baking (second order).}
\format{
A data frame of 14 observations on our variables.
\describe{
\item{\code{block}}{Factor indicating first or second block.}
\item{\code{time}}{Coded level of time treatment.}
\item{\code{temp}}{Coded level of temperature treatment.}
\item{\code{appeal}}{The palatability of the cake.}
}
}
\usage{
CakeBaking2
}
\description{
This is a second-order design in baking time and temperature,
run in two blocks.
The coded values -1, 0, 1 represent 33, 35, and 37 minutes for time,
and the coded values -1, 0, 1 for 340, 350, and 360 degrees
for temperature.  The response is an average palatability score,
with higher values being desirable. Block 1 is the same data
as the CakeBaking1 data set.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
