% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CakeHeights}
\alias{CakeHeights}
\title{Cake heights and oven positions.}
\format{
A data frame of eight observations on two variables.
\describe{
\item{\code{trt}}{Factor indicating the four treatments.}
\item{\code{height}}{Height of cakes in mm.}
}
}
\usage{
CakeHeights
}
\description{
Cakes can be baked either one at a time on the top or bottom oven rack,
or two at a time, with one on the top rack and one on the bottom rack.
We bake eight cakes, with two cake mix boxes randomly assigned to each of
the four treatments: top rack single, bottom rack single, top rack double,
bottom rack double. After the cake is baked and allowed to cool for 1 hour,
its height (mm) is measured at five locations, with the average of the five
measurements taken as the height for the cake.

Source: Oehlert (2022), originally from D. Schendel (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
