% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CalcExams}
\alias{CalcExams}
\title{Consistency of calculus exams.}
\format{
A data frame of 18 observations on four variables.
\describe{
\item{\code{school}}{A factor indicating where the observation was taken.}
\item{\code{exam}}{A factor indicating where the exam form.}
\item{\code{text}}{A factor indicating which textbook was used.}
\item{\code{score}}{Average test score.}
}
}
\usage{
CalcExams
}
\description{
Most universities teach many sections of introductory calculus,
and faculty are constantly looking for a method to evaluate
students consistently across sections.
Generally, all sections of intro-calculus take the final
exam at the same time, so a single exam is used for all sections.  An
exam service claims that it can supply different exams that consistently
evaluate students.  Some faculty doubt this claim, in part because they
believe that there may be an interaction between the text used and the exam used.

Three math departments (one each at Minnesota, Washington, and Berkeley)
propose the following experiment.  Three random final exams are obtained
from the service: E1, E2, and E3.  At Minnesota, the three exams will be
used in random order in the fall, winter, and spring quarters.  Randomization
will also be done at Washington and Berkeley.  The three schools all use
the same two intro calculus texts.  Sections of intro calculus at each
school will be divided at random into two groups, with half of the sections
using text A and the other half using text B.  The
mean test scores are tallied as a response.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
