% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CalfCounts}
\alias{CalfCounts}
\title{Calf count differences under control and medicated conditions}
\format{
A data frame of seven observations on one variable.
\describe{
\item{\code{differences}}{The differences in the number of pregnant cows.}
}
}
\usage{
CalfCounts
}
\description{
Seven bulls are each bred to two herds of cows (14 total herds).
For one randomly chosen herd per bull, the bull is given a drug that is supposed
	to increase fertility. The response is the number of pregnant cows under the treatment
condition minus the number of pregnant cows under the control condition.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
