% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CALM}
\alias{CALM}
\title{Pediatricians maintaining confidentiality.}
\format{
A data frame of 16 observations on five variables.
\describe{
\item{\code{complaint}}{The complaint in the fable.}
\item{\code{age}}{The age of the patient.}
\item{\code{time.known}}{How long the doctor has known the family.}
\item{\code{maturity}}{Maturity of the patient.}
\item{\code{fraction}}{The fraction of doctors maintaining confidentiality.}
}
}
\usage{
CALM
}
\description{
In a study of patient confidentiality, a large number of
pediatricians was surveyed.  Each pediatrician was given a "fable"
about a female patient less than 18 years old.  There
were sixteen different fables, the combinations of the factors complaint
(drug problem or sexually transmitted disease),
age (14 years or 17 years), the length of
time the pediatrician had known the family (less than 1 year
or more than 5 years), and the maturity of patient (immature for
age, 2 or mature for age).  The response at each combination of factor
levels is the fraction of doctors who would keep confidentiality
and not inform the patient's parents.

Source: Oehlert (2022), modeled on Moses (1987).
}
\references{
Moses, L. E. (1987). "Analysis of Bernoulli data from a 2^5 design done in blocks
of size four. In C. Mallows (Ed.), \emph{Design, Data, and Analysis},
pp. 275-290. New York: Wiley.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
