% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CheeseAminoAcid}
\alias{CheeseAminoAcid}
\title{Free amino acids in cheese after 168 days ripening.}
\format{
A data frame of eight observations on two variables.
\describe{
\item{\code{strain}}{The strain(s) of nonstarter bacteria added.}
\item{\code{freeAminoAcid}}{The amino acids measured after ripening.}
}
}
\usage{
CheeseAminoAcid
}
\description{
Nonstarter bacteria in cheese come from the environment of the
cheese factory and are not added by the maker (who adds the starter bacteria).
These nonstarter bacteria are throught to be crucial to cheese flavor. Nonstarter
bacterial can be isolated and added under laboratory conditions. Eight cheeses
will be made, and these are randomly assigned to four different nonstarter treatments:
none, strain A, strain B, strains A and B. The response is the total free amino
acids in the cheeese after 168 days of ripening.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
