% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CheeseAminoAcid56}
\alias{CheeseAminoAcid56}
\title{Free amino acids in cheese after 56 days ripening}
\format{
A data frame of twelve observations on three variables.
\describe{
\item{\code{r50}}{Factor indicating presence or absence of strain r50.}
\item{\code{r21}}{Factor indicating presence or absence of strain r21.}
\item{\code{freeAminoAcid}}{The amino acids measured after ripening.}
}
}
\usage{
CheeseAminoAcid56
}
\description{
Nonstarter bacteria in cheese come from the environment of the
cheese factory and are not added by the maker (who adds the starter bacteria).
These nonstarter bacteria are throught to be crucial to cheese flavor. Nonstarter
bacterial can be isolated and added under laboratory conditions. Twelve cheeses
will be made, and these are randomly assigned to four different nonstarter treatments:
none, R50, R21, strains R50 and R21. The response is the total free amino
acids in the cheeese after 56 days of ripening.

Source: Oehlert (2022), data from P. Swearingen (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
