% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ChickBodyWeight}
\alias{ChickBodyWeight}
\title{Body weight of chicks taking mineral supplements}
\format{
A dataframe of 6 observations on three variables.
\describe{
\item{\code{weight}}{The mean weight of chicks in grams.}
\item{\code{P}}{The level of phosphorus supplement.}
\item{\code{Ca}}{The level of calcium supplement.}
}
}
\usage{
ChickBodyWeight
}
\description{
Twenty-four pens of broiler chicks are randomly
assigned to six treatments, the factorial combinations of
calcium supplement (.5, .9, or 1.2\%) and phosphorus supplement
(.25 or .5\%). The response is average body weight of chicks in a
pen in grams. This data set only contains treatment means.

Source: Oehlert (2022), originally from Nelson et al. (1990).
}
\references{
Nelson, T. S., L. K. Kirby, and Z. B. Johnson (1990). "Effect of minerals
on the incidence of leg abnormalities in growing broiler chicks." \emph{Nutrition
Research} \strong{10}, 525-533.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
