% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ChickenSemen}
\alias{ChickenSemen}
\title{Effect of freezing on potency of chicken semen.}
\format{
A data frame of 16 observations on three variables.
\describe{
\item{\code{RIR}}{A factor indicating the Rhode Island Red semen is fresh or frozen.}
\item{\code{WL}}{A factor indicating the White Leghorn semen is fresh or frozen.}
\item{\code{fraction}}{The fraction of offspring with white feathers.}
}
}
\usage{
ChickenSemen
}
\description{
Artificial insemination is an important tool in agriculture,
but freezing semen for later use can reduce its potency
(ability to produce offspring).
Here we are trying to understand the effect of freezing
on the potency of chicken semen.
Four semen mixtures are prepared, consisting of equal
parts of either
fresh or frozen Rhode Island Red semen, and either fresh or
frozen White Leghorn semen.  Sixteen batches of Rhode Island Red hens
are assigned at random, four to each of the four treatments.
Each batch of hens is
inseminated with the appropriate mixture,
and the response measured is the fraction of
the hatching eggs that have white feathers and thus
White Leghorn fathers.

Source: Oehlert (2022), originally from Tajima (1987).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Tajima, A. (1987) \emph{Some Aspects of Preserving Checken Semen: Glycerol Effect,
Assay Method, and Application} Ph. D. thesis, University of Minnesota, St. Paul, MN.
}
\keyword{datasets}
