% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Cisplatin}
\alias{Cisplatin}
\title{Viable cancer cells after treatment with Cisplatin.}
\format{
A data frame of 12 observations on four variables.
\describe{
\item{\code{Cisplatin.z}}{The concentration of Cisplatin.}
\item{\code{Cisplatin}}{A factor version of \code{Cisplatin.z}.}
\item{\code{type}}{A factor indicating whether dye was added.}
\item{\code{count}}{The count of cells.}
}
}
\usage{
Cisplatin
}
\description{
One measure of the effectiveness of cancer drugs is their
ability to reduce the number of viable cancer cells in
laboratory settings.  In this experiment, the A549 line of
malignant cells is plated onto petri dishes with various
concentrations of the drug cisplatin.  After 7 days of incubation,
half the petri dishes at each dose are treated with a dye, and
the number of viable cell colonies per 500 mm^2 is determined
as a response for all petri dishes.
The dye is supposed to make the counting machinery more specific
to the cancer cells.

Source: Oehlert (2022), after Figure 1 of Alley, et al. (1982).
}
\references{
Alley, M. C., C. B. Uhl, and M. N. Lieber (1982). "Improved detection of drug
cytotoxicity in the soft agar colony formation assay through use of a
metabolizable tetrazolium salt," \emph{Life Sciences} \strong{31}, 3071-3078.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
