% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CloudBackup}
\alias{CloudBackup}
\title{Save/restore time for cloud file backup}
\format{
A dataframe of 30 observations on two variables.
\describe{
\item{\code{service}}{A factor with levels 1, 2, and 3 indicating the
backup service tested.}
\item{\code{updowntime}}{The time in seconds to upload a 100MB file to
a remote backup server and then download it back again.}
}
}
\usage{
CloudBackup
}
\description{
A synthetic dataset giving the time in seconds to
save a file to an online backup service and then recover it.
There are three cloud backup services tested.

The data are listed in temporal order.

Source: Oehlert (2022)
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
