% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CloudSeeding}
\alias{CloudSeeding}
\title{Rainfall in cloud seeding experiment}
\format{
A dataframe of 52 observations on two variables.
\describe{
\item{\code{seeding}}{A factor with two levels: nonseeded and seeded.}
\item{\code{rainfall}}{Rainfall amounts in acre feet.}
}
}
\usage{
CloudSeeding
}
\description{
Rainfall in acre-feet from 52 clouds, 26 of which
were seeded with silver oxide.

Source: Oehlert (2022), originally from Simpson, Olsen, and Eden (1975).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Simpson, J. A., A. Olsen, and J. C. Eden (1975). "A Bayesian analysis of a multiplicative
treatment effect in weather modification." \emph{Technometrics} \strong{17}, 161-166.
}
\keyword{datasets}
