% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Cockroaches}
\alias{Cockroaches}
\title{ATPase activity in chlordane-treated cockroaches}
\format{
A dataframe of 6 observations on one variable.
\describe{
\item{\code{pctchange}}{Change in activity.}
}
}
\usage{
Cockroaches
}
\description{
The coxal muscles from
one meso- and one metathoracic leg on opposite sides were
surgically extracted from each of six roaches. The roaches were
then treated with 50 micrograms of alpha-chlordane,
and coxal muscles from the two remaining meso- and metathoracic legs
were removed about two hours after treatment.
The response is the percentage change in the Na-K-ATPase activity in the muscles.

Source: Oehlert (2022), originally from Chu (1970).
}
\references{
Chu, Y. C. (1970). "Comparison of \emph{in vivo} and \emph{in vitro} inhibition
of ATPases by the insecticide Chlordane. Master's thesis, University of Minnesota,
St. Paul, MN.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
