% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{COEmissions}
\alias{COEmissions}
\title{CO Emissions of a gasoline engine under various conditions}
\format{
A dataframe of 46 observations on seven variables.
\describe{
\item{\code{load}}{Coded value of engine load.}
\item{\code{speed}}{Coded value of engine speed.}
\item{\code{advance}}{Coded value of spark advance.}
\item{\code{ratio}}{Coded value of air/fuel ratio.}
\item{\code{recycle}}{Coded value of exhaust gas recycle.}
\item{\code{block}}{The block of the observatoin.}
\item{\code{emissions}}{CO emissions.}
}
}
\usage{
COEmissions
}
\description{
An experiment is performed to determine how a gasoline
engine responds to various factors.  The response of interest
is CO emissions in grams per hour.  The design factors are
engine load, in Newton meters, range (30,70);
engine speed, in rpm, range (1000, 4000);
spark advance, in degrees, range (10, 30);
air-to-fuel ratio, dimensionless, range (13, 16.4);
and exhaust gas recycle, in percent, range (0, 10).
The experimental design has 46 observations in two blocks
of 23 each.
The design factors have been coded to the range (-1, 1).

Source: Oehlert (2022), originally from Chu (1970).
}
\references{
Draper, N. R., T. P. Davis, L. Pozueta, and D. M. Grove (1994). "Isolation
of degrees of freedom for Box-Behnken designs," \emph{Technometrics} \strong{36}, 283-291.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
