% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnett.R
\name{compare.to.best}
\alias{compare.to.best}
\title{Multiple comparisons with best.}
\usage{
compare.to.best(lmodel, term, confidence = 0.95, lowisbest = FALSE)
}
\arguments{
\item{lmodel}{A model object from \code{lm}, \code{lmer}, or \code{lme}.}

\item{term}{A term (variable name) in the linear model.}

\item{confidence}{The level of confidence for interval estimates of differences with control.}

\item{lowisbest}{A logical determining whether the highest or lowest response is best.}
}
\value{
A matrix with columns giving the estimated 
difference in mean response between each non-control level of the factor and the
best level, and the confidence limits for those differences.
}
\description{
Multiple comparisons of levels of a factor to the "best" level, which could
be either the highest or lowest response, with
error controlled via the Dunnett distribution.
}
\details{
\code{compare.to.best} computes one-sided confidence intervals for the difference
in mean between the best treatment and every other treatment. The output includes the 
differences and the one-sided allowance. Significant differences are marked with a star.
The true best treatment is in the unstarred group with the selected confidence rate.
By default, the best is considered to be a maximum response.  That can be
changed via \code{lowisbest}.

For linear mixed models, error df are estimated with the Kenward and Roger method.
}
\examples{
data(WeedControl)
# transform response for variance stability
weeds.separate <- lm(sqrt(100-pct.control)~treatment,data=WeedControl)
compare.to.best(weeds.separate,treatment,conf=.99,lowisbest=TRUE)

}
\seealso{
\code{\link{pairwise}}, \code{\link{compare.to.control}},
\code{\link[stats:pairwise.t.test]{stats::pairwise.t.test}},
\code{\link[stats:TukeyHSD]{stats::TukeyHSD}}
}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
