% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnett.R
\name{compare.to.control}
\alias{compare.to.control}
\title{Multiple comparisons with control.}
\usage{
compare.to.control(lmodel, term, control.level, confidence = 0.95)
}
\arguments{
\item{lmodel}{A model object from \code{lm}, \code{lmer}, or \code{lme}.}

\item{term}{A term (variable name) in the linear model.}

\item{control.level}{The level of the factor that represents the control.}

\item{confidence}{The level of confidence for interval estimates of differences with control.}
}
\value{
A matrix with columns giving the estimated 
difference in mean response between each non-control level of the factor and the
control level, and the lower and upper confidence limits for those differences.
}
\description{
Multiple comparisons of levels of a factor to a control level, with
error controlled via the Dunnett distribution.
}
\details{
\code{compare.to.control} computes two-sided confidence intervals
for the difference in mean between every non-control level of the factor
and the control level of the factor.  These intervals are simultaneous with
coverage rate given by the selected confidence parameter.  Significant differences
are marked with a star.

For linear mixed models, error df are estimated with the Kenward and Roger method.
}
\examples{
data(FruitFlyLifespan)
fit.flies <- lm(longevity~companions,data=FruitFlyLifespan)
compare.to.control(fit.flies,companions,5) # group 5 is no companions

}
\seealso{
\code{\link{pairwise}}, \code{\link{compare.to.best}},
\code{\link[stats:pairwise.t.test]{stats::pairwise.t.test}},
\code{\link[stats:TukeyHSD]{stats::TukeyHSD}}
}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
