% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ConcreteFaulting}
\alias{ConcreteFaulting}
\title{Faulting in jointed concrete pavements}
\format{
A dataframe with 108 observations on eight variables.
\describe{
\item{\code{base.z}}{The thickness of the base in inches.}
\item{\code{base}}{A factor version of the thickness.}
\item{\code{subgrade.z}}{The resilience of the subgrade layer.}
\item{\code{subgrade}}{A factor version of the subgrade resilience.}
\item{\code{wcratio.z}}{The water to concrete ratio in the mix.}
\item{\code{wcratio}}{A factor version of the water to concrete ratio.}
\item{\code{dowels}}{The presence or absence of dowels.}
\item{\code{faulting}}{The faulting.}
}
}
\usage{
ConcreteFaulting
}
\description{
Faulting is a difference in height on two sides of a concrete
pavement joint. Many factors can influence the amount of faulting
as the pavement ages. This study examines four factors: the
thickness of the base layer (5, 10, or 15 inches), the resilience
of the subgrade layer (5,000, 10,000, or 20,000 psi), the ratio
of water to cement in the mix (.33, .42, .55), and the presence
of dowel pins between successive pavement slabs (yes or no).
The response is faulting (in inches) after a 20 year life. There
are two runs for each factor-level combination, resulting
in 108 observations.

These data are based on a software package that simulates the behavior
of concrete (AASHTP 2002 Design Guide). In that sense, the "experiment"
merely recapitulates the assumptions and equations of the model.

Source: Oehlert (2022), originally from N. Funk (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
