% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ConcreteStrength}
\alias{ConcreteStrength}
\title{Strength of concrete with embedded fibers}
\format{
A dataframe with 15 observations on three variables.
\describe{
\item{\code{fiberPct.z}}{The percent fiber content.}
\item{\code{fiberPct}}{The percent fiber content as a factor.}
\item{\code{strength}}{Strength of the concrete cube in ksi.}
}
}
\usage{
ConcreteStrength
}
\description{
Fifteen concrete cubes are produced, each with one of five levels of
polypropylene fibers added to the mix. Fiber levels are randomly assigned to
concrete batches, and the strength of the concrete (in ksi, kilopound/sq in) is
measured as the response.

Source: Oehlert (2022), originally from Paskova and Meyer (1997).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Paskova, T. and C. Meyer (1997). "Low-cycle fatigue of plain and fiber-reinforced
concrete. \emph{ACI Materials Journal} \strong{94}, 273-285.
}
\keyword{datasets}
