% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CookieTexture}
\alias{CookieTexture}
\title{Cookie texture over time}
\format{
A dataframe of 220 observations on four variables.
\describe{
\item{\code{day.z}}{The day the cookie was measured.}
\item{\code{day}}{A factor version of \code{day.z}.}
\item{\code{recipe}}{A factor variable giving the recipe used for this cookie.}
\item{\code{texture}}{The texture of the cookie, as measured by force.}
}
}
\usage{
CookieTexture
}
\description{
Two-hundred twenty cookies are baked.  Half of them are
made with a control recipe, while the other half include some
raffinose.  The 110 cookies of each recipe are then
randomly divided into 11 groups of ten each.  Each group of
10 is put in a plastic bag.  On the next day, one bag is
chosen at random from each recipe, and the 10 cookies in that
bag are measured for texture.  The texture measurement is
the force required to push a 1.27 cm diameter probe into the
cookie.  This is repeated on each successive day until the last
bag of cookies has been measured on day 11.

Source: Oehlert (2022) synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
