% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Copepoda}
\alias{Copepoda}
\title{Copepoda counts after simulated snowmelt}
\format{
A data frame of six observations on two variables.
\describe{
\item{\code{pH}}{A factor indicating reduced or neutral pH.}
\item{\code{count}}{The count of Copepoda.}
}
}
\usage{
Copepoda
}
\description{
Copepoda are small crustaceans, some of which live in fresh water.
Six samples of wetland soil were subjected to an artificial snowmelt, with three
randomly chosen for neutral pH snowmelt and the other three receiving a reduced pH.
The response is the number of Copepoda removed from each microcosm in the 14 days
post snow melt.

Source: Oehlert (2022), originally from Dale (1992).
}
\references{
Dale, T. B. (1992) "Bilogical and chemical effects of acidified snowmelt on
seasonal wetlands in Minnesota. Master's Thesis, University of Minnesota, St. Paul, MN.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
