% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CopperZinc}
\alias{CopperZinc}
\title{Removal of copper and zinc from waste water}
\format{
A data frame of 32 observations on seven variables.
\describe{
\item{\code{pH}}{A factor indicating the pH of the reaction.}
\item{\code{initial}}{A factor the initial concentration of metal (mg/l).}
\item{\code{extract}}{A factor indicating the concentration of the extractant (percent by volume).}
\item{\code{medium}}{A factor indicating the medium of the solution.}
\item{\code{stirring}}{A factor indicating the stirring rate of the solution (rpm).}
\item{\code{zinc}}{Percent removal of zinc.}
\item{\code{copper}}{Percent removal of copper.}
}
}
\usage{
CopperZinc
}
\description{
Heavy metals should be removed from the waste water
stream before it is released into the environment. This
experiment examines how five factors affect the removal
of zinc and copper from water via liquid-liquid extraction.
There are five factors under consideration:
pH of initial solution (4.5 or 6.5), 
initial concentration of the metal (25 or 75 mg/l),
concentration of the extractant (5 or 10 \% by volume),
medium of the solution (sulfate or chloride),
and stirring rate (400 or 500 rpm). The 32 factor/level
combinations were assigned to 32 independent runs,
and two responses were measured: percent removal of zinc
and percent removal of copper.

Source: Oehlert (2022), originally from Berrama et al (2013) via
Lye (2019).
}
\references{
Berrama, T., N. Benaouag, F. Kaouah, and Z. Bendjama (2013) "Application of full
 factorial design to study the simultaneous removal of copper and zinc from aqueous 
solution by liquid-liquid extraction," \emph{Desalination and Water Treatment} \strong{51} 2135-2145.

Lye, Leonard M. (2019) \emph{Applications of DEO in Engineering and Science:
A Collection of 2^6 Case Studies}, Leonard Lye, St. John's, Newfoundland.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
