% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Corrosion}
\alias{Corrosion}
\title{Anti-corrosion coatings for aluminum.}
\format{
A dataframe of 64 observations on five variables.
\describe{
\item{\code{weeks.z}}{The number of weeks of weathering.}
\item{\code{weeks}}{A factor versioin of weeks.z.}
\item{\code{coating}}{A factor indicating the coating used.}
\item{\code{weathering}}{A factor indicating the weathering protocol.}
\item{\code{ocp}}{The open circuit potential (volts).}
}
}
\usage{
Corrosion
}
\description{
Aluminum corrosion is a major issue in the aerospace industry, and
a great deal of work has gone into finding coatings that will prevent, or at least delay,
corrosion. Electrochemistry states that only the more
reactive of two metals will react when two joined metals are in a corrosive environment.
For this reason, coatings that are magnesium rich continue to be developed
for aluminum. These are called "sacrificial coatings," because they are
literally consumed while protecting the aluminum underneath.
This experiment compares a commercially available
coating to three experimental  coatings, when weathered according to two different
protocols for corrosive conditions (B117 or D5894),
maintained for four different lengths of time (0, 1, 2, or 3 weeks).
The three
experimental coatings differ in the concentration of magnesium
pigment: 30\%, 40\%, or 45\%. (Too high a Mg concentration can
degrade other desirable properties such as adhesion or durability.)
Sixty-four aluminum samples are randomly assigned
to the thirty-two combinations of coating, weathering, and time. After treatment,
each sample is tested for its
"open circuit potential" (OCP, in volts) as measured by electrochemical impedance. Lower (more negative)
values are better.

Source: Oehlert (2022), data from T. Chen (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
