% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Cracks}
\alias{Cracks}
\title{Cracks in pavement.}
\format{
A dataframe with 11 observations on three variables.
\describe{
\item{\code{block}}{Factor indicating the level of traffic.}
\item{\code{treatment}}{Factor indicating the treatment applied.}
\item{\code{cracks}}{The number of cracks per mile.}
}
}
\usage{
Cracks
}
\description{
Winter road treatments to clear snow and ice can lead to cracking in
the pavement.
This experiment was conducted to compare four treatments:
sodium chloride, calcium chloride, a proprietary organic compound, and sand.
Traffic level was used as a blocking factor and a randomized complete block
experiment was conducted.  One observation is missing, because the
spreader in that district was not operating properly.  The response is new
cracks per mile of treated roadway

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
