% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DirtyDishes}
\alias{DirtyDishes}
\title{Comparison of dish detergent solutions.}
\format{
A dataframe with 36 observations on three variables.
\describe{
\item{\code{session}}{Factor indicating the washing session.}
\item{\code{treat}}{Factor indicating the detergent treatment.}
\item{\code{plates}}{The number of plates washed.}
}
}
\usage{
DirtyDishes
}
\description{
Nine different dishwashing solutions are compared.
The first four consist of base detergent I and 3, 2,
1, and 0 parts of an additive; solutions five through
eight consist of base detergent II and 3, 2,
1, and 0 parts of an additive; the last solution is
a control.  There are three washing basins and one
operator for each basin.  The three operators wash
at the same speed during each test, and the response
is the number of plates washed when the foam disappears.
The speed of washing is the same
for all three detergents used at any one session,
but could differ from session to session.

Source: Oehlert (2022), originally from John (1961).
}
\references{
John, P. W. M. (1961) "An application of a balanced incomplete block
design," \emph{Technometrics} \strong{3}, 51-54.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
