% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DiskSpeed}
\alias{DiskSpeed}
\title{Access time of disk drives.}
\format{
A dataframe with 20 observations on three variables.
\describe{
\item{\code{computer}}{Factor indicating the computer used in the test.}
\item{\code{brand}}{Factor indicating the brand of disk drive used in the test.}
\item{\code{time}}{Average access time.}
}
}
\usage{
DiskSpeed
}
\description{
An experiment to compare the average access times of
five brands of half-height computer disk drives
(denoted A through E).  There is a block on computer
holding the drive, but each computer can only hold four drives.
The response is verage access time.

Source: Oehlert (2022), originally from Nelson (1993).
}
\references{
Nelson, P. R. (1993). "Additional uses for the analysis of means and extended
tables of critical values," \emph{Technometrics} \strong{35}, 61-71.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
