% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DNA}
\alias{DNA}
\title{Optical density of DNA solutions.}
\format{
A dataframe of 180 observations on six variables.
\describe{
\item{\code{vol.z}}{The volume of the sample.}
\item{\code{vol}}{A factor versioin of vol.z.}
\item{\code{conc.z}}{The concentration of the sample.}
\item{\code{conc}}{A factor versioin of conc.z.}
\item{\code{user}}{A factor indicating the user.}
\item{\code{od260}}{The measured optical density of the sample.}
}
}
\usage{
DNA
}
\description{
One of the steps in a molecular biological analysis is
the quantification of DNA.  This can be done by measuring the
absorbance of ultra-violet light at 260 nm (called the optical
density).  The absorption
of light in the spectrophotometer
should be proportional to concentration of DNA, but
should not depend on the volume of sample used.  However, there
is a general belief that small samples, say less than 40 micro l,
lead to erroneous results.  In addition, the proportionality cited above
only holds over a range of concentrations; outside that range
nonlinear effects come into play. In theory, the optical density at 260 nm
should be 0.02 C,
where C is the concentration in ng/micro l.

This experiment measures the optical density at five concentrations
(10, 30, 60, 120, 480 ng/ul) and six volumes (15, 20, 30, 40, 50,
100 ul).  Three analysts are chosen at random from the 16 in the
lab.  Each of the analysts prepares two samples at each volume/concentration
combination and measures the optical density.

Source: Oehlert (2022), data from S. Charaniya (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
