% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DyeAdsorption}
\alias{DyeAdsorption}
\title{Dye Adsorption on Carbon}
\format{
A data frame of 18 observations on seven variables.
\describe{
\item{\code{dye.z}}{The initial concentration of dye (Mg/l).}
\item{\code{dye}}{A factor version of the initial concentration of dye.}
\item{\code{pH.z}}{The pH of the reaction.}
\item{\code{pH}}{A factor version of the pH of the reaction.}
\item{\code{temperature.z}}{The temperature of the reaction (C).}
\item{\code{temperature}}{A factor version of the temperature of the reaction (C).}
\item{\code{adsorbed}}{The percentage of dye adsorption.}
}
}
\usage{
DyeAdsorption
}
\description{
The effect of initial dye concentration, temperature,
and pH on the absorption of the dye Rhodamine 6G by activated
carbon beads incorporated with calcium alginate. The response
is the percentation of adsoprtion for 1 g/L of beads.

The treatments are the factor/level combinations of initial
concentration at three levels (100, 200, 300 Mg/l), pH
(7, 8, 9) and temperature (30 or 30 degrees C).

Source: Oehlert (2022), originally from Annadurai et al (2002) via
Lye (2019).
}
\references{
Annadurai, G., Ruey-Shin Juang, and Duu-Jong Lee (2002) "Factorial design
 analysis for adsorption of dye on activated carbon beads incorporated with 
calcium alginate," \emph{Advances in Environmental Research}, \strong{6},  191-198.

Lye, Leonard M. (2019) \emph{Applications of DEO in Engineering and Science:
A Collection of 2^6 Case Studies}, Leonard Lye, St. John's, Newfoundland.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
