% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{DyeRemoval}
\alias{DyeRemoval}
\title{Dye Removal via Oxidation}
\format{
A data frame of 90 observations on nine variables.
\describe{
\item{\code{time.z}}{The length of the reaction (minutes).} 
\item{\code{time}}{A factor version of the length of the reaction.}
\item{\code{oxidant.z}}{The presulfate concentration (mg/l).}
\item{\code{oxidant}}{A factor version of the persulfate concentration.}
\item{\code{initial.z}}{The initial dye concentration (mg/l).}
\item{\code{initial}}{A factor version of the initial dye concentration.}
\item{\code{temp.z}}{The temperature of the reaction (C).}
\item{\code{temp}}{A factor version of the temperature.}
\item{\code{CRE}}{The color removal efficiency (percent).}
}
}
\usage{
DyeRemoval
}
\description{
The dye methylene blue may be removed from aqueous solution via
an oxidation reaction with persulfate. This experiment studies how the
percentage of dye removed varies with reaction time (5, 10, 15, 20, or 25
minutes), persulfate concentration (355, 710, or 1065 mg/l),
the initial dye concentration (10, 15, or 20 mg/l), and the
process temperature (60 or 70 degrees C). The response is the color
removal efficiency (CRE in precentage).  There are 90 factor/level
combinations, each observed once.

Source: Oehlert (2022), originally from Kordkandi and Forouzesh (2014) via
Lye (2019).
}
\references{
Kordkandi, S. A. and Mojtaba Forouzesh (2014) "Application of full factorial
design for methylene blue dye removal using heat-activated persulfate oxidation,"
\emph{Journal of the Taiwan Institute of Chemical Engineers} \strong{45} 2597-2604.

Lye, Leonard M. (2019) \emph{Applications of DEO in Engineering and Science:
A Collection of 2^6 Case Studies}, Leonard Lye, St. John's, Newfoundland.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
