% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FatAcidity}
\alias{FatAcidity}
\title{Fat acidity in flour.}
\format{
A dataframe with 132 observations on five variables.
\describe{
\item{\code{weeks.z}}{The number of weeks the flour was stored.}
\item{\code{weeks}}{A factor version of \code{weeks.z}.}
\item{\code{extraction}}{The type of extraction.}
\item{\code{strength}}{The type of flour.}
\item{\code{strength}}{The fat acidity.}
}
}
\usage{
FatAcidity
}
\description{
Fat acidity is a measure of flour quality that depends
on the kind of flour, how  the flour has been treated, and
how long the flour is stored.  In this experiment there
are two types of flour (Patent or First Clear);
the flour treatment factor (extraction)
has eleven levels, and the flour has been stored for one of
six periods (0, 3, 6, 9, 15, or 21 weeks).
We observe only one unit for each factor-level combination.
The response is fat acidity in mg KOH/100 g flour.

Source: Oehlert (2022), originally from Nelson (1961).
}
\references{
Nelson, J. W. (1961) \emph{The Nature of Wheat Lipids and Their Role in Flour
Deterioration.} Ph. D. thesis, Univeristy of Minnesota, St. Paul, MN.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
