% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Fillings}
\alias{Fillings}
\title{Hardness of dental fillings.}
\format{
A dataframe of 120 observations on four variables.
\describe{
\item{\code{allow}}{A factor indicating the allow used.}
\item{\code{method}}{A factor indicating the method used.}
\item{\code{dentist}}{A factor indicating the dentist preparing the filling.}
\item{\code{hardness}}{The hardness (diamond pyramid number) of the filling.}
}
}
\usage{
Fillings
}
\description{
Dental fillings made with gold can vary in hardness depending
on how the metal is treated prior to its placement in the tooth.
Two factors are thought to influence the hardness:  the gold alloy and
the condensation method.  In addition, some dentists doing the work
are better at some types of fillings than others.

Five dentists were selected at random.
Each dentist prepares 24 fillings (in random order), one for each of
the combinations of method (three levels) and
alloy (eight levels).  The fillings were then
measured for hardness using the Diamond Pyramid Hardness Number
(big scores are better).

Source: Oehlert (2022), originally from Brown (1975).
}
\references{
Brown, M. B. (1975) "Exploring interactions effects in the ANOVA." \emph{Applied Statistics}
\strong{ 24}, 288-298.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
