% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FruitFlyLifespan}
\alias{FruitFlyLifespan}
\title{Fruit fly lifespans under reproductive stress}
\format{
A dataframe of 125 observations on two variables.
\describe{
\item{\code{companions}}{A factor giving the number and type of femal companions.}
\item{\code{longevity}}{The lifespan of the male fruit fly, in days.}
}
}
\usage{
FruitFlyLifespan
}
\description{
One hundred twenty five male fruit flies were randomly divided into
five groups of 25 and individually housed. In one group, males were kept by themselves.
In two groups, males were supplied with either one or eight receptive, virgin females
per day. In the final two groups, males were supplied with either one or eight
unreceptive (pregnant) femals per day. The response is the longevity (in days)
of the males.

Source: Oehlert (2022), originally from Hanley and Shapiro (1994).
}
\references{
Hanley, J. A. and S. H. Shapiro (1994). "Sexual activity and the lifespan of
male fruitflies: a dataset that gets attention." \emph{Journal of Statistics Education}
\strong{2}.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
