% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FruitPunch}
\alias{FruitPunch}
\title{Appeal of fruit punch.}
\format{
A dataframe of 18 observations on four variables.
\describe{
\item{\code{wm}}{The fraction of watermelon juice.}
\item{\code{pa}}{The fraction of pineapple juice.}
\item{\code{oj}}{The fraction of orange juice.}
\item{\code{appeal}}{The rating for the punch.}
}
}
\usage{
FruitPunch
}
\description{
A three-component fruit punch
mixture experiment, where the goal is to find
the most appealing mixture of watermelon juice,
pineapple juice, and orange juice.
Six different punches are produced, and eighteen judges are
assigned at random to the punches, three to a punch.

Source: Oehlert (2022), originally from Cornell (1985).
}
\references{
Cornell, J. A. (1985). Mixture experiments. In S. Kotz and N. Johnson (Eds.),
\emph{Encyclopedia of Statistical Sciences}, Volume 5, pp. 569-579. New York: Wiley.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
