% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GABA}
\alias{GABA}
\title{GABA release under anti-convulsant treatemant.}
\format{
A dataframe of 48 observations on four variables.
\describe{
\item{\code{Tri}}{Factor indicating presence or absence of Trifluoperazine.}
\item{\code{Dia}}{Factor indicating presence of absence of Diazepam.}
\item{\code{Cal}}{Factor indicating presence or absence of calmodulin.}
\item{\code{GABA}}{The amount of GABA released.}
}
}
\usage{
GABA
}
\description{
Anticonvulsant drugs may be effective because
they encourage the effect of the neurotransmitter
GABA (gamma-aminobutyric acid).  Calcium transport
may also be involved.  The present experiment randomly
assigned 48 rats to eight experimental conditions.  These
eight conditions are the factor-level combinations of three
factors, each at two levels.  The factors are the
anticonvulsant Trifluoperazine (brand name Stelazine)
present or absent,
the anticonvulsant Diazepam (brand name Valium)
present or absent,
and the calcium-binding protein calmodulin present
or absent.  The response is the amount of GABA released
when brain tissues are treated with 33 mM K+.

Source: Oehlert (2022), originally from de Belleroche, et al. (1982).
}
\references{
de Belleroche, J., A. Dick, and A. Wyrley-Birch (1982). "Anticonvulsants and
trifluoperazine inhibit the evoked release of GABA from cerebral cortex of
rat at different sites," \emph{Life Sciences} \strong{31}, 2875-2882.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
