% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GelStrength}
\alias{GelStrength}
\title{GABA release under anti-convulsant treatemant.}
\format{
A dataframe of 16 observations on six variables.
\describe{
\item{\code{percent.z}}{The percent of starch.}
\item{\code{percent}}{A factor version of \code{percent.z}.}
\item{\code{temp.z}}{The temperature of the gel.}
\item{\code{temp}}{A factor version of \code{temp.z}.}
\item{\code{starch}}{The source of the starch.}
\item{\code{strength}}{The strength of the gel.}
}
}
\usage{
GelStrength
}
\description{
An experiment studied the effects of starch source,
starch concentration, and temperature on the strength of
gels.  This experiment was completely randomized with
sixteen units.  There are four starch sources (adzuki
bean, corn, wheat, and potato), two starch
percentages (5\% and 7\%), and two temperatures
(22 C and 4 C).  The response is gel strength
in grams.

Source: Oehlert (2022), originally from Tjahjadi (1983).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Tjahjadi, C. (1983). \emph{Isolation and Characterization of Adzuki Bean
(Vigna angularis) Protein and Starch} Ph. D. thesis, University of Minnesota,
St. Paul, MN.
}
\keyword{datasets}
