% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GingerOil}
\alias{GingerOil}
\title{Microwave extration of essential oil from ginger}
\format{
A data frame of 16 observations on five variables.
\describe{
\item{\code{duration}}{A factor giving the duration of the experiment.}
\item{\code{wattage}}{A factor giving the wattage of the microwave.}
\item{\code{Type}}{A factor giving the preparation of the ginger.}
\item{\code{yield}}{The amount of wax oil extracted.}
}
}
\usage{
GingerOil
}
\description{
Ginger contains an essential oil that we would like
to extract without using a solvent, in this case
by using a microwave. This experiment
considers eight different treatments for extracting
this oil. These treatments are the factor/level combinations
of duration (10 or 30 minutes), wattage (288 or 640), and
preparation (crushed or sliced). The response
is the oil yield (in \%). Sixteen samples of
ginger are randomly assigned to the eight factor/level
combinations, two per combination. 

Source: Oehlert (2022), originally from Shah and Garg (2014) via
Lye (2019).
}
\references{
Lye, Leonard M. (2019) \emph{Applications of DEO in Engineering and Science:
A Collection of 2^6 Case Studies}, Leonard Lye, St. John's, Newfoundland.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Shah, Mumtaj and S. K. Garg (2014) :Application of 2^k full factorial design in
optimization of solvent-free microwave extraction of ginger essential oil,"
\emph{Journal of Engineering}, Article ID 828606, 5 pages.
}
\keyword{datasets}
