% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Graininess}
\alias{Graininess}
\title{Graininess of dried fruit pulp.}
\format{
A dataframe of 24 observations on five variables.
\describe{
\item{\code{batch}}{Factor indicating the batch of fruit pulp.}
\item{\code{temp}}{Factor indicating the temperature treatment.}
\item{\code{pH}}{Factor indicating the pH treatment.}
\item{\code{sugar}}{Factor indicating the sugar treatment.}
\item{\code{graininess}}{The graininess of the resulting product.}
}
}
\usage{
Graininess
}
\description{
Grains or crystals adversely affect the sensory qualities of foods
using dried fruit pulp.  A factorial experiment was conducted to determine
which factors affect graininess.  The factors were drying temperature
(three levels), acidity (pH) of pulp (two levels), and sugar content (two levels).
The experiment has two replications, with each replication using
a different batch of pulp.
Response is a measure of graininess

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
