% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Handkerchiefs}
\alias{Handkerchiefs}
\title{Whiteness of washed handerkerchiefs.}
\format{
A dataframe of 108 observations on six variables.
\describe{
\item{\code{tub}}{Factor indicating the tub of laundry.}
\item{\code{detergent}}{Factor indicating the detergent used.}
\item{\code{bleach}}{Factor indicating the bleach used.}
\item{\code{stain}}{Factor indicating the stain used.}
\item{\code{prewash}}{Factor indicating the prewash used.}
\item{\code{whiteness}}{The whiteness score of the washed handerchief.}
}
}
\usage{
Handkerchiefs
}
\description{
A consumer testing agency wishes to test the ability of
laundry detergents, bleaches, and prewash treatments to remove
soils and stains from fabric.  Three detergents are selected
(a liquid, an all-temperature powder, and a hot-water powder).  The two
bleach treatments are no bleach or chlorine bleach.  The three prewash
treatments are none, brand A, and brand B.  The  three stain treatments
are mud, grass, and gravy.  There are thus 54 factor-level combinations.

Each of 108 white-cotton handkerchiefs is numbered with a random code.
Nine are selected at random, and these nine are assigned at random to the
nine factor-level combinations of stain and prewash.  These nine handkerchiefs
along with four single sheets make a "tub" of wash.  This is repeated twelve
times to get twelve tubs.  Each tub of wash is assigned at random to one of
the six factor-level combinations of detergent and bleach.  After washing
and drying, the handkerchiefs are graded (in random order) for whiteness
by a single evaluator using a 1 to 100 scale, with 1 being whitest (cleanest).

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
