% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Herbicides}
\alias{Herbicides}
\title{Weed growth in soybeans with herbicides.}
\format{
A dataframe of 96 observations on four variables.
\describe{
\item{\code{location}}{Factor indicating location of the observation.}
\item{\code{treatment}}{Factor indicating the herbicide treatment.}
\item{\code{variety}}{Factor indicating the variety of soybeans.}
\item{\code{biomass}}{Biomass of weeds in kg/ha.}
}
}
\usage{
Herbicides
}
\description{
An experiment to determine how different soybean
varieties compete against weeds.  There were sixteen varieties of soybeans
and three
weed treatments: no herbicide, apply herbicide 2 weeks after
planting the soybeans, and apply herbicide 4 weeks after planting
the soybeans.  The measured response is weed biomass in kg/ha.
There were two replications of the experiment, one in St. Paul, MN,
and one in Rosemount, MN, for a total of 96 observations.

Source: Oehlert (2022), originally from Bussan (1995).
}
\references{
Bussan, A. J. (1995). \emph{Selectino for weed competitiveness among soybean
genotypes}. Master's thesis, University of Minnesota, St. Paul, MN.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
