% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{HighlyUnbalanced}
\alias{HighlyUnbalanced}
\title{A synthetic example of an unbalanced 2x2 factorial.}
\format{
A dataframe of 17 observations on three variables.
\describe{
\item{\code{A}}{Factor 1.}
\item{\code{B}}{Factor 2.}
\item{\code{y}}{A response.}
}
}
\usage{
HighlyUnbalanced
}
\description{
Synthetic data for a 2x2 factorial with cell counts
10, 1, 1, 5.

Source: Oehlert (2022).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
