% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IBD}
\alias{IBD}
\title{An incomplete block design}
\format{
A dataframe of 27 observations on three variables.
\describe{
\item{\code{block}}{Factor indicating the block.}
\item{\code{treatment}}{Factor indicating the treatment.}
\item{\code{resp}}{The response.}
}
}
\usage{
IBD
}
\description{
Nine treatments in nine blocks of size three.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
