% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ICDefects}
\alias{ICDefects}
\title{Defective integrated circuits on a wafer}
\format{
A dataframe of 54 observations on three variables.
\describe{
\item{\code{x}}{The horizontal position of the IC on the wafer.}
\item{\code{y}}{The vertical position of the IC on the wafer.}
\item{\code{good}}{A binary indicator for the status of the chip; 1=good, 0=bad.}
}
}
\usage{
ICDefects
}
\description{
The locations of integrated circuits on a silicon wafer, along
with the IC's status as good or bad.

Source: Oehlert (2022), origially from Taam and Hamada (1993).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Taam, W. and M. Hamada (1993). "Detecting spatial effects from factorial experiments:
An application from integrated-circuit manfacturing." \emph{Technometrics} \strong{35}, 149-160.
}
\keyword{datasets}
