% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IceCream}
\alias{IceCream}
\title{Consumer rating of ice cream}
\format{
A dataframe with 44 observations on three variables.
\describe{
\item{\code{gum}}{A factor indicating the type of gum in the recipe.}
\item{\code{protein}}{A factor indicating the type of non-cream protein.}
\item{\code{rating}}{The average rating for the batch of ice cream.}
}
}
\usage{
IceCream
}
\description{
Modern ice creams are not simple recipes.  Many use some
type of gum to enhance texture, and a non-cream protein source
(for example, whey protein solids).  A food scientist is trying to determine
how types of gum and protein added change a
sensory rating of the ice cream.  She runs a five by five factorial with two
replications using five gum types and five protein sources.  Unfortunately, six
of the units did not freeze properly, and these units were not rated.
Ice creams are rated by a trained panel, with higher ratings being better.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
