% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Icings}
\alias{Icings}
\title{Viscoscity of icings.}
\format{
A dataframe with 64 observations on nine variables.
\describe{
\item{\code{A}}{Level of first water pseudofactor.}
\item{\code{B}}{Level of second water pseudofactor.}
\item{\code{C}}{Level of sugar type 1.}
\item{\code{D}}{Level of sugar type 2.}
\item{\code{E}}{Level of stabilizer type 1.}
\item{\code{F}}{Level of stabilizer type 2.}
\item{\code{G}}{Level of stabilizer type 3.}
\item{\code{H}}{Level of setting agent.}
\item{\code{vicoscity}}{Coded viscoscity of the icing.}
}
}
\usage{
Icings
}
\description{
We wish to determine the tolerance of icings to
ingredient changes and variation in the preparation.
Ingredient changes are represented by factors C, D,
E, F, G, and H.  All are at two levels.  C and D
are two types of sugars; E, F, and G are three stabilizers;
and H is a setting agent.  The levels of these
factors represent changes in the amounts of these constituents
in the mix.  Variation in preparation
is modeled as the amount of water added to the product.
This has four levels and is represented as the
combinations of factors A and B.  The response we
measure is (coded) viscosity of the icing.
A quarter-fraction with 64 observations was run.

Source: Oehlert (2022), originally from Carroll and Dykstra (1958).
}
\references{
Carroll, M. B. and O. Dykstra (1958). Application of fractional factorials in a
food research laboratory. In V. Chew (Ed.), \emph{Experimental Designs in Industry},
pp. 224-234. New York: Wiley.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
