% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ImageID}
\alias{ImageID}
\title{Identification of flashed video images.}
\format{
A dataframe with 128 observations on nine variables.
\describe{
\item{\code{block}}{A factor indicating the block.}
\item{\code{A}}{Experimental factor A,  size, as -1 and 1.}
\item{\code{B}}{Experimental factor B,  shape, as -1 and 1.}
\item{\code{C}}{Experimental factor C,  color., as -1 and 1}
\item{\code{D}}{Experimental factor D,  orientation, as -1 and 1.}
\item{\code{E}}{Experimental factor E, duration, as -1 and 1.}
\item{\code{F}}{Experimental factor F, vertical, as -1 and 1.}
\item{\code{G}}{Experimental factor G, horizontal, as -1 and 1.}
\item{\code{fracID}}{The fraction (out of 15) of images identified (noticed).}
}
}
\usage{
ImageID
}
\description{
This experiment studies how image properties affect visual perception.
Each subject looks at a white computer
screen.  At random intervals averaging about 5 seconds,
we will put a small image on the screen for a very short time.
The subject is supposed to click the
mouse button when she sees an image on the screen.  The experiment
takes place
in sixteen ten-minute sessions to prevent fatigue;
during each session we present
120 images.  In fact, these are eight images repeated fifteen times
each and presented in random order.
We record
as the response the fraction of times that the mouse is clicked
for a given image type.

We wish to study 128 different images, the factorial combinations
of seven factors each at two levels: size of image, shape of
image, color of image, orientation of image, duration of image,
vertical location of image, and horizontal location of image.
We anticipate session to session variability, and we run a confounded factorial
with sixteen blocks of size eight using defining contrasts ABG, ACF, BCE, and ABCD.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
