% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Incinerator}
\alias{Incinerator}
\title{Soil Mercury from an incinerator.}
\format{
A dataframe with 11 observations on four variables.
\describe{
\item{\code{soil}}{A factor indicating the type of soil.}
\item{\code{distance}}{A factor indicting distance from the incinerator.}
\item{\code{carbon}}{The concentration of carbon.}
\item{\code{mercury}}{The concentration of mercury.}
}
}
\usage{
Incinerator
}
\description{
A study has been conducted on the environmental impact of
an industrial incinerator.  One of the concerns is the emission
of heavy metals from the stack, and one way to measure the impact
is by looking at metal accumulations in soil and seeing if
nearby sites have more metals than distant sites (presumably
due to deposition of metals from the incinerator).

Eleven sites of one hectare each (100 m by 100 m)
were selected around the incinerator.  Five sites
are on agricultural soils, while the other six
are on forested soils.  Five of the sites
were located near the incinerator (on their
respective soil types), while
the other sites were located far from the
incinerator.  At each site,
nine locations are randomly selected within the
site and mineral soil sampled at each location.  We then
measure the mercury content in each sample (mg/kg).

Complicating any comparison is the fact that heavy metals are
generally held in the organic portion of the soil, so that
a soil sample with more carbon will tend to have more heavy
metals than a sample with less carbon, regardless of the deposition
histories of the samples, soil type, etc.
For this reason, we also measure the
carbon fraction of each sample (literally the fraction of the soil sample
that was carbon).

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
