% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{InsectCatch}
\alias{InsectCatch}
\title{Insect catch in traps}
\format{
A dataframe with 15 observations on three variables.
\describe{
\item{\code{trap}}{A factor indicating the trap.}
\item{\code{period}}{A factor indicating the period the insects were collected.}
\item{\code{catch}}{The average catch for the trap type in the period.}
}
}
\usage{
InsectCatch
}
\description{
Average catch of insects in three different
kinds of traps across five periods.

Source: Snedecor and Cochran (1967).
}
\references{
Snedecor, G. W. and Cochran, W. G. (1967) \emph{Statistical Methods},
Iowa State University Press: Ames, Iowa.
}
\keyword{datasets}
