% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactionplot.R
\name{interactplot}
\alias{interactplot}
\title{Interaction plots}
\usage{
interactplot(
  x.factor,
  trace.factor,
  response,
  type = c("b", "l", "p"),
  legend = TRUE,
  trace.label = deparse(substitute(trace.factor)),
  fixed = FALSE,
  xlab = deparse(substitute(x.factor)),
  ylab = ylabel,
  lty = nc:1,
  col = 1,
  pch = c(1L:9, 0, letters),
  leg.bg = par("bg"),
  leg.bty = "n",
  confidence = 0,
  pooled = TRUE,
  sigma2 = NULL,
  df = NULL,
  jitterwidth = 0.1,
  at = NULL,
  data = NULL,
  ...
)
}
\arguments{
\item{x.factor}{A factor whose levels will form the x axis.}

\item{trace.factor}{A second factor (or interaction) whose levels will form the traces.}

\item{response}{The variable giving the response.}

\item{type}{with lines, points, or both.}

\item{legend}{Logical. Should a legend be included describing the traces?}

\item{trace.label}{Title for the legend.}

\item{fixed}{Logical. Are traces labeled in original order or
resorted by values on the right?}

\item{xlab}{x axis label.}

\item{ylab}{y axis label.}

\item{lty}{Line types for traces.}

\item{col}{Plotting color(s).}

\item{pch}{Plotting characters for traces.}

\item{leg.bg}{Parameter passed to legend().}

\item{leg.bty}{Parameter passed to legend().}

\item{confidence}{Level of confidence (between 0 and 1) for confidence
intervals on cell means.}

\item{pooled}{Logical.  Should confidence intervals be based on
a pooled estimate of variance or on individual cell variances?}

\item{sigma2}{External pooled variance for use in confidence intervals.}

\item{df}{External degrees of freedom for use in confidence intervals.}

\item{jitterwidth}{How far (horizontally) should points be spread so
that confidence intervals do not overlap?}

\item{at}{Optional vector giving horizontal positions for the cells.}

\item{data}{An  optional data frame.}

\item{...}{Other arguments passed to graphics functions.}
}
\value{
None. Called for its side effect of drawing an interaction plot.
}
\description{
Produce an interaction plot, with optional confidence limits 
on group means.
}
\details{
This function is a modification of the builtin
\code{\link[stats:interaction.plot]{stats::interaction.plot()}} function.
\code{interactplot} can plot confidence intervals for each cell mean,
with optional confidence intervals jittered for observability.
\code{interactplot} can only plot cell means, whereas 
\link[stats:interaction.plot]{stats::interaction.plot()} can plot other functions
of cell data.  

If you have multiple factors, for example, A, B, and C, and you with to plot
the combinations of B and C against A as the x factor, use \code{B:C}
as the trace factor.

You may include confidence intervals for means by setting \code{confidence}
to something between 0 and 1.  By default, intervals are based on a pooled
estimate of variance that combines variability of all the groups. Alternatively,
you may get within group variances by using \code{pooled = FALSE}.  Note,
the non-pooled estimates of error have many fewer degrees of freedom and the
confidence intervals will be correspondingly longer.  

Both pooled and unpooled estimates of error will be inappropriate if there are
additional active factors that are not reflected in the plot.  To permit
sensible plots in such cases you can directly specify an error
variance and degrees of freedom to use.

Confidence intervals will overplot each other unless they are offset 
horizontally.  The \code{jitterwidth} argument allows you to control the
amount of horizontal offset.

Other arguments should act in the same way as for \code{interaction.plot()}.
}
\examples{
 data(ZincRetention)
 interactplot(m.protein,m.zinc,retention,data=ZincRetention)
 # Can have combinations of factors as x or trace
 interactplot(m.protein,m.zinc:d.zinc,retention,data=ZincRetention)
 interactplot(m.zinc:d.zinc,m.protein,retention,data=ZincRetention)
 # add intervals while specifying error and error df
 interactplot(m.protein,d.zinc,retention,confidence=.95,sigma2=9,df=7,
      data=ZincRetention)
 # These intervals are too long because the variability due to a significant
 # third factor is included in error.
 interactplot(m.protein,d.zinc,retention,confidence=.95,
    main="These intervals are too long",data=ZincRetention)
 # at is useful for unequally spaced quantitative factors
 interactplot(m.protein,d.zinc,retention,at=c(10,15,20,30),data=ZincRetention)

}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
