% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Interchanges}
\alias{Interchanges}
\title{Speed at interchanges.}
\format{
A dataframe with 60 observations on five variables.
\describe{
\item{\code{sign}}{A factor indicating whether the interchange received the sign.}
\item{\code{timing.z}}{The timing interval for the on-ramp.}
\item{\code{timing}}{A factor version of \code{timing.z}}
\item{\code{interchange}}{A factor indicating which interchange is in use.}
\item{\code{speed}}{The average speed of cars.}
}
}
\usage{
Interchanges
}
\description{
Traffic engineers are experimenting with two ideas.  The first is
that erecting signs that say "Accident Reduction Project Area" along freeways
will raise awareness and thus reduce accidents.  Such signs may have an effect
on traffic speed.  The second idea is that metering the flow of vehicles onto
on-ramps will spread out the entering traffic and lead to an average increase
in speed on the freeway.  The engineers conduct an experiment to determine how
these two ideas affect average traffic speed.

First, twenty more-or-less equivalent
freeway interchanges
are chosen, spread well around a single metropolitan area and not too close to
each other.  Ten of these interchanges are chosen at random to get "Accident
Reduction Project Area" signs (in both directions); the other ten receive no
signs.  Traffic lights are installed on all on-ramps to meter traffic.  The
traffic lights can be turned off (that is, no minimum spacing between entering
vehicles) or be adjusted to require 3 or 6 seconds between entering vehicles.
Average traffic speed 6:30-8:30 AM and 4:30-6:30 PM will be measured at each
interchange on three consecutive Tuesdays, with our response
being the average of morning and evening speeds.
At each interchange, the three settings of the traffic
lights are assigned at random to the three Tuesdays.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
