% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Interlaboratory}
\alias{Interlaboratory}
\title{Interlaboratory comparison of bacteria in milk.}
\format{
A dataframe with 40 observations on three variables.
\describe{
\item{\code{sample}}{A factor indicating the contaminating sample.}
\item{\code{lab}}{A factor indicating the laboratory making the count.}
\item{\code{count}}{The number of colony forming units per ul.}
}
}
\usage{
Interlaboratory
}
\description{
Milk is tested after Pasteurization to assure that
Pasteurization was effective.  This experiment was conducted
to determine variability in test results between laboratories,
and to determine if the interlaboratory differences depend
on the concentration of bacteria.

Five contract laboratories are selected at random from
those available in a large metropolitan area.  Four levels
of contamination are chosen at random by choosing four samples
of milk from a collection of samples at various stages of
spoilage.
A batch of fresh milk from a dairy was obtained and split
into 40 units.  These 40 units are assigned at random to
the twenty combinations of laboratory and contamination sample.
Each unit is contaminated with 5 ml from its selected sample,
marked with a numeric code, and sent to the selected laboratory.
The laboratories count the bacteria in each sample by
serial dilution plate counts without knowing that they
received four pairs, rather than eight separate samples.
The response is colony forming units per ul.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
