% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IVP}
\alias{IVP}
\title{Irrigation, variety, and pesticide effects on yield.}
\format{
A dataframe with 24 observations on five variables.
\describe{
\item{\code{rep}}{A factor indicating the replication.}
\item{\code{pesticide}}{A factor indicating the pesticide being used.}
\item{\code{irrig}}{A factor indicating the irrigation being used.}
\item{\code{variety}}{A factor indicating the variety being used.}
\item{\code{yield}}{The crop yield.}
}
}
\usage{
IVP
}
\description{
An experiment was conducted to study the effects of irrigation,
crop variety, and aerially sprayed pesticide on grain yield.  There were
two replicates.  Within each replicate, three fields were chosen and randomly
assigned to be sprayed with one of the pesticides.  Each field was then
divided into two east-west strips; one of these strips was chosen at
random to be irrigated, and the other was left unirrigated.  Each east-west
strip was split into north-south plots, and the two varieties were
randomly assigned to plots.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
