% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{JapaneseBeetles}
\alias{JapaneseBeetles}
\title{Japanese beetles in the garden.}
\format{
A dataframe with 12 observations on three variables.
\describe{
\item{\code{bed}}{A factor indicating the garden bed.}
\item{\code{spray}}{A factor indicating the spray being used.}
\item{\code{bugs}}{The count of Japanese beetles.}
}
}
\usage{
JapaneseBeetles
}
\description{
An experiment to determine the best pesticide for Japanese Beetles.
We have six garden beds with beans, and the garden store has three different
sprays that claim to keep the beetles off the beans.
Sprays drift on the wind, so we cannot spray
very small areas. We divide each garden bed into two plots
and use a different spray on each plot.
The response is the number of Japanese beetles we find in each plot.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
