% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{KeyboardingPain}
\alias{KeyboardingPain}
\title{Hours of pain after keyboarding.}
\format{
A dataframe with 12 observations on three variables.
\describe{
\item{\code{type}}{A factor indicating the type of keyboard.}
\item{\code{KbdHours}}{The hours spent keyboarding.}
\item{\code{PainHours}}{The the subjective hours of pain.}
}
}
\usage{
KeyboardingPain
}
\description{
A company wishes to choose
an ergonomic keyboard for its computers to reduce
the severity of repetitive motion disorders (RMD) among
its staff.  Twelve staff known to have
mild RMD problems are randomly assigned to three keyboard
types.  The staff keep daily logs of the amount of
time spent keyboarding and their subjective assessment
of the RMD pain.  After 2 weeks, we get the total number
of hours spent keyboarding and the total number of hours
in RMD pain.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
