% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Laminates}
\alias{Laminates}
\title{Laminate lengths.}
\format{
A dataframe of 56 observations on three variables.
\describe{
\item{\code{laminate}}{A factor indicating the laminate.}
\item{\code{location}}{A factor indicating the location on the laminate.}
\item{\code{length}}{The coded length of the substrate.}
}
}
\usage{
Laminates
}
\description{
Implantable pacemakers contain a small circuit board called
a substrate.  Multiple substrates are made as part of a single
"laminate."  In this experiment, seven laminates are
chosen at random. We choose eight substrate locations and
measure the length of the substrates at those eight locations
on the seven substrates.  The response is a coded length
(10,000 x [response - 1.45]).

Source: Oehlert (2022),  data  from Todd Kerkow (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
