% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LaserTransmission}
\alias{LaserTransmission}
\title{Percent laser transmission through PVC}
\format{
A dataframe of 90 observations on four variables.
\describe{
\item{\code{thickness.z}}{The thickness of the PVC in mm.}
\item{\code{thickness}}{A factor version of \code{thickness.z}.}
\item{\code{sanding}}{A factor indicating the sanding treatment.}
\item{\code{transmission}}{The percentage of light transmitted through the PVC.}
}
}
\usage{
LaserTransmission
}
\description{
Ninety pieces of clear PVC are assigned at random to one of
18 treatments, five pieces per treatment. A low power laser is aimed
at the PVC, and the response is the percentage of light that is transmitted
through the PVC to a receptor on the other side. The 18 treatments are
the combinations of thickness (1.57, 3.18, 4.78, 6.35, 9.53, or 12.7 mm) and
surface treatment (none, front sanded, front and back sanded). We expect
that greater thickness and more sanding (roughness in this case) will lead
to reduced light transmission.

Source: Oehlert (2022),  data  from J. Van de Ven (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
