% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LeafSprings}
\alias{LeafSprings}
\title{Free height of leaf springs.}
\format{
A dataframe of 48 observations on six variables.
\describe{
\item{\code{B}}{A variable indicating furnace temperature (-1,1).}
\item{\code{C}}{A variable indicating heating time (-1,1).}
\item{\code{D}}{A variable indicating transfer time (-1,1).}
\item{\code{E}}{A variable indicating hold-down time (-1,1).}
\item{\code{O}}{A variable indicating quench temperature (-1,1).}
\item{\code{height}}{The free height.}
}
}
\usage{
LeafSprings
}
\description{
Data
from a quality experiment on the manufacture of leaf springs
for trucks.  The free height should be as close to 8 inches
as possible, with minimum variation.   There are four
inner noise factors, each at two levels: furnace temperature (B),
heating time (C), transfer time (D), and hold-down time (E).
There was one outer noise factor: quench oil temperature (O).
The design is a 2^(5-1) with three replications.  Analyze
as a 2^(4-1) design in the inner
noise factors with six replications, because quench-oil temperature
is not easily controlled in factory conditions.

Source: Oehlert (2022), originally from Pignatiello and Ramberg (1985).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Pignatiello, J. J. and J. S. Ramberg (1991). "Top ten triumphs and tragedies
of Genichi Taguchi," \emph{Quality Engineering} \strong{4}, 211-225.
}
\keyword{datasets}
