% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LethalDosage}
\alias{LethalDosage}
\title{Guinea pig cardiac relaxant lethal dosages}
\format{
A dataframe of 40 observations on two variables.
\describe{
\item{\code{drug}}{A factor indicating the drug used.}
\item{\code{dosage}}{The lethal dosage.}
}
}
\usage{
LethalDosage
}
\description{
Forty guinea pigs are randomly assigned to four different cardiac
relaxants, ten animals per drug.  Anesthetized animals are infused with the drug
until death occurs, with response being the total lethal dose. Small dosages are
more effective.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
