% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcontrast.R
\name{linear.contrast}
\alias{linear.contrast}
\alias{linear.contrast.default}
\title{Linear contrasts for linear models}
\usage{
linear.contrast(
  lmodel,
  term,
  contr.coefs,
  confidence = 0.95,
  joint = FALSE,
  byvar = NULL,
  contr.labels = NULL,
  allpairs = FALSE,
  controlpairs = 0,
  termname = NULL,
  bonferroni = FALSE,
  scheffe = FALSE,
  returnedf = FALSE,
  exactBF = FALSE,
  ...
)

\method{linear.contrast}{default}(
  lmodel,
  term,
  contr.coefs,
  confidence = 0.95,
  joint = FALSE,
  byvar = NULL,
  contr.labels = NULL,
  allpairs = FALSE,
  controlpairs = 0,
  termname = NULL,
  bonferroni = FALSE,
  scheffe = FALSE,
  returnedf = FALSE,
  exactBF = FALSE,
  ...
)
}
\arguments{
\item{lmodel}{A model object from \code{lm}, \code{lmer}, \code{lme}, or \code{bglmm}.}

\item{term}{A term (variable name) in the linear model.}

\item{contr.coefs}{A real vector or matrix contrast coefficients, 
each column summing to zero, with one row for each level of the factor \code{term}.}

\item{confidence}{The level of confidence for interval estimates of contrasts.}

\item{joint}{Logical flag for whether to compute an overall assessment for the
situation where all of the contrasts are zero. This is an F-test in the non-Bayesian
case and a (exact, not approximate) Bayes factor for objects of class \code{bglmm}
(from \code{bcfcdae::bglmm()}).}

\item{byvar}{If we have a single contrast in one variable, we may optionally evaluate 
it at each level of a second variable, called the by variable.}

\item{contr.labels}{An optional character vector of labels for the contrasts.}

\item{allpairs}{A logical flag indicating whether to make all pairwise comparisons.}

\item{controlpairs}{The number of the control treatment if pairwise comparisons with 
the control are desired.}

\item{termname}{A character variable giving the name of the treatment factor;
an alternative spacification to \code{term}.}

\item{bonferroni}{A logical flag indicating whether confidence intervals and p-values should
be adjusted with a Bonferroni correction; not applicable for an object of class \code{bglmm}.}

\item{scheffe}{A logical flag indicating whether confidence intervals and p-values should
be adjusted with a Scheffe correction; not applicable for an object of class \code{bglmm}.}

\item{returnedf}{A logical flag indicating whether to return the error degrees of freedom;
not applicable for an object of class \code{bglmm}.}

\item{exactBF}{A logical flag for whether to use an approximate Bayes factor,
which is fast, or an exact Bayes factor, which is slow. Applicable only for
an object of class \code{bglmm} (from \code{bcfcdae::bglmm()}).}

\item{...}{Other parameters passed to \code{bcfcdae::bglmm()}.}
}
\value{
If \code{lmodel} is of class \code{bglmm} (from \code{bcfcdae::bglmm()}), the value is a
matrix with one row for each contrast and columns for posterior estimate,
posterior standard deviation, lower and upper endpoints for a posterior
interval, and (approximate) Bayes factor for the full model relative to the model
with the contrast
constrained to be zero.
Otherwise, the value is a matrix with one
row for each contrast and columns for estimate, standard error,
t-value (null hypothesis of zero), p-value, and lower and upper
confidence limits. When \code{joint} is true, the value is a list
with components for the basic output as well as the overall Bayes
factor (for \code{bglmm} objects) or F
and optionally its error degrees of freedom.
}
\description{
Do inference for linear contrasts of factor treatment coefficients in linear models.
}
\details{
This function is used for computing linear contrasts in the
levels of a factor. The confidence level for the intervals can be set via the
the \code{confidence} argument. This also applies in the Bayesian context.

A single contrast is computed when \code{contr.coefs} is a vector
summing to zero with one element for each level of the contrast
variable \code{term}.  Multiple contrasts in \code{term} are
represented by a matrix with one row for each level of \code{term},
and each column summing to zero.

Either \code{allpairs} or \code{controlpairs} can be used in place of
specifying contrast coefficients.  In the first case, all pairwise
comparisons will be computed.  In the second case, if controlpairs is
the number of a treatment, you will get the pairwise differences
of the other treatments with this treatment.

A by variable allows you compute a contrast in \code{variable} separately
for each level of \code{byvar}.  If you use a by variable,
that variable and its interaction with the contrast variable must
be explicit terms in the model. Also, by variables only work with single contrasts.
For \code{bglmm} objects, no Bayes factors are reported when you use a by
variable.

By default, contrasts are simply labeled by number. You may set
your own labels with the \code{contr.labels} argument, which should
be a character vector with one element for each contrast.

In a non-Bayesian setting, it is possible to compute a joint F-test of the null hypothesis that
all of the contrasts are zero.  The numerator degrees of freedom
for this test will be at most the number of contrasts you compute but
can be less if the contrasts are not linearly independent.

In a non-Bayesian setting, you may specify either Bonferroni
or Scheffe corrections for multiple comparisons. The Bonferroni
correction will adjust for the number of contrasts in this function
call, whereas the Scheffe correction will adjust for all possible
contrasts, including contrasts suggested by the data.

Getting estimated error degrees of freedom 
for the output of \code{lmer()} is done with the Kenward and Roger method.

The approximate Bayes factor for objects of class \code{bglmm} employs a BIC-style approximation. It should
be more accurate as the sample size increases. Beyond using BIC instead of
the Bayes factor, it also employs a normal approximation to the likelihood
rather than maximizing a constrained likelihood.

\emph{Note:} \code{linear.contrast} should not be confused with the
\code{contrast} option in R. The latter establishes the basis or
parameterization for a factor, which might, or might not, be a
contrast in the sense that \code{linear.contrast} uses the term.
}
\examples{
data(ConcreteFaulting)
# fit a linear model
fit <- lm(faulting~dowels*base*subgrade*wcratio,data=ConcreteFaulting)
# compare first two levels of base
linear.contrast(fit,base,c(1,-1,0)) 
# compare first two levels of base separately by levels of subgrade; overall test
linear.contrast(fit,base,c(1,-1,0),byvar="subgrade",joint=TRUE)
# all pairwise comparisons of base; HSD default
linear.contrast(fit,base,allpairs=TRUE)
# all pairwise comparisons of base with custom labels
linear.contrast(fit,base,allpairs=TRUE,contr.labels=c("1 vs 2","1 vs 3","2 vs 3"))
# all pairwise comparisons of base using Bonferroni
linear.contrast(fit,base,allpairs=TRUE,bonferroni=TRUE)
}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
