% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Lithium}
\alias{Lithium}
\title{Serum lithium concentrations}
\format{
A dataframe of 24 observations on four variables.
\describe{
\item{\code{treatment}}{A factor indicating level of treatment.}
\item{\code{period}}{A factor indicating the time period.}
\item{\code{subject}}{A factor indicating the subject.}
\item{\code{concentration}}{The serum lithium concentration.}
}
}
\usage{
Lithium
}
\description{
We study the blood concentrations of lithium
12 hours after administering lithium carbonate,
using either a 300 mg capsule, 250 mg capsule,
450 mg time delay capsule, or 300 mg solution.  There are
twelve subjects, each of whom will be used twice, 1 week apart.  We
anticipate that the responses will be different in the
second week, so we block on subject and week.
The response is the serum lithium level (micro equivalents per liter).

Source: Oehlert (2022), originally from Westlake (1974).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Westlake, W. J. (1974). "The use of balanced incomplete block designs in
comparative bioavailability trials." \emph{Biometrics} \strong{30}, 319-327.
}
\keyword{datasets}
