% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{LoafVolume}
\alias{LoafVolume}
\title{Volume of loaves as a function of flours used.}
\format{
A dataframe of 18 observations on six variables.
     \describe{
     \item{\code{day}}{A factor giving the day for the loaf.}
     \item{\code{x1}}{The fraction of flour 1 used.}
     \item{\code{x2}}{The fraction of flour 2 used.}
     \item{\code{x3}}{The fraction of flour 3 used.}
     \item{\code{x4}}{The fraction of flour 4 used.}
     \item{\code{volume}}{The volume of the loaf.}
     }
}
\usage{
LoafVolume
}
\description{
Millers want to make bread flours that bake into large loaves.
They need to mix flours from four varieties of wheat, so
they run an experiment with different mixtures and measure
the volume of the resulting loaves (ml/100 g dough).  The
experiment was performed on 2 separate days.

     Source: Oehlert (2022), originally from Draper, et al. (1993).
}
\references{
Draper, N. R., S. M. Lewis, P. W. M. John, P. Prescott, A. M. Dean, and
M. G. Tuck (1993). "Mixture designs for four components in orthogonal blocks,"
\emph{Technometrics} \strong{35}, 268-276.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
