% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MealyBugs}
\alias{MealyBugs}
\title{Changes in Mealy Bug counts on cycads after treatment}
\format{
A dataframe of 30 observations on four variables.
     \describe{
     \item{\code{treatment}}{The treatment used on the branch.}
     \item{\code{plant}}{A factor indicating the plant from which each branch was taken.}
     \item{\code{change}}{The change (before-after) for the patch. }
\item{\code{branch}}{A factor indicating the branch (plant by treatment).}
     }
}
\usage{
MealyBugs
}
\description{
Cycads (plants that look vaguely like palms)
can be infested with mealybug, and we test
three treatments to control the bugs: water (a control), horticultural
oil (a standard no-mammalian-toxicity pesticide), and
fungal spores in water (\emph{Beauveria bassiana\/}, a fungus that
grows exclusively on insects).  Five infested cycads are removed
to a testing area.  Three branches are randomly chosen on each cycad,
and two 3 cm by 3 cm patches are marked on each branch;
the number of mealybugs in these patches is noted.
The three branches on each cycad are randomly assigned to the three treatments.
After three days, the patches are counted again, and the response
is the change in the number of mealybugs (before - after).
Data here are the average responses on each branch.

     Source: Oehlert (2022), originally from Scott Smith (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
