% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Melatonin}
\alias{Melatonin}
\title{Hamster testes weight after melatonin treatment}
\format{
A dataframe of 4 observations on three variables.
\describe{
\item{\code{dose}}{The dose of melatonin used in each treatment group.}
\item{\code{ave.weight}}{Average testes weight in each treatment group.}
\item{\code{sd.weight}}{Standard devaition of testes weight in each treatment group.}
}
}
\usage{
Melatonin
}
\description{
Thirty-two male hamsters were randomly assigned to one of four
doses of melatonin daily 1 hour prior to lights out for 12 weeks. Response is the
paired testes weight in mg. Data are summaries (mean and standard deviation) of the
responses for the 8 animals in each treatment group.

Source: Oehlert (2022), originally from Rollag (1982).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Rollag, M. D. (1982) "Ability of tryptophan derivatives to mimic melatonin's
action upon the Syrian hamster reproductive system," \emph{Life Sciences}
\strong{31}, 2699-2707.
}
\keyword{datasets}
