% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MentosGeyser}
\alias{MentosGeyser}
\title{Soda remaining after adding Mentos(R)}
\format{
A data frame with 36 observations on six variables.
\describe{
\item{\code{type}}{The type of soft drink (DS, D7, or 7).}
\item{\code{elapsed}}{A factor indicating the time elapsed (minutes) since opening the bottle.}
\item{\code{elapsed.z}}{A numeric variable indicating the time elapsed (minutes) since opening the bottle.}
\item{\code{tablets}}{A factor indicating the number of Mentos(R) added.} 
\item{\code{tablets.z}}{A numeric variable indicating the number of Mentos(R) added.} 
\item{\code{remaining}}{The volume of soft drink remaining (mL).}
}
}
\usage{
MentosGeyser
}
\description{
This experiment explores the effects of elapsed time
after opening a soft drink and number of Mentos(R)
added on the volume of soft drink remaining in the bottle.
There are 36 .5L bottles of soft drink; 12 each of
Diet Sunkist(R),
Diet 7up(R),
and 7up(R).
The 12 bottles of each
type are randomly assigned to the six factor/level
combinations of time (0 or 20 minutes after opening) and number of tablets
(1, 4 or 7 Mentos(R)), two bottles to each combination. 
In random order, each bottle was given its assigned
treatment, and the amount (mL) of beverage remaining 30
seconds after the tablet drop was measured as response.

Source Oehlert (2022), originally from T. Steichen (pers. comm.)
}
\keyword{datasets}
