% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Milfoil}
\alias{Milfoil}
\title{Milfoil biomass under weevil treatments.}
\format{
A dataframe of 32 observations on five variables.
\describe{
\item{\code{tank}}{Factor denoting the tank for the current observation.}
\item{\code{A}}{Factor indicating the parental diet (Eurasion or northern).}
\item{\code{B}}{Factor indicating the hatched diet (Eurasion or northern).}
\item{\code{C}}{Factor indicating the raised diet (Eurasion or northern).}
\item{\code{biomass}}{Biomass of Eurasion milfoil after four months.}
}
}
\usage{
Milfoil
}
\description{
Eurasian water milfoil is a nonnative plant that is
taking over many lakes in Minnesota and driving out the native
northern milfoil.  However, there is a
native weevil (an insect) that eats milfoil and may be
useful as a control.  We wish to investigate how eight
treatments affect the damage the weevils do to Eurasian milfoil.
The treatments are the
combinations of whether a weevil's parents were raised
on Eurasian or northern, whether the weevil was hatched
on Eurasian or northern, and whether the weevil
grew to maturity on Eurasian  or northern.

We have eight tanks (big aquariums), each of which is
subdivided into four sections.  The subdivision is accomplished
with a fine mesh that lets water through, but not weevils.
The tanks are planted with equal amounts of Eurasian milfoil.
We try to maintain uniformity between tanks, but there will
be some tank to tank variation due to differences in light
and temperature.  The tanks are planted in May, then
weevils are introduced. In September, milfoil
biomass is measured as response.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
