% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MilkChiller}
\alias{MilkChiller}
\title{Sensory ratings of milk using different chilling post-Pasteurization.}
\format{
A dataframe of 16 observations on five variables.
\describe{
\item{\code{block}}{Factor denoting the block for the current observation.}
\item{\code{A}}{Factor indicating spacing.}
\item{\code{B}}{Factor indicating temperature.}
\item{\code{C}}{Factor indicating flow rate.}
\item{\code{response}}{Sensory score for the milk.}
}
}
\usage{
MilkChiller
}
\description{
Milk is chilled immediately after Pasteurization.
The goal in chiller design is to get high flow at low
capital and operating costs while still chilling the milk quickly
enough to maintain sensory qualities.
Basic chiller design is a set of refrigerated plates
over which the hot milk is pumped.  We are investigating
the effect of the spacing between the plates (two levels),
the temperature of the plates (two levels), and the flow rate
of the milk (two levels) on the perceived quality of the resulting
milk.  There is a fresh batch of raw milk each day, and we expect
batch to batch differences in quality.  Because of the time involved in
modifying the chiller, we can use at most four factor-level
combinations in a day.

This experiment uses two replicates, confounding ABC and
BC in the two replicates.
The processed milk is judged daily by a trained expert who
is blinded to the treatments used.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
