% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MilkFiltration}
\alias{MilkFiltration}
\title{Filtration times for raw milk.}
\format{
A dataframe of 48 observations on seven variables.
\describe{
\item{\code{farm}}{Factor denoting the farm for the current observation.}
\item{\code{order}}{Factor indicating whether this is a first, second, or third filtration.}
\item{\code{weight}}{Factor giving the filter weight for the current observation.}
\item{\code{loft}}{Factor giving the loft height for the current observation.}
\item{\code{bot.bond}}{Factor giving the bonding solution for the bottom suface.}
\item{\code{top.bond}}{Factor giving the bonding solution for the top suface.}
\item{\code{filt.time}}{Filtration time.}
}
}
\usage{
MilkFiltration
}
\description{
Milk can be strained through filter disks to remove dirt and debris.
Filters are made by surface-bonding fiber webs to both sides
of a disk.  This experiment is concerned with how the construction
of the filter affects the speed of milk flow through the filter.

We have a 2^4 factorial structure for the filters.  The factors
are fiber weight (normal or heavy), loft (thickness of the filter,
normal or low), bonding solution on bottom surface (A or B),
and bonding solution on top surface (A or B).

We pour a measured amount of milk through
the disk and record the filtration time as the response.  We
expect considerable variation from farm to farm, so we block on
farm.  We also expect variation from milking to milking, so
we want all measurements at one farm to be done at a single milking.
However, only three filters can be satisfactorily used at a single milking.

Sixteen farms were selected.  At each farm there will be three
strainings at one milking, with the milk strained first with
one filter, then a second, then a third.  Each treatment will
be used three times in the design: once as a first filter, once
as second, and once as third. Response is filtration time in seconds.

Source: Oehlert (2022), originally from Connor (1958).
}
\references{
Connor, W. S. (1958). "Experiences with incomplete block designs." In V. Chew (Ed.),
\emph{Experimental Designs in Industry}, pp. 193-206. New York: Wiley.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
