% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{MilkProduction}
\alias{MilkProduction}
\title{Milk production of cows under different feed regimes.}
\format{
A dataframe of 54 observations on five variables.
\describe{
\item{\code{cow}}{Factor giving the cow for the current observation.}
\item{\code{period}}{Factor giving the time period for the current observation.}
\item{\code{treatment}}{Factor giving the treatment for the current observation.}
\item{\code{previous}}{Factor giving the treatment for the previous time period for this cow.}
\item{\code{milk}}{Milk production in pounds per six weeks.}
}
}
\usage{
MilkProduction
}
\description{
This experiment studies the effect of feed type on
milk production in cows.  There is
large cow to cow variation in production, so each cow receives all
treatments.
Milk production for a given cow also tends to decrease during
any given lactation, so time period is also blocked.    The treatments
are roughage, limited grain, and full grain.
The response is the milk production during the
six week period the cow is on a given feed.

Source: Oehlert (2022), originally from Cochran, Autrey, and Cannon (1941),
via Bellavance and Tardif (1995).
}
\references{
Bellavance, F. and S. Tardif (1995). "A nonparametric approach to the analysis of
three-treatment three-period crossover designs," \emph{Biometrika} \strong{82},
865-875.

Cochran, W. G, K. M. Autrey, and C. Y. Autrey (1941). "A double change-over design
for dairy cattle feeding experiments," \emph{Journal of Dairy Science} \strong{24}, 937-951.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
