% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasse.R
\name{MixedTools}
\alias{MixedTools}
\alias{mixed.ems}
\alias{ems}
\alias{mixed.hasse}
\alias{Hasse}
\alias{mixed.tests}
\alias{mixed.power}
\alias{power}
\title{Tools for a classical approach to balanced mixed-effects factorials.}
\usage{
mixed.ems(formula, levels, random = NULL, restricted = TRUE, grid = FALSE, ...)

mixed.hasse(
  formula,
  levels,
  random = NULL,
  cex = 1,
  vert.off = 0.2,
  lwd = 1,
  xeps = 0.1,
  max.depth = 10,
  show.super = TRUE,
  show.sub = TRUE,
  alt.names = NULL,
  ...
)

mixed.tests(
  formula,
  levels,
  random = NULL,
  restricted = TRUE,
  grid = FALSE,
  ...
)

mixed.power(
  formula,
  levels,
  elements,
  random = NULL,
  restricted = TRUE,
  alpha = 0.05,
  ...
)
}
\arguments{
\item{formula}{A one-sided formula (~terms) describing the factorial model.}

\item{levels}{An integer vector giving the levels of the factors and replication
in the model.}

\item{random}{A (possibly NULL) character vector giving the names of random factors.}

\item{restricted}{A logical flag indicating whether restricted model assumptions
(zero sums across levels of fixed factors) should be used for mixed terms; default
is TRUE.}

\item{grid}{A logical flag indicating whether a tabular form of the results should
be returned; see Details. Default is FALSE.}

\item{...}{Other parameters. These are currently ignored, allowing the user to employ
the same arguments in all of these utilities for classical mixed effects.}

\item{cex}{Character size in Hasse diagrams; default is 1.7.}

\item{vert.off}{Offset for how far line segments should end from plotting position
of nodes in Hasse diagram; default is .2 (i.e., 20 percent of inter-level distance).}

\item{lwd}{Width of lines in Hasse diagram; default is 2.}

\item{xeps}{Larger values squeeze the Hasse diagram into the center of the plot. Should be between 0
and .5; default is .1.}

\item{max.depth}{Maximum depth for the Hasse diagram.}

\item{show.super}{Include superscript on Hasse diagram.}

\item{show.sub}{Include subscript on Hasse diagram.}

\item{alt.names}{A named list of character strings giving alternative labels for
for the Hasse diagram nodes corresponding to the component names.}

\item{elements}{A list of named components giving the values of the "representative
elements", with the name of the component indicating the effect.
For a random effect, its element is its variance. For a fixed effect,
its element is the sum of the squared effects across all levels divided by degrees
of freedom. Unspecified elements are assumed to be zero. Residual variance is named "Error".}

\item{alpha}{Significance level to be used in computing power; default is .05.}
}
\value{
\code{mixed.ems} computes the expected mean square for each term in the model as a vector
of character strings. If \code{grid==TRUE}, the result is a list including a
square integer matrix with
one row/column for each effect in the model. Element i,j (row, column) gives the multiplier
for the contribution of the jth effect to the ems of the ith effect.

\code{mixed.tests} returns the (approximate) F-test for each term in
the model as a vector of character strings of the form "MS[term1]/MS[term2]",
"(MS[term1]+MS[term2])/(MS[term3]+MS[term4])", etc. If \code{grid==TRUE},
the result is a list including a matrix with one row for every test and one column
for every MS. Element (i,j) (row, column) is 1 if MS j appears in the numerator of the
test for effect i and -1 if it appears in the denominator.

\code{mixed.power} returns a matix with one row for each test and columns for power and for expected
values and (Satterthwaite approximate) degrees of freedom of the numerator and denominator
of the (approximate) F test.

\code{mixed.hasse} has no return value and is used for the side effect of plotting the
Hasse diagram.
}
\description{
Classical analysis of balanced mixed-effects factorials begins by computing
the ANOVA as if all effects were fixed, calculating the expected mean squares
of the terms in the ANOVA acknowledging the random nature of some factors,
and creating (sometimes approximate) F-tests for effects and interactions based
on the expected mean squares. This has been largely supplanted by modern REML
approaches, but some classical tools are still useful, particularly in regard
to power and understanding the design. Tools here compute expected mean squares,
display the form of (approximate) F-tests, draw Hasse diagrams, and compute
power.
}
\details{
These functions assume \strong{balanced} designs. They give incorrect
results for unbalanced designs, although many of the insights may be similar.

For a k-factor design, the \code{levels} argument must have k+1 elements, representing
the number of levels of the k factors and the number of replications. The factors must
be entered in the same order as they appear in the R model. This order is not always
obvious (R rearranges terms), so looking at \code{attr(terms(formula),"factors")}
to see the order of rows can be helpful.

Contributions to EMS
from random effects are shown as "V[effect]", meaning the variance component for that
effect. Contributions from fixed effects are shown as "Q[effect]", meaning the sum of
the squared effects divided by degrees of freedom for the effect (a variance of a sort).
For example, for a two-factor interaction of factors with a and b levels, Q would
be the sum of the ab squared effects divided by (a-1)(b-1). This calculation should
also be used when specifying the
"representative element" for a non-null effect in \code{mixed.power} via the \code{element}
argument.

Interactions such as a:b are not valid names in R, so you will need to quote the name
when creating the argument for \code{elements}. For example, \code{elements = list(a=1, b=2, "a:b"=6, 
Error=10)}.
}
\examples{
 # These examples are for three-factor models with factors A, B, and C having
 # 3, 5, and 7 levels respectively. There are n=2 replications at each ABC
 # combination.
 #
 # EMS for fully crossed model with A and B random.
 mixed.ems(~A*B*C,levels=c(3,5,7,2),random=c("A","B")) # restricted is default
 mixed.ems(~A*B*C,levels=c(3,5,7,2),random=c("A","B"),restricted=FALSE)
 mixed.ems(~A*B*C,levels=c(3,5,7,2),random=c("A","B"),grid=TRUE) # return grid form
 # nest B in A
 mixed.ems(~(A/B)*C,levels=c(3,5,7,2),random=c("A","B"))
 # F tests for test models
 mixed.tests(~A*B*C,levels=c(3,5,7,2),random=c("A","B")) # restricted is default
 mixed.tests(~A*B*C,levels=c(3,5,7,2),random=c("A","B"),restricted=FALSE)
 mixed.tests(~(A/B)*C,levels=c(3,5,7,2),random=c("A","B"))
 # Hasse diagrams
 mixed.hasse(~A*B*C,levels=c(3,5,7,2),random=c("A","B"))
 # nest B in A
 mixed.hasse(~(A/B)*C,levels=c(3,5,7,2),random=c("A","B"))
 # just the skeleton diagram
 mixed.hasse(~(A/B)*C,levels=c(3,5,7,2),random=c("A","B"),
   show.sub=FALSE,show.super=FALSE)
 # power
 mixed.power(~A*B*C,levels=c(3,5,7,2),random=c("A","B"),
    elements=list(A=2,C=1,Error=5))
 # AC variability reduces power for A and C
 mixed.power(~A*B*C,levels=c(3,5,7,2),random=c("A","B"),
    elements=list(A=2,C=1,"A:C"=2,Error=5))

}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
