% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{model.effects}
\alias{model.effects}
\alias{model.effects.default}
\title{Get all model effects (including redundant effects)}
\usage{
model.effects(lmodel, termname, data = NULL)

\method{model.effects}{default}(lmodel, termname, data = NULL)
}
\arguments{
\item{lmodel}{A model object of class \code{lm}, \code{lme}, \code{lmerMod}, or
\code{bglmm}.}

\item{termname}{A character string giving the name of a term in the model.}

\item{data}{An optional data frame.}
}
\value{
A vector or matrix giving the complete set of term effects (coefficients).
}
\description{
Construct vector or table of all (generalized) linear model coefficients,
filling in redundant effects by exploting zero sum constraints.
}
\details{
This function only works for models that were fit using the
\code{"contr.sum"} model contrasts. These models have effects that add
zero across any subscript. Fittting functions, e.g. \code{lm}, leave out
the redundant coefficients. This function merely computes the redundant
coefficients by enforcing the zero sum constraint.
}
\examples{
data(SproutingBarley)
# transform response
barleyfit <- lm(sqrt(sprouting)~water*weeks,data=SproutingBarley)
coef(barleyfit) # only non-redundant coefficients get fit
model.effects(barleyfit,water)
model.effects(barleyfit,weeks)
model.effects(barleyfit,water:weeks)
\dontrun{
model.effects(barleyfit,weeks:water) # won't find it
}


}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
