% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Nonstarters}
\alias{Nonstarters}
\title{Free amino acids in cheese with nonstarter bacteria.}
\format{
A data frame of twelve observations on three variables.
\describe{
\item{\code{r50}}{Factor indicating presence or absence of strain r50.}
\item{\code{r21}}{Factor indicating presence or absence of strain r21.}
\item{\code{days.z}}{The number of days of aging.}
\item{\code{days}}{A factor version of \code{days.z}.}
\item{\code{cheese}}{A factor indicating the cheese being studied (replication).}
\item{\code{freeAA}}{The amino acids measured after ripening.}
}
}
\usage{
Nonstarters
}
\description{
Cheese is made by bacterial fermentation of Pasteurized
milk.  Most of the bacteria are purposefully added to
do the fermentation; these are the starter cultures.  Some
"wild" bacteria are also present in cheese; these are
the nonstarter bacteria.  One hypothesis is that nonstarter
bacteria may affect the quality of a cheese, so that
otherwise identical cheese making facilities produce
different cheeses due to their different indigenous
nonstarter bacteria.

Two strains of nonstarter bacteria were isolated at
a premium cheese facility: R50-10 and R21-2.  We will
add these nonstarter bacteria to cheese to see if they
affect quality.  Our four treatments will be control,
addition of R50, addition of R21, and addition of a
blend of R50 and R21. Twelve cheeses are made, three
for each of the four treatments, with the treatments
being randomized to the cheeses.  Each cheese is then
divided into four portions, and the four portions for
each cheese are randomly assigned to one of four aging
times: 1 day, 28 days, 56 days, and 84 days.  Each portion
is measured for total free amino acids (a measure of
bacterial activity) after it has aged for its specified
number of days.

Source: Oehlert (2022), data from P. Swearingen (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
