% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{NotchedBoards}
\alias{NotchedBoards}
\title{Strength of notched and un-notched boards}
\format{
A dataframe of 91 observations on two variables.
\describe{
\item{\code{shape}}{A factor indicating the treatment applied to the board (notched or uniform).}
\item{\code{power}}{The breaking strength in pounds.}
}
}
\usage{
NotchedBoards
}
\description{
Twenty-six boards are cut to dimension 26 inches, by 2.5 inches
by .75 inches. Thirteen of the boards are randomly selected, and these boards
are planed to a uniform thickness of .625 inches. The remaining thirteen
boards have a notch cut in the center that is 1 inch wide and
.125 inch deep (that is, .625 inch of wood remains under the notch). Each
board is then supported at the ends and pressure is applied in the center
until the board fails (center-point loading at .1 in/minute across a span
of 24 inches). The response is the breaking strength of the boards, in pounds.

Source: Oehlert (2022),  data  from D. Shmulsky (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
