% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{OdorIntensity}
\alias{OdorIntensity}
\title{Odor intensities of various formulations.}
\format{
A dataframe of 32 observations on six variables.
\describe{
\item{\code{block}}{Factor denoting the block for the current observation.}
\item{\code{A}}{Factor indicating the ratio of component 1 to component 5.}
\item{\code{B}}{Factor indicating the ratio of component 2 to component 5.}
\item{\code{C}}{Factor indicating the ratio of component 3 to component 5.}
\item{\code{D}}{Factor indicating the ratio of component 4 to component 5.}
\item{\code{concentration}}{Threshold concentration (scaled).}
}
}
\usage{
OdorIntensity
}
\description{
Individuals perceive odors at different intensities.
We have a procedure that allows us to determine the concentration
of a solution at which an individual first senses the odor (the
threshold concentration).  We would like to determine how
the threshold concentrations vary over sixteen solutions.
However, the threshold-determining
procedure is time consuming and any individual judge can
only be used to find threshold concentrations for four solutions.

Each solution is a combination of five compounds in various ratios.
The sixteen solutions are formed by manipulating four factors, each at
two levels.  Factor 1 is the ratio of the concentration
of compound 1 to the concentration of compound 5.  Factors
2 through 4 are are similar.

We have eight judges.  Two judges are assigned at random to each
of the solution sets [(1), bc, abd, acd],
[a, abc, bd, cd],
[ab, ac, d, bcd], and
[b, c, ad,abcd].  We then determine the threshold
concentration for the solutions for each judge.
The threshold concentrations are normalized by dividing
by a reference concentration.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
