% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Oleoresin}
\alias{Oleoresin}
\title{Oleoresin collected under acid treatments.}
\format{
A dataframe of 28 observations on four variables.
\describe{
\item{\code{acid.z}}{A numerical variable giving the acid percent strength.}
\item{\code{acid}}{A factor version of \code{acid.z}.}
\item{\code{tree}}{A factor indicating the tree used.}
\item{\code{yield}}{The amount of resin collected (grams).}
}
}
\usage{
Oleoresin
}
\description{
The oleoresin of trees is obtained by cutting a tapping
gash in the bark and removing the resin that collects
there.  Acid treatments can also improve collection.
In this experiment, four trees (\emph{Dipterocarpus kerrii})
will be tapped seven times each.  Each of the tappings
will be treated with a different strength of
sulfuric acid (0, 2.5, 5, 10, 15, 25, and 50\% strength),
and the resin collected from each tapping is
the response (in grams).

Source: Oehlert (2022), originally from Bin Jantan, Bin Ahmad,
and Bin Ahmad (1987).
}
\references{
Bin Jantan, I., A. S. Bin Ahmad, and A. R. Bin Ahmad (1987)
"Tapping of oleo-resin from \emph{Dipterocarpus kerrii},
\emph{The Malaysian Forester} \strong{50}, 343-353.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
