% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{OneCell}
\alias{OneCell}
\title{A synthetic example of a one-cell interaction.}
\format{
A dataframe of 32 observations on five variables.
\describe{
\item{\code{A}}{Factor 1.}
\item{\code{B}}{Factor 2.}
\item{\code{C}}{Factor 3.}
\item{\code{D}}{Factor 4.}
\item{\code{y}}{A response.}
}
}
\usage{
OneCell
}
\description{
Synthetic data for a 2x2x2x2 factorial with a one-cell interaction.

Source: Oehlert (2022), originally from Oehlert (1994).
}
\references{
Oehlert, Gary W. (1994) "Isolating one-cell interactions," \emph{Technometrics}
\strong{36}, 403-408.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
