% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{OrangePulpSilage}
\alias{OrangePulpSilage}
\title{Moisture content of orange pulp silage}
\format{
A data frame of 12 observations on two variables.
\describe{
\item{\code{treatment}}{A factor denoting the treatment for each observation.}
\item{\code{moisture}}{The moisture content of the silage (percent).}
}
}
\usage{
OrangePulpSilage
}
\description{
Twelve orange pulp silage samples were randomly divided into four
groups of three, each group receiving its own treatment. Each group received a
different treatment (control, salt (NaCl), formic acid, beet pulp). The response
is the moisture content of the silage.

Source: Oehlert (2022), originally from Caro, et al. (1990).
}
\references{
Caro, M. R., E. Zamora, L. Leon, F. Cuello, J. Salinas, D. Megias, M. J. Cubero,
and Z. Contreras (1990). "Isolation and identification of \emph{Listeria monocytogenes}
in vegetable byproduct silages containing preservative additives and destined for animal
feeding. \emph{Animal Feed Science and Technology} \strong{31}, 285=291.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
