% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PacemakerDelamination}
\alias{PacemakerDelamination}
\title{Delamination of pacemaker substrates}
\format{
A dataframe with 16 observations on four variables.
\describe{
\item{\code{flow}}{A factor indicating the airflow rate.}
\item{\code{time.z}}{Firing profile time.}
\item{\code{time}}{A factor version of firing profile time.}
\item{\code{fraction}}{The fraction of substrates delaminating.}
}
}
\usage{
PacemakerDelamination
}
\description{
Implantable heart pacemakers contain small circuit boards called substrates.
These substrates are assembled, cut to shape, and fired. Some of the substrates will separate,
or delaminate, making them useless. The purpose of this experiment was to study the effects
of three factors on the rate of delamination. The factors were A: firing profile time,
8 versus 13 hours with the theory suggesting 13 hours is better; B: furnace airflow, low versus
high, with theory suggesting high is better; and C: laser, old versus new, with theory suggesting
new cutting lasers are better.

A large number of raw, assembled substrates are divided into sixteen groups. These
sixteen groups are assigned at random to the eight factor-level combinations of the three
factors, two groups per combination. The substrates are then processed, and the response
is the fraction of substrates that delaminate.

Source: Oehlert (2022), originally from T. Kerkow (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
