% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PacemakerPins}
\alias{PacemakerPins}
\title{Strength of pacemaker solder joints}
\format{
A dataframe with 48 observations on three variables.
\describe{
\item{\code{operator}}{A factor indicating which operator did the soldering.}
\item{\code{substrate}}{A factor indicating the substrate number.}
\item{\code{strength}}{Strength of the solder joint in pounds.}
}
}
\usage{
PacemakerPins
}
\description{
Twelve "substrates" are randomly assigned to four operators,
three substrates per operator. Each operator solders four pins on each substrate.
Finally, each solder joint is assessed by measuring its shear strength.

Source: Oehlert (2022), originally from T. Kerkow (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
