% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PageFaults}
\alias{PageFaults}
\title{Page faults in computers.}
\format{
A dataframe of 54 observations on five variables.
\describe{
\item{\code{alg}}{A factor indicating the paging algorithm used.}
\item{\code{init}}{A factor indicating the initialization sequence.}
\item{\code{size}}{A factor indicating the size of the program.}
\item{\code{ram}}{A factor indicating the RAM allocation for the program.}
\item{\code{faults}}{The number of page faults observed.}
}
}
\usage{
PageFaults
}
\description{
Some computers divide memory into pages.  A page fault
occurs when a needed page of memory is not actually in RAM and needs
to be loaded. Doing this entails removing one of the RAM-resident
memory pages, and this experiment studies algorithms for choosing the page.

The experiment consists of running different programs on a
computer under different configurations and counting the number
of page faults.
There were two paging algorithms to study, and this is
the factor of primary interest.  The other factors are the
sequence in which the system was initialized (three levels),
the size of the program (three levels), and the memory allocation
for the program (three levels). We run these 54 combinations in random
order and get the count of page faults as response.

We expect all three of the
non-algorithm factors to be important and interact. The issue
is what does algoritm do and whether/how it interacts with the others.

Source: Oehlert (2022), originally from ... I think these data are
from Barry Margolin back in 1977, but I'm not certain.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
