% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairwise}
\alias{pairwise}
\title{Pairwise comparisons}
\usage{
pairwise(lmodel, term, confidence = 0.95, type = "hsd")
}
\arguments{
\item{lmodel}{A model object from \code{lm}, \code{lmer}, or \code{lme}.}

\item{term}{A term (variable name) in the linear model.}

\item{confidence}{The level of confidence for interval estimates of pairwise differences.}

\item{type}{The type of multiplicity correction to use.}
}
\value{
A list of class \code{pairwise} with named elements "ests",
     "sds", "signif", "title", and "groupings"; see details below.
}
\description{
Compute all pairwise comparisons among levels of a factor
in a linear model using various multiplicity corrections.
}
\details{
This function computes pairwise comparisons among
the levels of a factor in a linear model.  \code{term} must be a
factor in the the model \code{lmodel}. \code{type} can be one of
"hsd" (Tukey's Honest Significant Different, using the Studentized range),
"bsd" (Bonferroni Significant Difference, using a Bonferroni correction),
"lsd" (Lease Significant Difference, which makes no correction for multiplicity),
"regwr" (Ryan-Einot-Gabriel-Welsch test that controls the strong familywise error rate),
or "snk" (Student-Neuman-Keuls step down test).

\code{pairwise} is an alternative to \code{\link[stats:pairwise.t.test]{stats::pairwise.t.test}}.
or \code{\link[stats:TukeyHSD]{stats::TukeyHSD}}. The
functions have different choices of methods for controlling multiplicity,
and \code{pairwise} works on the output of a (possibly multiway factorial) linear
model, whereas the others work directly on data and a
grouping factor.

The value is a list that is not generally used directly.  Instead, it is printed
via the method \code{\link{print.pairwise}} or plotted via the method
\code{\link{plot.pairwise}} or displayed via \code{\link{sidelines}}.
The returned list has named elements "ests",
"sds", "signif", "title", and "groupings".  The title is simply a character string used
to label the printed output or the plot.  ests is a (symmetric) matrix containing the
pairwise differences, and signif is a logical matrix with i,j element TRUE if the i,j
pair of treatments is significantly different and FALSE otherwise. sds is a little
more complicated. The upper triangle contains the nominal significant differences for
comparing a pair of treatments.  The lower triangle is just the reflection of the
upper triangle except when using a pure stepdown method (regwr or snk).  In this case,
if an pair of treatments is declared to be not significantly different because an
enclosing pair is not significantly different, then the lower triangle in that element
will contain -sigd, where sigd is the significant difference for the enclosing pair
that was not significantly different. Finally, groupings will be NULL if all differences are
significant.  Otherwise, groupings is a logical matrix;
the columns of groupings indicate maximal consecutive (by rank order) groupings of
treatments that do not differ significantly from each other (TRUE meaning not
significantly different).
}
\examples{
data(FruitFlyLifespan)
fit.flies <- lm(longevity~companions,data=FruitFlyLifespan)
# all pairwise differences, .01 error rate, control strong family-wise error rate
pairwise(fit.flies,companions,type="regwr",confidence=.99)
data(ZincRetention)
fit.zinc <- lm(retention~m.zinc*m.protein+d.zinc,data=ZincRetention)
# also works for one factor in a factorial
pairwise(fit.zinc,m.protein,type="hsd")

}
\seealso{
\code{\link{print.pairwise}}, \code{\link{sidelines}}, \code{\link{compare.to.best}},
 \code{\link{compare.to.control}},
\code{\link{plot.pairwise}}, \code{\link[stats:pairwise.t.test]{stats::pairwise.t.test}},
\code{\link[stats:TukeyHSD]{stats::TukeyHSD}}
}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
