% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ParticleBoard}
\alias{ParticleBoard}
\title{Density of particle board}
\format{
A dataframe of 18 observations on six variables.
\describe{
\item{\code{slash.z}}{The percentage of slash chips used.}
\item{\code{slash}}{A factor versio of the percentage of slash chips used.}
\item{\code{resin.z}}{The percentage of resin used.}
\item{\code{resin}}{A factor version of the percentage of resin used.}
\item{\code{target.z}}{The target density.}
\item{\code{target}}{A factor version of the target density.}
\item{\code{density}}{The achieved density.}
}
}
\usage{
ParticleBoard
}
\description{
Eighteen samples of particle board are created (in random order) by
varying the target density (42 or 48 lb/ft^3), the fraction of resin used
(6, 9, or 12\%), and the fraction of slash (waste wood chips) used (0, 25 or 50\%).
The response is the achieved density of the particle board.

Source: Oehlert (2022), originally from Boehner (1975).
}
\references{
Boehner, A. W. (1975) \emph{The Effect of Three Species of Logging Slash on the
Properties of Aspen Planer Shavings Particleboard}. Ph. D. Thesis, University
of Minnesota, St. Paul, MN.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
