% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Pasteurization}
\alias{Pasteurization}
\title{Bacteria counts in milk after Pasteurization}
\format{
A dataframe of 15 observations on two variables.
\describe{
\item{\code{treatment}}{A factor indicating the treatment used.}
\item{\code{count}}{The bacterial count.}
}
}
\usage{
Pasteurization
}
\description{
Bacterial counts (via serial dilution) for 15 samples of milk randomly
assigned to three different Pasteurization methods.

Source: Oehlert (2022)
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
