% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PeaGermination}
\alias{PeaGermination}
\title{Germinating pea seeds}
\format{
A dataframe of 8 observations on three variables.
\describe{
\item{\code{germinated}}{The number of germinated seeds in the experimental unit.}
\item{\code{nongerminated}}{The number of non-germinated seeds, simply 20-germinated.}
\item{\code{treatment}}{A factor with four levels showing the treatment given to the
experimental unit. Levels are: towel_12, soil_12, towel_24, soil_24.}
}
}
\usage{
PeaGermination
}
\description{
Results of an experiment looking at the effects of pre-soaking pea seeds
either 12 or 24 hours and placing them on a substrate of either potting soil or paper
towel. Each experimental unit used 20 seeds, and the response is the number of seeds
that germinated after five days.

Source: Oehlert (2022), modeled on original data from K. Becklund (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
