% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permsign.R
\name{permsign.test}
\alias{permsign.test}
\title{Randomization/Permutation Sign Test}
\usage{
permsign.test(
  x,
  y = NULL,
  data = NULL,
  ntrials = 10000,
  plot = FALSE,
  bins = 20,
  nullvalue = 0,
  confidence = 0.95
)
}
\arguments{
\item{x}{A vector of numeric data, typically arising as a vector of differences.}

\item{y}{An optional vector of data.}

\item{data}{An optional data frame}

\item{ntrials}{Integer number of random sets of signs to use in simulating
the permutation distribution.}

\item{plot}{Logical flag for whether to plot a histogram of the permuation distribution.}

\item{bins}{Number of bins to use in the (optional) histogram.}

\item{nullvalue}{Optional null value to subtract from the data.}

\item{confidence}{The coverage rate for a two-sided interval.}
}
\value{
A list with components x0, lower.p, upper.p, twosided.p containing the
observed mean of the deviations of the data from the null value
and the lower, upper, and two-sided p-values for testing
the null hypothesis.
}
\description{
Sign test for zero mean of differences based on randomizing the signs.
}
\details{
\code{permsign.test} is a permutation test for the null hypothesis that
the mean of \code{x-nullvalue} is zero (or the mean of \code{x-y-nullvalue} if
 \code{y} is not NULL). The basis for the test is that the signs of the deviations from the null value are
independent with a 50/50 channce of positive or negative. The mean of the
deviations of the observed data from the null value is the test statistic, and the null distribution is estimated from
a random sample of of \code{ntrials} means with the signs of the deviations from the null value
randomized.

\code{permsign.test(x)} is an alternative to the t-test or the sign test.
\code{permsign.test(x-y)} (or equivalently \code{permsign.test(x,y)})
is an alternative to the paired t-test.

If \code{plot} is TRUE, then a histogram of the sampled null distribution is
printed with a vertical line drawn at the observed value. You may adjust the
number of bins in the histogram usings \code{bins}.
}
\examples{
data(RunStitch)
# paired two sample test using permuted differences
permsign.test(Standard,Ergonomic,data=RunStitch)
# test using permuted signs of differences
permsign.test(Standard-Ergonomic,data=RunStitch) # the same

}
\seealso{
\code{\link[stats:t.test]{stats::t.test}}
}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
