% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PineOleoresin}
\alias{PineOleoresin}
\title{Oleoresin tapped from pine trees}
\format{
A dataframe of 24 observations on three variables.
\describe{
\item{\code{shape}}{A factor indicating the shape of the cut.}
\item{\code{treatment}}{A factor indicating acid treatment or control.}
\item{\code{resin}}{The collected resin.}
}
}
\usage{
PineOleoresin
}
\description{
Pine oleoresin is obtained by tapping the trunks of pine trees.
Tapping is done by cutting a hole in the bark and collecting the
resin that oozes out.  This experiment compares four shapes for
the holes and the efficacy of acid treating the holes.  Twenty-four
pine trees are selected at random from a plantation, and the
24 trees are assigned at random to the eight combinations of
hole shape (circular, diagonal slash, check, rectangular)
and acid treatment (yes or no).  The response is total grams of
resin collected from the hole.

Source: Oehlert (2022),  originally from Low and Bin Mohd.Ali (1985).
}
\references{
Low, C. K. and A. R. Bin Mohd. Ali (1985) "Experimental tapping of pine
oleoresin." \emph{The Malaysian Forester} \strong{48} 248-253.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
