% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PlantShoots}
\alias{PlantShoots}
\title{Plasma leucine concentration in pigs}
\format{
A data frame of 132 observations on four variables.
\describe{
\item{\code{orient}}{A factor indicating the orientation of the seeds.}
\item{\code{medium}}{A factor indicating the medium in which the seeds were soaked.}
\item{\code{day}}{A factor indicating the day on which these seeds were started.}
\item{\code{fracNormal}}{The fraction of shoots that are normal.}
}
}
\usage{
PlantShoots
}
\description{
Plant shoots can be encouraged in tissue culture by exposing
the cotyledons of plant embryos to cytokinin, a plant
growth hormone.  However, some shoots become watery, soft,
and unviable; this is vitrification.  An experiment was
performed to study how the orientation of
the embryo during exposure to cytokinin and the type of growth
medium after exposure to cytokinin affect the
rate of vitrification.
There are six treatments, which are the
factorial combinations of orientation (standard and experimental)
and medium (three kinds).  On a given day, the experimenters
extract embryos from white pine seeds and randomize them
to the six treatments.  The embryos are exposed using
the selected orientation for 1 week, and then go onto
the selected medium.
The experiment was repeated 22 times on different starting days.
The response is the fraction of shoots that are normal

Source: Oehlert (2022), originally from David Zlesak (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
