% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PlasmaLeucine}
\alias{PlasmaLeucine}
\title{Plasma leucine concentration in pigs}
\format{
A data frame of 35 observations on four variables.
\describe{
\item{\code{source}}{A factor indicating the source of protein in the diet.}
\item{\code{percent.z}}{The protein concentration in the diet.}
\item{\code{percent}}{A factor version of \code{percent.z}.}
\item{\code{leucine}}{The free plasma leucine level.}
}
}
\usage{
PlasmaLeucine
}
\description{
An animal nutrition experiment was conducted to study
the effects of protein in the diet on the level of
leucine in the plasma of pigs.  Pigs were randomly assigned
to one of twelve treatments.  These treatments are
the combinations of protein source (fish meal, soybean meal,
and dried skim milk) and protein concentration in the
diet (9, 12, 15, or 18 percent).  The response is the
free plasma leucine level in mcg/ml.

Source: Oehlert (2022), originally from Windels (1964).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Windels, H. F. (1964). \emph{The Influence of Diet and Duration of Fast upon
Plasma Levels of Free Leucine, Isoleucine, and Valine in the Growing Pig},
Ph. D. thesis, University of Minnesota, St. Paul, MN.
}
\keyword{datasets}
