% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrastplot.R
\name{plot.linear.contrast}
\alias{plot.linear.contrast}
\title{Plot linear contrasts
     
Plot a summary of linear contrasts produced by \code{linear.contrast}.}
\usage{
\method{plot}{linear.contrast}(
  x,
  title = "",
  adjust.method = "none",
  cex = 1,
  marginlab = FALSE,
  xaxis = TRUE,
  plines = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class "linear.contrast".}

\item{title}{The title for the plot.}

\item{adjust.method}{Method to adjust p-values}

\item{cex}{Expansion factor for contrast labels}

\item{marginlab}{If TRUE, contrast labels are put in the margin.}

\item{xaxis}{If TRUE, a horizontal line at y=0 is added.}

\item{plines}{If not NULL, a vector of p-value cutoffs where horizontal lines are drawn.}

\item{...}{Additional graphical parameters.}
}
\value{
Nothing returned.
}
\description{
Plot linear contrasts
     
Plot a summary of linear contrasts produced by \code{linear.contrast}.
}
\details{
The plot method makes a plot displaying the
results of linear contrasts from the \code{linear.contrast} function.  Each contrast
is shown as a line segment representing the confidence (or posterior) interval along 
the vertical axis with the contrasts ordered by p-value (or Bayes factor for contrasts
from a \code{bglmm} object) along the horizontal axis.
There is a separate curve showing the p-value (or Bayes factor) for each contrast.
For non-Bayesian results, you may opt to adjust the p-values in \code{x} by one
of the adjustment methods available to \code{p.adjust}. Note: the plotted intervals
are the original individual intervals regardless of the choice of p-value adjustment.
}
\examples{
 data(FruitFlyLifespan)
 fit.flies <- lm(longevity~companions,data=FruitFlyLifespan)
flies.allpairs <- linear.contrast(fit.flies,companions,allpairs=TRUE)
 plot(flies.allpairs,cex=.5,ylim=c(-35,45),plines=.05)
 # or control strong familywise error rate
 plot(flies.allpairs,cex=.5,ylim=c(-35,45),plines=.05,adjust="holm")

}
\seealso{
\code{\link{linear.contrast}}, \code{\link{p.adjust.methods}}.
}
