% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseplot.R
\name{plot.pairwise}
\alias{plot.pairwise}
\title{Plot pairwise comparisons
     
Plot a summary of all pairwise comparisons among levels of a factor.}
\usage{
\method{plot}{pairwise}(x, title = x$title, ...)
}
\arguments{
\item{x}{An object of class "pairwise".}

\item{title}{The title for the plot.}

\item{...}{Currently ignored.}
}
\value{
Nothing returned.
}
\description{
Plot pairwise comparisons
     
Plot a summary of all pairwise comparisons among levels of a factor.
}
\details{
The plot method makes a plot displaying the
results of pairwise comparisons from the \code{pairwise} function.  Each comparison
is shown as a line segment from the (i,j) means to the (j,i) means.  Significant differences
are shown in red; nonsignificant differences are plotted in black.  For a significant difference,
the cutoff for significance is plotted as a gray point on the red line.  Nonsignificant
differences are bounded in a gray square going out to the required cutoff.
For pure stepdown procedures (snk and regwr), the gray square for a comparison is not
plotted if it is enclosed in a broader nonsignificant comparison.
}
\examples{
data(FruitFlyLifespan)
fit.flies <- lm(longevity~companions,data=FruitFlyLifespan)
pairwise.flies <- pairwise(fit.flies,companions,type="snk",confidence=.99)
plot(pairwise.flies)
data(ZincRetention)
fit.zinc <- lm(retention~m.zinc*m.protein+d.zinc,data=ZincRetention)
# also works for one factor in a factorial
pairwise.zinc <- pairwise(fit.zinc,m.protein,type="hsd")
plot(pairwise.zinc)

}
\seealso{
\code{\link{pairwise}}, \code{\link{print.pairwise}},
     \code{\link{sidelines}}, \code{link{plot.linear.contrast}},
 \code{\link[stats:pairwise.t.test]{stats::pairwise.t.test}},
     \code{\link[stats:TukeyHSD]{stats::TukeyHSD}}
}
