% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PoliticalAds}
\alias{PoliticalAds}
\title{Perception of political campaigns}
\format{
A dataframe of 113 observations on two variables.
\describe{
\item{\code{condition}}{The treatment applied to the subject.}
\item{\code{response}}{The perception of the campaign on a 1 (positive) to 7 (negative) scale.}
}
}
\usage{
PoliticalAds
}
\description{
One-hundred thirteen people were randomly assigned into five groups.
Each group will receive some information about a political campaign,
and then make a determination about whether the campaign has been
using positive or negative advertising.  They make that rating on a
1 to 7 scale, with 1 being most positive and 7 being most negative.
Group one receives the transcript of a television ad that attacks
an opposing candidate.  Group two receives the transcript of the ad
plus an editorial describing the campaign as generally positive.
Group three receives the transcript of the ad
plus an editorial describing the campaign as generally negative.
Groups four and five receive only the positive and negative
editorials, respectively.

Source: Oehlert (2022), synthetic data adapted from A. Bos (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
