% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PopcornRatios}
\alias{PopcornRatios}
\title{Volume ratio of popped to unpopped popcorn.}
\format{
A dataframe of 32 observations on five variables.
\describe{
\item{\code{pType}}{A factor indicating the type of popcorn used.}
\item{\code{pAmt}}{A factor indicating the amount of popcorn used (in cups).}
\item{\code{oType}}{A factor indicating the type of oil used.}
\item{\code{oAmt}}{A factor indicating the volume of oil used (in tablespoons).}
\item{\code{ratio}}{The popped to unpoped volume ratio.}
}
}
\usage{
PopcornRatios
}
\description{
Thirty-two batches of popcorn are produced, two each for the combinations
of popcorn kernel amount (1/8 or 1/4 cup), popcorn type (generic or
gourmet), oil type (canola or "popcorn" oil), and oil amount (1 or 2 tablespoons).
The response is the ratio of the volumes of the popped and unpopped popcorn.

Source: Oehlert (2022), data from J. McLaren (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
