% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PotatoChips}
\alias{PotatoChips}
\title{Sensory aspects of regular and reduced fat potato chips.}
\format{
A dataframe of 48 observations on four variables.
\describe{
\item{\code{judge}}{A factor indicating the judge rating the chips.}
\item{\code{period}}{A factor indicating first or second tasting period.}
\item{\code{fat}}{A factor indicating regular or reduced fat.}
\item{\code{score}}{The sensory score of the chips.}
}
}
\usage{
PotatoChips
}
\description{
A sensory experiment was conducted to determine if consumers
have a preference between regular potato chips (A) and reduced-fat
potato chips (B).  Twenty-four judges will rate both types of
chips; twelve judges will rate the chips in the order regular fat, then
reduced fat; and the other twelve will have the order
reduced fat, then regular fat.  We anticipate judge to judge
differences and possible differences between the first and
second chips tasted.  The response is a liking scale, with
higher scores indicating greater liking

Source: Oehlert (2022),  data  from Monica Coulter (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
