% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{power.f.test}
\alias{power.f.test}
\title{Power for F tests}
\usage{
power.f.test(
  means = NULL,
  ns = NULL,
  sigma2 = NULL,
  ncp = NULL,
  df1 = NULL,
  df2 = NULL,
  alpha = 0.05
)
}
\arguments{
\item{means}{A vector of assumed treatment means.}

\item{ns}{Sample sizes for each level of the treatment factor.}

\item{sigma2}{The error variance.}

\item{ncp}{The noncentrality parameter.}

\item{df1}{Numerator degrees of freedom.}

\item{df2}{Denominator degrees of freedom.}

\item{alpha}{Type I error rate for test.}
}
\value{
The power for the test.
}
\description{
Compute power for F tests with multiple possible definitions of noncentrality
and degrees of freedom.
}
\details{
Specify either a vector of means, a vector of sample
sizes, and an error variance, or specify a noncentrality parameter,
numerator degrees of freedom, and denominator degrees of freedom. If the
sample sizes are all equal, then the samples size can be specified as a
scalar.

The noncentrality parameter is computed in the following way:
\code{grandmean <- sum(ns*means)/sum(ns)}
\code{ncp <- sum(ns*(means-grandmean)^2)/sigma2}

\code{ncp}, \code{df1}, \code{df2} can be vectors or scalars. Values will be
recycled as needed if they are not all the same length.
}
\examples{
 # means 1:5, ni==4, error variance 4.5, alpha .05 (default)
 power.f.test(means=c(1,2,3,4,5),ns=4,sigma2=4.5)
 # same computation, but via degrees of freedom and non-centrality
 power.f.test(ncp=4*sum(((1:5)-3)^2)/4.5,df1=4,df2=15) # same thing

}
\seealso{
\code{sample.size.f.test}
}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
