% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{PowerPerception}
\alias{PowerPerception}
\title{Perception of interpersonal power}
\format{
A dataframe of 91 observations on two variables.
\describe{
\item{\code{expectation}}{The treatment applied to the subject.}
\item{\code{power}}{The perception of power.}
}
}
\usage{
PowerPerception
}
\description{
Ninety-three student volunteers are told that they will be having
a conversation with a member of their same gender. The students are
randomly assigned to three treatments, 31 students per treatment.
One group of students will be told that their conversation partner is
is an extravert, a second group is told that their partner is an
introvert, and the third group is given no information. Prior to
beginning the conversations, subjects fill out a questionaire.

In fact, there is no conversation; the quantities derived from the pre-conversation
questionaire are the responses. The response shown here is "perception of
power" in the upcoming conversation, a scale derived from 18 items in the
questionaire. Higher responses on this scale indicate perceptions of greater power.

Source: Oehlert (2022), synthetic data adapted from A. Baldwin (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
