% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{print.pairwise}
\alias{print.pairwise}
\title{Print pairwise comparisons}
\usage{
\method{print}{pairwise}(x, ...)
}
\arguments{
\item{x}{An object of class "pairwise".}

\item{...}{Currently ignored.}
}
\value{
Nothing returned.
}
\description{
Print a summary of all pairwise comparisons among levels of a factor
in a linear model using various multiplicity corrections.
}
\details{
The print method presents a printed output for displaying the
results of pairwise comparisons from the \code{pairwise} function.  For each comparison
it lists the estimated difference, the significant difference, and a confidence interval.
For pure stepdown methods (snk and regwr), confidence intervals for nonsignificant differences
use the margin of error for any enclosing nonsignificant comparison.

Significant differences are marked with a star.
}
\seealso{
\code{\link{pairwise}}, \code{\link{sidelines}},
\code{\link{plot.pairwise}}, \code{\link[stats:pairwise.t.test]{stats::pairwise.t.test}},
\code{\link[stats:TukeyHSD]{stats::TukeyHSD}}
}
