% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ProductScore}
\alias{ProductScore}
\title{Product score.}
\format{
A dataframe with 64 observations on nine variables.
\describe{
\item{\code{A}}{Level of factor 1.}
\item{\code{B}}{Level of factor 2.}
\item{\code{C}}{Level of factor 3.}
\item{\code{D}}{Level of factor 4.}
\item{\code{E}}{Level of factor 5.}
\item{\code{F}}{Level of factor 6.}
\item{\code{G}}{Level of factor 7.}
\item{\code{H}}{Level of factor 8.}
\item{\code{y}}{The produce score.}
}
}
\usage{
ProductScore
}
\description{
A product is produced that should have a score as close to
2 as possible.  Eight factors are believed to influence the score, and a
completely randomized experiment is conducted using 64 units and sixteen
treatments in a 2_IV^(8-4) fractional-factorial treatment structure.
Factors are named in order of increasing expense.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
