% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{QCM}
\alias{QCM}
\title{Quartz Crystal Microbalance.}
\format{
A dataframe of 96 observations on six variables.
\describe{
\item{\code{type}}{The type of coating.}
\item{\code{crystal}}{A factor indicating the crystal (nested in type, so 1,2,3,1,2,3 not 1,2,3,4,5,6).}
\item{\code{env}}{The environment.}
\item{\code{intensity}}{A factor indicating the intensity.}
\item{\code{y}}{The frequency change.}
\item{\code{int.z}}{A centered, quantitative variable indicating intensity based on
the assumption that intensity is equally spaced.}
}
}
\usage{
QCM
}
\description{
A quartz crystal will oscillate when an alternating current is applied.
However, the frequency of the oscillation will change
if some material is adsorbed to the surface of the crystal. This is the
basis of the quartz crystal microbalance (QCM), which can measure in nanogram
quantities. In a QCM, the crystal will typically be coated with some
material and placed in some gaseous or liquid environment. The
substance to be measured is introduced to the environment, and the
concentration is inferred from the change in frequency of the crystal.

In this experiment, we have two types of crystals: those coated with
gold and those coated with titanium dioxide. We have three random crystals
from each type of coating. The environments we will use are air and water.
The main purpose of this experiment is to investigate the change in
frequency when a third factor called intensity is varied across eight levels.
For each individual crystal, we make 16 observations on frequency change,
one for each of the combinations of environment and intensity. These 96
observations are done in random order. Note, there is not actually anything
being "weighed."

Source: Oehlert (2022),  data  from K. Greden (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
