% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunnett.R
\name{qdunnett}
\alias{qdunnett}
\alias{dunnett}
\title{Quantiles of the Dunnett distribtion.}
\usage{
qdunnett(ngroups, errordf, alpha = 0.025, twosided = FALSE)
}
\arguments{
\item{ngroups}{Number of groups (including control) for making comparisons to control.}

\item{errordf}{Error degrees of freedom when computing significance levels.}

\item{alpha}{Percent point for the Dunnett critical value.}

\item{twosided}{Logical flag indicating whether the Dunnett distribution should}
}
\value{
A list with the requested quantile of the Dunnett distribution (multivariate t)
 and the difference between the obtained probability and the requested probability.
}
\description{
Quantiles of the Dunnett distribtion.
}
\examples{
qdunnett(5,20,alpha=.01,twosided=TRUE)
}
