% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Quackgrass}
\alias{Quackgrass}
\title{Quackgrass invasion of big bluestem}
\format{
A dataframe of 24 observations on four variables.
\describe{
\item{\code{pct.Bluestem}}{A numeric variable giving the percentage of biomass that is bluestem.}
\item{\code{treatment}}{A factor with six levels indicating the treatment the plot received.}
\item{\code{water}}{A factor with two levels indicating the water treatment the plot received.}
\item{\code{N}}{The nitrogen dose each plot received.}
}
}
\usage{
Quackgrass
}
\description{
Twenty four plots are randomly assigned to six different combinations of
nitrogen addition (200, 400, 600, or 800 mg N/Kg soil) and irrigation (non or 1 cm/week).
Big bluestem is planted and allowed to establish itself.
After one year, quackgrass is seeded. After another year, biomass is harvested, and
the response is the percentage of biomass that is big bluestem.

Source: Oehlert (2022), based on Wedin (1990).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Wedin, E. (1990) \emph{Nitrogen Cycling and Competition and Grass Species,}
Ph. Thesis, University of Minnesota, Minneapolis, MN.
}
\keyword{datasets}
