% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RampMeters}
\alias{RampMeters}
\title{Performance of ramp meters.}
\format{
A dataframe of 16 observations on ten variables.
\describe{
\item{\code{minRel.z}}{The minimum release rate.}
\item{\code{minRel}}{A factor version of the minimum release rate.}
\item{\code{maxWait.z}}{The maximum wait time.}
\item{\code{maxWait}}{A factor version of the maximum wait time.}
\item{\code{maxRel.z}}{The maximum release rate.}
\item{\code{maxRel}}{A factor version of the maximum release rate.}
\item{\code{Incr.z}}{The demand increment.}
\item{\code{Incr}}{A factor version of the demand increment.}
\item{\code{PI}}{The performance index for the freeway.}
\item{\code{PI2}}{A version of PI with the fourth entry of 69 replaced by 690.}
}
}
\usage{
RampMeters
}
\description{
Ramp meters are stop lights on the entrance ramps to freeways in urban settings. They
are used to control the number of cars trying to merge onto the freeway
at any given time. Proper use of a ramp meter can lead to shorter
overall waiting times and fewer accidents. This experiment considers the
effects of four factors on a performance index for the ramp.
The experimental factors are minimum release rate (150 or 350 vehicles per hour),
maximum allowed waiting time (3.5 or 6 minutes), maximum release rate
on the ramp (1650 or 1800 vehicles per hour), and an increment in demand
above normal traffic (125 or 250 vehicles/hour).
The performance
index is a combination of the number of vehicles that need to wait longer
than the desired maximum, the total time waited above the desired maximum,
and the total distance the queue on the ramp extends beyond the beginning of the ramp.

This experiment is not performed on the driving public, but rather is performed
using a  traffic simulator for the entire transportation system in
a region. The simulator is extremely complex and includes many random inputs, so the response
to adjusting parameters like the ramp meter parameters can only be determined by
experiment.

Source: Oehlert (2022),  data adapted from A. Beegala (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
