% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RatLiverWeight}
\alias{RatLiverWeight}
\title{Rat liver weight}
\format{
A dataframe of 29 observations on two variables.
\describe{
\item{\code{weight}}{The liver weight, as a percentage of body weight.}
\item{\code{diet}}{A factor indicating the assigned diet.}
}
}
\usage{
RatLiverWeight
}
\description{
Twenty-nine rats were randomly assigned to one of four diet treatments.
Response is rate liver weight as a percentage of body weight. Synthetic data.
Source: Oehlert (2022).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
