% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ReactionTimes}
\alias{ReactionTimes}
\title{Reaction Times while on Cell Phone}
\format{
A data frame of 54 observations on six variables.
\describe{
\item{\code{meters.z}}{The following distance in meters.}
\item{\code{meters}}{A factor version of the following distance.}
\item{\code{duration.z}}{The duration of the call before braking (s).}
\item{\code{duration}}{A factor version of the duration.}
\item{\code{conditions}}{The time of day for the experiment.}
\item{\code{reactionTime}}{The reaction time of the driver (ms).}
}
}
\usage{
ReactionTimes
}
\description{
Using a cell phone while driving affects reaction time, but is it
just using a cell phone, or do other factors affect reaction time?
In this study, 54 male adults aged 22--24 years were asked to drive in a simulator under
various conditions, and their reaction times would be measured. At some
point during the simulation, the drivers begin talking on the cell phone.
After a certain time on the phone has elapsed, the ``car" ahead of the
driver hits the brakes, and the time between the leading car hitting
the brakes and the driver in the simulator hitting the brakes is the
measured response.

The 54 individuals were randomized to 18 conditions, which are 
the factor/level combinations of: meters (10, 15, or 20; the trailing
distance behind the leading car), conditions (day or night driving), and
duration (30, 60, or 90 seconds of cell phone conversation before the leading
car hits the brakes). The response is the reaction time (in milli seconds).

Source: Oehlert (2022), originally from Al-Darrab, et al (2009) via
Lye (2019).

Note: the original experiment procedure was more complex and on
actual roads. It also used each of 27 drivers twice, although details
were not given.
}
\references{
Al-Darrab, I. A., Zahid A. Khan, and Shiekh I. Ishrat (2009) ``An experimental
study on the effect of mobile phone conversation on drivers’ reaction time in braking response,"
\emph{Journal of Safety Research}, \strong{40} 185-189.

Lye, Leonard M. (2019) \emph{Applications of DEO in Engineering and Science:
A Collection of 2^6 Case Studies}, Leonard Lye, St. John's, Newfoundland.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
