% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ResinLifetimes}
\alias{ResinLifetimes}
\title{Resin lifetimes}
\format{
A dataframe of 37 observations on four variables.
\describe{
\item{\code{temp.z}}{A numeric variable giving the temperature (in degrees C) at which
the unit was stressed.}
\item{\code{logTime}}{The log_10 lifetime of the resin before failure.}
\item{\code{temp}}{A factor version of \code{temp.z}, included for convenience.}
\item{\code{Time}}{The lifetime in hours (as \code{10^Time}), included for convenience.}
}
}
\usage{
ResinLifetimes
}
\description{
Lifetime in hours of an encapsulating resin for gold-aluminum bonds in
integrated circuits mean to operate at 120 degrees C. These samples were stressed at
higher temperatures to shorten the lifetime, which would then be extrapolated back to
120 C.

Source: Oehlert (2022), originally from Nelson (1990).
}
\references{
Nelson, W. (1990). \emph{Accelerated Testing}. New York: Wiley.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
