% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RiceMalt}
\alias{RiceMalt}
\title{Free alpha amino acid in rice malt}
\format{
A dataframe of 12 observations on five variables.
\describe{
\item{\code{day.z}}{The number of days for germination.}
\item{\code{day}}{A factor version of the number of days for germination.}
\item{\code{temperature.z}}{The temperature in degrees C.}
\item{\code{temperature}}{A factor version of the temperature.}
\item{\code{FAN}}{The average free alpha amino acid for this treatment.}
}
}
\usage{
RiceMalt
}
\description{
Free alpha amino acid was measured in twenty-four batches of rice malt that
had been assigned randomly to 12 treatments, the factorial combinations of three levels
of germination time (in days) and four levels of temperature (in degrees C). Data here
are the twelve treatment means. The (corrected) total sum of squares in the full data set
was 8097.

Source: Oehlert (2022), originally from Aniche and Okafor (1989).
}
\references{
Aniche, N. G. and N. Okafor (1989) "Studies on the effect of germination time and
temperature on the malting of rice." \emph{Journal of the Institute of Brewing}
\strong{95}, 165-167.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
