% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RLS}
\alias{RLS}
\title{Replicated Latin Square}
\format{
A dataframe of 32 observations on four variables.
\describe{
\item{\code{row}}{A factor indicating the row of the square (1-4).}
\item{\code{column}}{A factor indicating the column of the squares (1-8).}
\item{\code{treatment}}{A factor indicating the treatment.}
\item{\code{response}}{The experimental response.}
}
}
\usage{
RLS
}
\description{
A 4x4 Latin Square with two replications.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
