% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RocketFuel}
\alias{RocketFuel}
\title{Elasticity of rocket fuel mixtures}
\format{
A dataframe of 10 observations on four variables.
\describe{
\item{\code{binder}}{The pseudo-component for the binder.}
\item{\code{oxidizer}}{The pseudo-component for the oxidizer.}
\item{\code{fuel}}{The pseudo-component for the fuel.}
\item{\code{elasticity}}{The elasticity of the fuel mixture.}
}
}
\usage{
RocketFuel
}
\description{
Three components of a rocket propellant are the binder, the
oxidizer, and the fuel. We want to find the mixtures that yield coefficients
of elasticity less than 3000.
All components must be present and there are minimum proportions,
so the investigators used a pseudocomponents design.

Source: Oehlert (2022), originally from Kurotori (1966).
}
\references{
Kurotori, I. S. (1966). "Experiments with mixtures of components having lower bounds."
\emph{Industrial Quality Control}, May, 592-596

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
