% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Rockets}
\alias{Rockets}
\title{Azimuth errors of rockets}
\format{
A dataframe of 81 observations on four variables.
\describe{
\item{\code{day}}{A factor indicating the day of the launch.}
\item{\code{slantRng}}{A factor indicating the slant range of the launch.}
\item{\code{temp}}{A factor indicating the propellant temperture of the launch.}
\item{\code{asError}}{The azimuth error of the launch.}
}
}
\usage{
Rockets
}
\description{
An army rocket development
program was investigating the effects of slant range and
propellant temperature on the accuracy of rockets.  The
overall objective of this phase of the program was to determine
how these variables affect azimuth error (that is, side to side as
opposed to distance) in the rocket impacts.

Three levels
were chosen for each of slant range and temperature.
The following
procedure was repeated on 3 days.  Twenty-seven rockets are
grouped into nine sets of three, which are then assigned to
the nine factor-level combinations in random order.  The
three rockets in a group are fired all at once in a single
volley, and the azimuth error recorded. (Note that meteorological
conditions may change from volley to volley.)

Source: Oehlert (2022), originally from Bicking (1958).
}
\references{
Bicking, C. A. (1958) "Experiences and needs for design in ordnance
experimentation. In V. Chow (Ed.), \emph{Experimental Designs in Industry},
pp. 247-252. New York, Wiley.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
