% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Rotations}
\alias{Rotations}
\title{Soybean weight under crop rotation.}
\format{
A dataframe of 32 observations on four variables.
\describe{
\item{\code{weight}}{The weight of 100 randomly chosen seeds.}
\item{\code{locYr}}{A factor indicating the location/year combination for the plot.}
\item{\code{rotation}}{A factor indicating the rotation used for the plot.}
\item{\code{variety}}{A factor indicating the variety used for the plot.}
}
}
\usage{
Rotations
}
\description{
An experiment to see how variety of soybean
and crop rotation practices affect soybean productivity.
There are two varieties used, Hodgson 78 and BSR191.
These varieties are each used
in four different 5-year rotation patterns with corn.
The rotation
patterns are (1) four years of corn and then soybeans (C-C-C-C-S),
(2) three years of corn and then two years of soybeans (C-C-C-S-S),
(3) soybean and corn alternation (S-C-S-C-S), and (4) five years of
soybeans (S-S-S-S-S).
Here we only have data from the fifth year.

This experiment was conducted twice in Waseca, MN, and twice in
Lamberton, MN. Two groups of eight plots were chosen at each location.  The first
group of eight plots at each location was randomly assigned to the variety-rotation
treatments in 1983.  The second group was then assigned
in 1984.  Responses were measured in 1987 and 1988 (the fifth
years) for the two groups.

The response of interest is the weight (g) of 100 random seeds from
soybean plants

Source: Oehlert (2022), originally from Whiting (1990).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Whiting, K. R. (1990) \emph{Host-Specific Pathogens and the Corn/Soybean Rotation
Effect}. Ph. D. thesis, University of Minnesota, St. Paul, MN.
}
\keyword{datasets}
