% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Ruffe}
\alias{Ruffe}
\title{Weight change of yellow perch under competition from ruffe.}
\format{
A dataframe of 16 observations on four variables.
\describe{
\item{\code{encl}}{A factor indicating the enclosure.}
\item{\code{density}}{A factor indicating the density of fish.}
\item{\code{ruffe}}{A factor indicating the addition of ruffe.}
\item{\code{wtChange}}{The weight change (grams).}
}
}
\usage{
Ruffe
}
\description{
Yellow perch and ruffe are two fish species that compete.
An experiment is run to determine the effects of fish density
and competition with ruffe on the weight change in yellow perch.
There are two
levels of fish density (low and high) and two levels of
competition (ruffe absent and ruffe present). Sixteen
tanks are arranged in four enclosures of four tanks each.
Within each enclosure, the four tanks are randomly assigned to
the four factor-level combinations of density and competition.
The response is the change in the weight of perch after 5 weeks
(in grams).

Source: Oehlert (2022), originally from Julia Frost (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
