% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{RunStitch}
\alias{RunStitch}
\title{Average time to runstitch a collar in standard and ergonomic conditions.}
\format{
A dataframe of 30 observations on two variables.
\describe{
\item{\code{Standard}}{Runstitch time under standard conditions.}
\item{\code{Ergonomic}}{Runstitch time under ergonomic conditions.}
}
}
\usage{
RunStitch
}
\description{
Thirty employees will runstitch collars using a standard
workspace set up and an ergonomic workspace setup. The response
is the average time to runstich collars (in seconds). For each
employee a coin is tossed; heads means that the employee will
use the standard first, and tails indicates that the employee
	will use ergonomic first.

Source: Oehlert (2022), originally from Bezjak and Knez (1995).
}
\references{
Bezjak, Z. and B. Knez (1995)  "Workplace design and loadings in the
process of sewing garments," \emph{International Journal of Clothing Science
and Technology} 7, 89-101.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
