% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SadAds}
\alias{SadAds}
\title{Willingness to buy after viewing a video.}
\format{
A dataframe of 22 observations on three variables.
\describe{
\item{\code{show}}{A factor indicating the category of program.}
\item{\code{feel}}{How the subject feels after the video.}
\item{\code{buy}}{How likely the subject is to purchase after the video.}
}
}
\usage{
SadAds
}
\description{
Advertisers wish to determine if program content affects
the success of their ads on those programs.  They produce
two videos, one containing  a depressing drama and some ads,
the second containing an upbeat comedy and the same ads.  Twenty-two
subjects are split at random into two groups of eleven, with
the first group watching the drama and the second group watching
the comedy.  After the videos, the subjects are asked several questions,
including "How do you feel?" and "How likely are you to buy?"
one of the products mentioned in the ads.  "How do you feel" was on a
1 to 6 scale, with 1 being happy and 6 being sad.
"How likely are you to buy?" was also on a 1 to 6 scale, with 6
being most likely.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
