% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{sample.size.f.test}
\alias{sample.size.f.test}
\title{Sample size for F tests}
\usage{
sample.size.f.test(
  reqpower,
  alpha = 0.05,
  means = NULL,
  sigma2 = NULL,
  ncp1 = NULL,
  ngrps = NULL,
  propvec = NULL
)
}
\arguments{
\item{reqpower}{The minimum power needed.}

\item{alpha}{Type I error rate for test.}

\item{means}{A vector of assumed treatment means.}

\item{sigma2}{The error variance.}

\item{ncp1}{The noncentrality parameter when all sample sizes are 1.}

\item{ngrps}{The number of treatment means (number of groups in the num).}

\item{propvec}{An optional vector giving the relative sizes requested for different treatments.}
}
\value{
A list including the calculated sample sizes and the achieved power for the test.
}
\description{
Compute the sample size needed for an F test to achieve a given power,
 with multiple possible definitions of noncentrality.
}
\details{
In addition to the requested power and type I error rate,
you must specify either the means and the error
variance \emph{or} the number of groups and \code{ncp1}
(the noncentrality parameter when
all sample sizes are equal to 1).

When calling the function using means, you may also supply propvec,
which should be a positive vector of the same length as the number of means.
The function will compute sample sizes that are approximately proportional
to the elements of the vector.
}
\examples{
 sample.size.f.test(.8,means=1:5,sigma2=4.5)
 sample.size.f.test(.8,ncp1=sum(((1:5)-3)^2)/4.5,ngrps=5) # same thing
 # more samples in middle needs more total data
 sample.size.f.test(.8,means=1:5,sigma2=4.5,propvec=c(1,2,2,2,1))
 # more samples at ends needs fewer total data
 sample.size.f.test(.8,means=1:5,sigma2=4.5,propvec=c(2,1,1,1,2))

}
\seealso{
\code{power.f.test}
}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
