% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{sample.size.t.interval}
\alias{sample.size.t.interval}
\title{Sample size for confidence intervals}
\usage{
sample.size.t.interval(width, sigma2, sumw2, g, alpha = 0.05)
}
\arguments{
\item{width}{The maximum width (twice the margin of error) for the confidence interval.}

\item{sigma2}{A the true error variance.}

\item{sumw2}{The sum of squared linear coefficients.}

\item{g}{The number of groups.}

\item{alpha}{The error rate for the interval (1-alpha coverage).}
}
\value{
The minimum per group sample size to achieve the coverage and width.
}
\description{
Compute the per group sample size needed for a confidence interval for a linear combination
of g group means to have coverage 1-alpha and be no wider than a determined width.
}
\details{
The sample variance is assumed to be a pooled MSE from a
model with g groups and equal per group sample sizes.
}
\examples{
 # .02 margin of error, .25 error variance, pairwise comparison, 5 groups
sample.size.t.interval(.04,.25,2,5)

}
\seealso{
\code{sample.size.f.test} and \code{sample.size.bayes.interval}
}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
