% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SeatSoftness}
\alias{SeatSoftness}
\title{Softness of cold-foamed car seats.}
\format{
A dataframe with 8 observations on eight variables.
\describe{
\item{\code{A}}{Level of factor 1.}
\item{\code{B}}{Level of factor 2.}
\item{\code{C}}{Level of factor 3.}
\item{\code{D}}{Level of factor 4.}
\item{\code{E}}{Level of factor 5.}
\item{\code{F}}{Level of factor 6.}
\item{\code{G}}{Level of factor 7.}
\item{\code{softness}}{The observed softness.}
}
}
\usage{
SeatSoftness
}
\description{
Seven factors are believed to control the softness of
cold-foamed car seats, and an experiment was conducted
to determine how these factors influence the softness.
A 2_III^(7-4) design was run with generators
I = ABD = ACE = BCF = ABCG.  The
response is the average softness of the seats

Source: Oehlert (2022), originally from Bergman and Hynen (1997)
}
\references{
Bergman, B. and A. Hynen (1997) "Dispersion effects from unreplicated designs
in the 2^k-p series," \emph{Technometrics} \strong{39}, 191-198.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
