% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SewageBacteria}
\alias{SewageBacteria}
\title{Antibiotic resistant bacteria in sewage.}
\format{
A dataframe of 48 observations on four variables.
\describe{
\item{\code{medium}}{A factor indicating LB or CAGY.}
\item{\code{origin}}{A factor indicating sludge or effluent.}
\item{\code{antibiotic}}{A factor indicating which antibiotic was added.}
\item{\code{count}}{The bacterial count (per uL of sample).}
}
}
\usage{
SewageBacteria
}
\description{
Bacteria may be adapting to non-zero levels of antibiotics in sewage.
We will take some sewage, treat it with some kind of antibiotic,
grow it on a plate, and then count the bacteria that survive.
Forty-eight preparations were made, three each for the combinations
of origin of sample (activated sludge or effluent), growth medium
(LB or CAGY), and antibiotic (Amoxicillin, Tetracycline, Tylosin,
or none). Bacterial counts (per uL of sample) were determined by serial plate counting.

Source: Oehlert (2022), data adapted from S. Ghosh (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
