% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ShadedCoffee}
\alias{ShadedCoffee}
\title{Coffee yield under shade trees.}
\format{
A dataframe of 12 observations on four variables.
\describe{
\item{\code{density.z}}{The planting density (per hectare) of the shade trees.}
\item{\code{density}}{A factor version of \code{density.z}.}
\item{\code{location}}{A factor giving the location.}
\item{\code{yield}}{The yield of coffee.}
}
}
\usage{
ShadedCoffee
}
\description{
Shade trees in coffee plantations may increase or decrease
the yield of coffee, depending on several environmental and
ecological factors.  Robusta coffee was planted at three
locations in Ghana.  Each location was divided into four plots,
and trees were planted at densities of 185, 90, 70,
and 0 trees per hectare.  Data are the yields of coffee
(kg of fresh berries per hectare) for the 1994-95 cropping season.

Source: Oehlert (2022), originally from Amoah, et al. (1997).
}
\references{
Amoah, F. M., K. Osei-Bonsu, and F. K. Oppong (1997). "Response of
improved robusta coffee to location and management practices in Ghana."
\emph{Experimental Agriculture} \strong{cc}, 103-111.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
