% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Shrinkage}
\alias{Shrinkage}
\title{Shrinkage of speedometer cable casings.}
\format{
A dataframe with 16 observations on ten variables.
\describe{
\item{\code{A}}{Level of liner OD.}
\item{\code{B}}{Level of liner die.}
\item{\code{C}}{Level of liner material.}
\item{\code{D}}{Level of liner line speed.}
\item{\code{E}}{Level of wire braid type.}
\item{\code{F}}{Level of braiding tension.}
\item{\code{G}}{Level of wire diameter.}
\item{\code{H}}{Level of liner tension.}
\item{\code{J}}{Level of liner temperature.}
\item{\code{K}}{Level of coating material.}
\item{\code{L}}{Level of coating die type.}
\item{\code{M}}{Level of melt temperature.}
\item{\code{N}}{Level of screen pack.}
\item{\code{O}}{Level of cooling method.}
\item{\code{P}}{Level of line speed.}
\item{\code{strength}}{The observed shrinkage.}
}
}
\usage{
Shrinkage
}
\description{
An experiment was conducted to determine the factors that affect
the amount of shrinkage in speedometer cable casings.  There were
fifteen factors, each at two levels, but the design used only sixteen factor-level
combinations (2_III^(15-11)).  The generators were
I = -DHM = -BHK = BDF = BDHO = -AHJ = -ADE = ADHN
= -ABC = ABHL = ABDG = -ABDHP, and the factors were:
liner OD (A); liner die (B); liner material (C);
liner line speed (D); wire braid type (E); braiding
tension (F); wire diameter (G); liner tension (H);
liner temperature (J); coating material (K);
coating die type (L);  melt temperature (M);
screen pack (N); cooling method (O); and line speed (P).
The response is the average of four shrinkage
measurements.

Source: Oehlert (2022), originally from Quinlan (1985)
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Quinlan, J. (1985). Product improvement by application of Taguchi meth- ods. In
\emph{American Supplier Institute News} (special symposium ed.), pp. 11-16,
Dearborn, MI: American Supplier Institute.
}
\keyword{datasets}
