% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidebyside.R
\name{sidebyside}
\alias{sidebyside}
\title{Side by Side plot}
\usage{
sidebyside(
  lmodel,
  p.cex = 1,
  t.cex = 1,
  left.margin = 8,
  show.intercept = FALSE,
  n.box = 30,
  ...
)
}
\arguments{
\item{lmodel}{A model object from one of \code{lm}, \code{lme}, \code{lmer}, or \code{bglmm}.}

\item{p.cex}{Expansion factor for drawn points in plot.}

\item{t.cex}{Expansion factor for text labeling drawn in margins.}

\item{left.margin}{Size of left margin (where term labels are drawn) in "lines".}

\item{show.intercept}{Should the intercept be included in the plot.}

\item{n.box}{Cutoff for displaying a boxplot instead of individual points.}

\item{...}{Other arguments passed to \code{plot}.}
}
\value{
No value returned other than producing the plot.
}
\description{
A side by side plot for the effects of a linear model (including fixed and/or random effects).
}
\details{
A side by side plot shows the estimated effects
for each term in a linear model. Terms are displayed separately as
points along a horizontal line, with a term label written
in the left margin.

Term labels can be quite long, so \code{cex} and
\code{left.margin} can be used to make term labels smaller
(e.g., \code{cex=.5}) or the left margin wider (e.g.,
\code{left.margin=12}).

\code{t.cex} can be used to change the size of the plotted points.

The default is not to plot the intercept/constant in a model,
as it is often of a very different magnitude than the
effects.

Terms with many values (particulary residuals), can be cluttered.
For terms with as many as \code{n.box} values, a boxplot will be plotted
instead of individual points.
}
\examples{
faults.orig <- lm(faults~(alg+ram+size+init)^3,data=PageFaults)
faults.log <- lm(log(faults)~(alg+ram+size+init)^3,data=PageFaults)
# see effects on original scale
sidebyside(faults.orig)
# interactions are relatively smaller when analyzed on log scale
sidebyside(faults.log)

}
\author{
Gary W. Oehlert \email{gary@umn.edu}
}
