% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidelines.R
\name{sidelines}
\alias{sidelines}
\title{Display pairwise comparisons}
\usage{
sidelines(x, ...)
}
\arguments{
\item{x}{An object of class "pairwise".}

\item{...}{Currently ignored.}
}
\value{
Nothing returned.
}
\description{
Display an "underline"  summary of all pairwise comparisons among levels of a factor.
}
\details{
The sidelines method presents an "underline" diagram for displaying the
results of pairwise comparisons from the \code{pairwise} function. Any two
treatments that are covered by a single line are not significantly different.

This implementation arranges the output vertically, so the "underlines" are actually
running vertically along the side, hence sidelines.
}
\examples{
data(FruitFlyLifespan)
fit.flies <- lm(longevity~companions,data=FruitFlyLifespan)
pairwise.flies <- pairwise(fit.flies,companions,type="hsd",confidence=.99)
sidelines(pairwise.flies)
data(ZincRetention)
fit.zinc <- lm(retention~m.zinc*m.protein+d.zinc,data=ZincRetention)
# also works for one factor in a factorial
pairwise.zinc <- pairwise(fit.zinc,m.protein,type="hsd")
sidelines(pairwise.zinc)

}
\seealso{
\code{\link{pairwise}}, \code{\link{print.pairwise}},
\code{\link{plot.pairwise}}, \code{\link[stats:pairwise.t.test]{stats::pairwise.t.test}},
\code{\link[stats:TukeyHSD]{stats::TukeyHSD}}
}
