% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SiliconThickness}
\alias{SiliconThickness}
\title{Thickness of silicon wafers.}
\format{
A dataframe with 16 observations on nine variables.
\describe{
\item{\code{A}}{Level of rotation method.}
\item{\code{B}}{Level of wafer code.}
\item{\code{C}}{Level of deposition temperature.}
\item{\code{D}}{Level of deposition time.}
\item{\code{E}}{Level of arsenic flow rate.}
\item{\code{F}}{Level of HCl etch temperature.}
\item{\code{G}}{Level of HCl flow rate.}
\item{\code{H}}{Level of nozzle position.}
\item{\code{thickness}}{The observed thickness of the silicon.}
}
}
\usage{
SiliconThickness
}
\description{
Silicon wafers for integrated circuits are grown in
a device called a susceptor, and a response of interest
is the thickness of the silicon.  Eight factors, each at two levels,
were believed to contribute: rotation method (A),
wafer code (B), deposition temperature (C), deposition
time (D), arsenic flow rate (E), HCl etch temperature (F),
HCl flow rate (G), and nozzle position (H).
A 2_IV^(8-4) design was run with generators
I = ABCD = BCEF = ACEG = BCEH.  The average
thickness of the silicon is the response.

Source: Oehlert (2022), originally from Shoemaker, Tsui, and Wu (1991)
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Shoemaker, A. C., K.-L. Tsui, and C. F. J. Wu (1991) "Economical experimentation
methods for robust design," \emph{Technometrics} \strong{33}, 415-427.
}
\keyword{datasets}
