% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Sires}
\alias{Sires}
\title{Average daily weight gain by calves.}
\format{
A dataframe of 24 observations on two variables.
\describe{
\item{\code{bull}}{A factor indicating the bull.}
\item{\code{gain}}{The average daily weight gain.}
}
}
\usage{
Sires
}
\description{
Four bulls are selected at random, and each bull is bred to
multiple randomly selected cows. The response is average daily weight gain
of the calves.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
