% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Softness}
\alias{Softness}
\title{Softness of clothes under different detergents}
\format{
A dataframe of 120 observations on three variables.
\describe{
\item{\code{judge}}{The consumer doing the rating.}
\item{\code{treat}}{The detergent in which the clothes were washed.}
\item{\code{softness}}{The softness rating.}
}
}
\usage{
Softness
}
\description{
Thirty consumers are asked to rate the softness of clothes washed
by ten different detergents, but  each consumer rates only four different
detergents.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
