% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SoupCrackers}
\alias{SoupCrackers}
\title{Absorbance of soup on crackers}
\format{
A data frame of 18 observations on five variables.
\describe{
\item{\code{replication}}{The day the observations were made.}
\item{\code{type}}{A factor giving the kind of soup.}
\item{\code{temp}}{A factor giving the temperature treatment of the soup.}
\item{\code{temp.z}}{The temperature of the soup.}
\item{\code{absorbed}}{The amount of soup absorbed onto the cracker (g).}
}
}
\usage{
SoupCrackers
}
\description{
Many people enjoy dipping crackers in their soup. This experiment
explores the amount of soup that is absorbed by, or adhered to, dipping
crackers. We consider three different types of soup: a condensed, cream-based
soup (Cream of Mushroom, need to add water); a condensed water-based soup (Tomato,
need to add water);
and a ready-to-eat pre-made soup (Steak and Potatoes, no additional water needed).
We also consider three serving temperatures: 100, 110, and 120 degrees F.
}
\details{
In this experiment, nine bowls of soup will be prepared in random order,
one each for the combinations of soup type and serving temperature.
For each bowl, 200g (approximately) of the soup is added and the
combination of soup and bowl is weighed. Then five unbroken
saltine crackers are placed in the soup. After two minutes, the crackers
are removed (along with any soup that clings to them). The soup 
and bowl are again weighed, the decrease in weight (in g) is 
response. The experiment is then repeated the next day.

Source: Oehlert (2022), originally S. Kleba (pers. comm.)
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
