% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SproutingBarley}
\alias{SproutingBarley}
\title{Number of barley seeds sprouting}
\format{
A dataframe of 30 observations on five variables.
\describe{
\item{\code{weeks.z}}{The age of the seeds.}
\item{\code{weeks}}{A factor version of age.}
\item{\code{water.z}}{The amount of water used in germination.}
\item{\code{water}}{A factor version of water.}
\item{\code{sprouting}}{The number of seeds sprouting.}
}
}
\usage{
SproutingBarley
}
\description{
Barley seeds are divided into 30 lots of 100 seeds each.  The 30 lots are divided at
random into ten groups of three lots each, with each group receiving
a different treatment.  The ten treatments are the factorial
combinations of amount of water used for sprouting with two levels (4 or 8 ml)
and age of the seeds with five levels (1, 3, 6, 9, or 12 weeks).
The response measured is the number of seeds sprouting.

Source: Oehlert (2022), originally from Hareland and Madson (1989).
}
\references{
Hareland, G. A. and M. A. Madson (1989) "Barley dormancy and fatty acid composition
of lipids isolated from freshly harvested and stored kernels" \emph{Journal of the
Institute of Brewing} \strong{95}, 437-442.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
