% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SteelBars}
\alias{SteelBars}
\title{Adhesive strength between steel bars.}
\format{
A dataframe of 24 observations on four variables.
\describe{
\item{\code{day}}{Factor indicating the day.}
\item{\code{time}}{Coded level of curing time.}
\item{\code{temp}}{coded level of curing temperature.}
\item{\code{strength}}{Shear strength of the bond.}
}
}
\usage{
SteelBars
}
\description{
Curing time and temperature affect the shear strength of
an adhesive that bonds galvanized steel bars.  This
experiment was repeated on 2 separate days.  Twenty-four
pieces of steel are obtained by random sampling from
warehouse stock.  These are grouped into twelve pairs;
the twelve pairs are glued and then cured with one of
nine curing treatments assigned at random.  The treatments
are the three by three factorial combinations of
temperature (375, 400, and 450 degrees F, coded -1, 0, 2)
and time (30, 35, or 40 seconds, coded -1, 0, 1).  Four
pairs were assigned to the center point, and one pair to
all other conditions.  The response is
shear strength (in psi).

Source: Oehlert (2022),  originally from Khuri (1992).
}
\references{
Khuri, A. I. (1992). "Response surface models with random block effects,"
\emph{Technometrics} \strong{34}, 26-37.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
