% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{StringCoating}
\alias{StringCoating}
\title{Percent retention of guitar string coating}
\format{
A dataframe of 48 observations on seven variables.
\describe{
\item{\code{time.z}}{The curing time in minutes.}
\item{\code{time}}{A factor version of the curing time.}
\item{\code{temp.z}}{The curing temperature, in degrees C.}
\item{\code{temp}}{A factor version of the curing temperature.}
\item{\code{rate.z}}{The rate at which the sample is extracted from the dipping coater, in inches/minute.}
\item{\code{rate}}{A factor version of the rate.}
\item{\code{ptc.retained}}{The percentage of coating remaining after the salt treatment.}
}
}
\usage{
StringCoating
}
\description{
Many guitar strings are made (at least partially) of steel. Perspiration
and oils from the body can corrode the steel, so manufacturers use coatings
to prevent corrosion. This experiment studies how the manufacturing process
affects the durability of these coatings. Coatings are applied by dipping the
specimen in the coating, extracting it from the coating, and then curing the coating.
Forty-eight steel samples are randomly
assigned to the 16 factor/level combinations of rate (the speed at which the
steel is extracted from the dip coater, 6 or 12 inches per minute), the curing
temperature (20, 60, 90, or 120 degrees C), and the curing time (30 or 60 minutes).
After curing, the the coated steel samples are swirled for two minutes in a .9\% NaCl in water
solution to simulate exposure to perspiration.

The steel samples are weighed three times: before coating, after curing, and after
the NaCl treatment. From these, you can determine the weight of coating added and
the weight of coating remaining after the NaCl treatment. The response is the
percentage of coating retained after the salt treatment.

Source: Oehlert (2022),  data  from N. Ellering (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
