% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Substrates}
\alias{Substrates}
\title{Disk Drive Substrate Voltages}
\format{
A dataframe of 16 observations on five variables.
\describe{
\item{\code{operator}}{A factor indicating the operator for the observation.}
\item{\code{machine}}{A factor indicating the machine for the observation.}
\item{\code{day}}{A factor indicating the day for the observation.}
\item{\code{substrate}}{A factor indicating the substrate for the observation.}
\item{\code{amplitude}}{The response (in microvolts/100).}
}
}
\usage{
Substrates
}
\description{
Disk drive substrates may affect the amplitude of the
signal obtained during read back.  A manufacturer compares
four substrates: aluminum (A), nickel-plated aluminum (B), and two
types of glass (C and D).  Sixteen disk drives will be made,
four using each of the substrates.  It is felt that the machine used in manufacturing,
the operator of the machine,
and day of production may have an effect on the drives,
so these three effects were blocked.  The design and
The response is in microvolts/100.

Source: Oehlert (2022), originally from Nelson (1993).
}
\references{
Nelson, P. R. (1993). "Additional uses for the analysis of means and extended
tables of critical values," \emph{Technometrics} \strong{35}, 61-71.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
