% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{SulfurRemoval}
\alias{SulfurRemoval}
\title{Sulfur removal from high-sulfur coal}
\format{
A data frame of 32 observations on six variables.
\describe{
\item{\code{pH}}{A factor giving the pH of the reaction.}
\item{\code{size}}{A factor giving the particle size (um).}
\item{\code{iron}}{A factor giving the concentration of iron (mmol).}
\item{\code{density}}{A factor giving the density of the pulp (percent).}
\item{\code{time}}{A factor giving the leaching time (days).}
\item{\code{removed}}{The percentage of sulfur removed.}
}
}
\usage{
SulfurRemoval
}
\description{
Burning coal containing sulfur produces acid rain, so 
this experiment studies a process to remove sulfur from coal.
There are five factors, each at two levels: pH (1.5 or 2.5),
particle size (180 or 500 um), Fe^2+ (0 or 60 mmol),
pulp density (2 or 10\%), and leaching time (6 or 14 days).
The response of interest is percentage of sulfur removed.

Source: Oehlert (2022), originally from Golshani, et al (2013) via
Lye (2019).
}
\references{
Golshani, T., Jorjani, E., Chelgani S. Chehreh, Shafaei, S. Z., and Nafechi Y. Heidari (2013)
"Modeling and process optimization for microbial desulfurization of coal by using a 
two-level full factorial design," \emph{International Journal of Mining Science and Technology}
\strong{23} 261-265.

Lye, Leonard M. (2019) \emph{Applications of DEO in Engineering and Science:
A Collection of 2^6 Case Studies}, Leonard Lye, St. John's, Newfoundland.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
