% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Tensile}
\alias{Tensile}
\title{Tensile strength of composite materials}
\format{
A dataframe of 25 observations on two variables.
\describe{
\item{\code{batch}}{A factor indicating the batch of material.}
\item{\code{strength}}{The tensile strength (in coded units) of the sample.}
}
}
\usage{
Tensile
}
\description{
Composite materials used in the manufacture of aircraft components must be
tested to determine tensile strength. A manufacturer tests five random specimens
from five randomly selected batches. Results are "coded".

Source: Oehlert (2022), originally from Vangel (1992).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Vangel, Mark G. (1992) "New Methods for One-Sided Tolerance Limits for a One-Way Balanced
Random-Effects ANOVA Model." \emph{Technometrics} \strong{34}, 176-185.
}
\keyword{datasets}
