% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TextBooks}
\alias{TextBooks}
\title{Ratings of statistics textbooks.}
\format{
A dataframe with 7 observations on two variables.
\describe{
\item{\code{text}}{A factor with two levels giving the text used.}
\item{\code{pct.vge}}{The percentage of students giving very good or excellent ratings.}
}
}
\usage{
TextBooks
}
\description{
Seven sections of a statistics class are randomly assigned to two
potential textbooks, one the current text (Old) and the other a potential replacement (New).
The randomization will select four sections for the new text, with the other three getting
the old text.
At the end of the semester, students evaluate their text, and the response recorded here is
the percentage of the students who rate the text as "very good" or "excellent".

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
