% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Thermocouples}
\alias{Thermocouples}
\title{Thermocouple temperature differences}
\format{
A dataframe of 64 observations on one variable.
\describe{
\item{\code{temp.diff}}{The temperature difference in degrees C.}
}
}
\usage{
Thermocouples
}
\description{
Five thermocouples were placed in a furnace to measure relative temperature
at different locations in the furnace. This data set consists of the temperature difference
between thermocouples 3 and 5 for 64 consecutive readings.

Source: Oehlert (2022), originally from Christensen and Blackwood (1993).
}
\references{
Christensen, R. and L. G. Blackwood (1993). "Tests for precision and accuracy of multiple measuring
devices." \emph{Technometrics} \strong{35}, 411-420.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
