% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Tires}
\alias{Tires}
\title{Durability of tires.}
\format{
A dataframe with 30 observations on two variables.
\describe{
\item{\code{brand}}{A factor indicating the brand of tire.}
\item{\code{durability}}{The treadlife of the tire (in thousands of miles).}
}
}
\usage{
Tires
}
\description{
Five tire types (brand/model combinations like Goodyear/Arriva)
in the size 175/80R-13 are chosen at random from those available in a metropolitan
area, and six tires of each type are taken at random from warehouses.
The tires are placed (in random order) on a machine thaT will test tread
durability and report a response in thousands of miles.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
