% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TissueTearing}
\alias{TissueTearing}
\title{Weight needed to tear a tissue}
\format{
A dataframe with 96 observations on two variables.
\describe{
\item{\code{treatment}}{A factor with four levels giving the treatment used in the run. Levels are
A_dry, A_wet, B_dry, B_wet.}
\item{\code{tear.wt}}{The tearing weight in grams.}
}
}
\usage{
TissueTearing
}
\description{
Weight is added to a facial tissue stretched across an open jar top until the tissue tears.
There were 96 experimental runs with four conditions randomly assigned to the runs. The conditions are
brand A dry, brand A wet, brand B dry, and brand B wet.

Source: Oehlert (2022), originally from Joel Rumsch (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
