% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TissueVaporization}
\alias{TissueVaporization}
\title{Width of tissue vaporization under a laser scalpel}
\format{
A dataframe with 12 observations on three variables.
\describe{
\item{\code{width}}{The width in mm of the vaporized tissue.}
\item{\code{power.z}}{The power of the laser scalpel, in Watts.}
\item{\code{power}}{A factor version of the power of the laser scalpel, in Watts.
Included for convenience.}
}
}
\usage{
TissueVaporization
}
\description{
A laser scalpel will be drawn across rat liver at a rate of .01 inches/second.
Interesting responses include the width and depth of the tissue vaporized by the laser, as
well as the width and depth of the non-vaporized tissue that was damaged. This data set
includes the width of the vaporized tissue at four differ power levels of the laser.

Source: Oehlert (2022), adapted from original data from D. Deepa (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
