% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TowelStrength}
\alias{TowelStrength}
\title{Strength of wet paper towels}
\format{
A dataframe with 15 observations on two variables.
\describe{
\item{\code{brand}}{Factor indicating the brand of paper towel.}
\item{\code{pennies}}{The number of pennies when the towel breaks.}
}
}
\usage{
TowelStrength
}
\description{
Everyone likes a strong, durable paper towel (absorbency is another desirable
property, but this experiment is about durability). Five rolls of paper towels
are purchased, one each from five brands (B, V, S, S2, and T). The first
three brands are name brands, and the last two are store brands. From each
roll, we randomly select three towels. In random order, each towel is
dipped in water until wet, gently squeezed by hand to remove absorbed water,
then spread across the top of a bowl and held in place by clothespins. Then,
pennies are gently placed on the suspended towel until the towel breaks.
The number of pennies until the towel is the measure of strength.

Source: Oehlert (2022), original data from A. Frosch (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
