% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TrampledPlants}
\alias{TrampledPlants}
\title{Grass recovery after trampling}
\format{
A dataframe with 20 observations on three variables.
\describe{
\item{\code{passes.z}}{The number of passes for this lane.}
\item{\code{passes}}{A factor version of the number of passes.}
\item{\code{height}}{The height of the vegetation one year later.}
}
}
\usage{
TrampledPlants
}
\description{
Twenty "lanes" in an alpine meadow will be trampled by having someone in
lug soled boots walk repeatedly over them in a natural gait.  The lanes were randomly assigned to
one of five different treatments: 0, 25, 75, 200, or 500 passes by the trampler.
	The response is the height (in cm)
of the vegetation one year after trampling.

Source: Oehlert (2022), originally from Cole (1993).
}
\references{
Cole, D. N. (1993). "Trampling effects on mountain vegetation in Washington, Colorado,
New Hampshire, and North Carolina." Research Paper INT-464, U.S. Forest Service,
Intermountain Research Station.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
