% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TropicalGrasses}
\alias{TropicalGrasses}
\title{Tropical grass production}
\format{
A dataframe with 48 observations on seven variables.
\describe{
\item{\code{height.z}}{The height in inches.}
\item{\code{height}}{A factor version of the height.}
\item{\code{fert.z}}{Fertilizer in cwt/acre/year.}
\item{\code{fert}}{A factor version of fertilizer.}
\item{\code{interval.z}}{Cutting interval in weeks.}
\item{\code{interval}}{A factor version of interval.}
\item{\code{yield}}{The yield as cwt/acre.}
}
}
\usage{
TropicalGrasses
}
\description{
A study looked into the management of various tropical grasses
for improved production, measured as dry matter yield in hundreds of pounds
per acre over a 54-week study period.  The management variables were height
of cut (1, 3, or 6 inches), the cutting interval (1, 3, 6, or 9 weeks), and
amount of nitrogen fertilizer (0, 8, 16, or 32 hundred pounds of ammonium
sulfate per acre per year).   Forty-eight plots were assigned in completely
randomized fashion to the 48 factor-level combinations.

Source: Oehlert (2022), originally from Richards (1965).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Richards, J. A. (1965) "Effects of fertilizers and management on the three
promising tropical grasses in Jamaica" \emph{Experimental Agriculture} \strong{1}, 281-288.
}
\keyword{datasets}
