% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Tubules}
\alias{Tubules}
\title{Hormones affecting seminiferous tubules}
\format{
A dataframe of 24 observations on four variables.
\describe{
\item{\code{litter}}{A factor indicating the litter of the boar.}
\item{\code{GH}}{A factor indicating whether growth hormone was used.}
\item{\code{FSH}}{A factor indicating whether follicle stimulating hormone was applied.}
\item{\code{length}}{The length of the seminiferous tubules, in meters.}
}
}
\usage{
Tubules
}
\description{
Hormones can alter the sexual development of animals.
This experiment studies the effects of growth hormone
(GH) and follicle-stimulating hormone (FSH) on the
length of the seminiferous tubules in pigs.  The treatments
are control, daily injection of GH, daily injection of
FSH, and daily injection of GH and FSH.  Twenty-four
weanling boars are used, four from each of six litters.
The four boars in each litter are randomized to the
four treatments.  The boars are castrated at 100 days
of age, and the length (in meters!) of the seminiferous tubules
determined as response

Source: Oehlert (2022), originally from Swanlund, et al. (1995).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Swanlund, D. J., M. R. N'Diaye, K. L. Loseth, J. L. Pryor, and B. G. Crabo (1995)
"Diverse testicular responses to exogenous growth hormone and follicle-stimulating
hormone in prepubertal boars," \emph{Biology of Reproduction} \strong{53}, 749-757.
}
\keyword{datasets}
