% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TurkeyPoults}
\alias{TurkeyPoults}
\title{Turkey poult weight gain}
\format{
A dataframe of 71 observations on three variables.
\describe{
\item{\code{group}}{A factor indicating the group of pens (1-9, or 10-18, etc.).}
\item{\code{treatment}}{A factor indicating (as 1-9) the treatment given to the pen.}
\item{\code{gain}}{Average weight gain for poults in the pen, g/day.}
}
}
\usage{
TurkeyPoults
}
\description{
Seventy-two pens of eight male turkey poults (chicks) each were set out.
Pens 1-9 were randomly assigned to nine treatments. Treatment 1 is
a control treatment; treatments 2 through 9 contain alfalfa meal.
Treatments 2 through 5 contain alfalfa meal type 22; treatments
6 through 9 contain alfalfa meal type 27.  Treatments 2 and 6 are
2.5\% alfalfa, treatments 3 and 7 are 5\% alfalfa, treatments 4
and 8 are 7.5\% alfalfa.  Treatments 5 and 9 are also 7.5\%
alfalfa, but they have been modified to have the same calories
as the control treatment. After this, pens 10-18 were randomly assigned,
19-27 were randomly assigned, and so on.

The response for each pen is the average weight gain (g/day) for the poults
in the pen over ages 7-14 days. The response for pen 66 was not observed.

Source: Oehlert (2022), data from T. Ergul (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
