% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{TwoSeriesDataA}
\alias{TwoSeriesDataA}
\title{Synthetic single replication of a 2^4 experiment}
\format{
A dataframe of 16 observations on five variables.
\describe{
\item{\code{A}}{Factor 1.}
\item{\code{B}}{Factor 2.}
\item{\code{C}}{Factor 3.}
\item{\code{D}}{Factor 4.}
\item{\code{y}}{The response.}
}
}
\usage{
TwoSeriesDataA
}
\description{
Synthetic data representing the results of a single replication of a
2^4 experiment.

Source: Oehlert (2022).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
