% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2series.R
\name{TwoSeriesPlots}
\alias{TwoSeriesPlots}
\title{Half-Normal or Pareto Effects Plots}
\usage{
TwoSeriesPlots(
  fit,
  alpha = 0.05,
  data = NULL,
  B = NULL,
  type = c("half", "pareto")[1],
  block = NULL,
  pse = FALSE,
  recode = FALSE,
  signif.only = TRUE,
  term.labels = NULL,
  cex.fac = par("cex.lab"),
  cex.lab = par("cex.lab"),
  cex.pch = par("cex"),
  cex.legend = par("cex.lab"),
  main = NULL,
  subtitle = NULL,
  ...
)
}
\arguments{
\item{fit}{Either a linear model formula or an object of class lm for a two-series model.}

\item{alpha}{The significance level for testing effects. This is a  strong familywise error rate for
the Basso-Salmaso permutation and a per comparison error rate for the Lenth test.}

\item{data}{An optional data frame.}

\item{B}{The number of random permutations to use for testing effects using the Basso-Salmaso test.}

\item{type}{Either "half" (default, for half-normal) or "pareto", indicating the type of plot to make.}

\item{block}{NULL or integer vector. If non-null, the terms corresponding to the entries in \code{block} will be labeled "Block".}

\item{pse}{Logical flag indicating whether effects significant according to the Lenth PSE approach should also be labeled.}

\item{recode}{Logical flag (default FALSE) indicating whether factor names should be replaced with A, B, C, etc.}

\item{signif.only}{Logical flag. If TRUE (the default), then labels are only shown for significant effects.}

\item{term.labels}{NULL or list. If NULL, point labels are automatically determined according to the setting of recode (i.e. A,B,C etc. for recode=TRUE and
natural effect names otherwise) and signif.only (i.e. all effects are labelled if signif.only=FALSE, only significant effects are labelled if
signif.only=TRUE). Otherwise, term.labels can be used for manual labelling of certain effects and should be a list with components index
 (integer vector referring to position of terms to be labelled once pivoted terms are eliminated) and label (character vector of labels) components.}

\item{cex.fac}{numeric. Factor label character size.}

\item{cex.lab}{numeric. Labels character size.}

\item{cex.pch}{numeric. Points character size.}

\item{cex.legend}{numeric. Legend size in case of recodes.}

\item{main}{NULL or character. Title of plot. If NULL, automatic title is generated.}

\item{subtitle}{NULL or character. Sub title of plot.}

\item{...}{Additional graphical parameters to be passed to plot.}
}
\description{
This function is a modification of FrF2::DanielPlot, which is itself a modification
of BsMD::DanielPlot. The changes implemented here include:
\itemize{
\item Added Pareto plots.
\item Significant effects are selected using the Basso-Salmaso permutation test.
\item Removed (full) normal plot of effects.
\item Data are always on the x-axis.
\item Optional selection of significant effects using the Lenth PSE (with marking for individual and simultaneous margin of error).
\item Plotting characters are set and not optional.
 \item More error checking.
}
}
\details{
The model must be a saturated linear model for a (fractional) factorial 2-level design. Completely aliased effects are allowed, but
 partial aliasing is not allowed. Effects (except for the intercept) are
displayed in a Pareto or half-normal plot with the effects on the x-axis.

Significant effects according to the Basso-Salmaso permutation test (strong familywise error rate) are shown with a solid dot, while other effects
are shown with an open dot. If \code{signif.only} is true, only the significant effects are labeled. If
\code{pse} is also true, any additional effects that are significant using the (per comparison error rate) Lenth
procedure will also be labeled, but they will not be given the solid dot. This will also show
per comparison and simultaneous margin of error for the Lenth pse.

 B will be set to a reasonable value if unspecified.
}
\examples{
 \dontrun{data(ImageID)}
 \dontrun{TwoSeriesPlots(fracID~A*B*C*D*E*F*G,data=ImageID,block=block)}
 \dontrun{TwoSeriesPlots(fracID~A*B*C*D*E*F*G,data=ImageID,
      term.labels=list(index=c(5,127),label=c("I'm term E","I'm a block")))}

}
\references{
Basso, D. and L. Salmaso (2006). "A Discussion of Permutation Tests Conditional to Observed Responses in
Unreplicated 2^m Full Factorial Designs," \emph{Communications in Statistics - Theory and Methods} \strong{35}, 83-97.

Lenth, R.V. (1989) "Quick and easy analysis of unreplicated factorials," \emph{Technometrics} \strong{31}, 469-473.
}
\seealso{
\code{\link[FrF2:DanielPlot]{FrF2::DanielPlot}}
}
\author{
Ernesto Barrios, modified by Ulrike Groemping, further modified by Gary W. Oehlert (\email{gary@umn.edu}). Code
for the permutation test by Gary W. Oehlert.
}
