% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{UnpoppedKernels}
\alias{UnpoppedKernels}
\title{Unpopped popcorn kernels}
\format{
A dataframe of 36 observations on four variables.
\describe{
\item{\code{brand}}{A factor indicating the brand of popcorn used.).}
\item{\code{watts.z}}{The wattage of the oven used.}
\item{\code{watts}}{A factor version of watts.}
\item{\code{unpopped}}{The count of unpopped kernels.}
}
}
\usage{
UnpoppedKernels
}
\description{
Thirty-six 3 oz. bags of microwave popcorn are popped, and the number
of unpopped kernels in each bag is recorded. The 36 runs are randomly assigned to
the twelve combinations of microwave wattage (500, 700, or 1000) and
brand ("P", "A", "J", or "O"). Ovens are allowed to cool for 10 minutes with
the door open between runs, and popping continues until there is a three
second gap between consecutive pops.

Source: Oehlert (2022), data from P. Stenberg (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
