% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Verapamil}
\alias{Verapamil}
\title{ACTH release under Verapamil treatment}
\format{
A dataframe of 36 observations on seven variables.
\describe{
\item{\code{verapamil.z}}{The amount of Verapamil in the treatment.}
\item{\code{verapamil}}{A factor version of \code{verapamil.z}.}
\item{\code{crf.z}}{The amount of CRF in the treatment.}
\item{\code{crf}}{A factor version of \code{crf.z}.}
\item{\code{calcium.z}}{The amount of calcium in the treatment.}
\item{\code{calcium}}{A factor version of \code{calcium.z}.}
\item{\code{ACTH}}{The amount of ACTH released under treatment.}
}
}
\usage{
Verapamil
}
\description{
An experiment investigated the release of the hormone ACTH from
rat pituitary glands under eight treatments:  the
factorial combinations of CRF (0 or 100 nM; CRF is believed to increase
ACTH release), calcium (0 or 2 mM of CaCl2), and Verapamil (0 or 50 uM;
Verapamil is thought to block the effect of calcium).  Thirty-six rat
pituitary cell cultures are assigned at random to the factor-level combinations,
with control (all treatments 0) getting 8 units, and other combinations
getting 4.

Source: Oehlert (2022), originally from Giguere, Lefevre, and Labrie (1982).
}
\references{
Giguere, V., G. Lefevre, and F. Labrie (1982) "Site of calcium requirement
for stimulation of ACTH release in rat anterior pituitary cells in culture by
synthetic ovine corticotropin-releasing factor," \emph{Life Sciences} \strong{31},
3057-3062.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
