% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{VisiplumePlant2}
\alias{VisiplumePlant2}
\title{Ratio of Visiplume readings to standard method}
\format{
A dataframe of 6 observations on one variable.
\describe{
\item{\code{ratio}}{The ratio of the Visiplume reading to the standard measurement.}
}
}
\usage{
VisiplumePlant2
}
\description{
A "Visiplume" is a remote sensing instrument that measures SO2 concentration
	by comparing ultraviolet absorption in clear air and polluted air. This data set reports
the ratio of the Visiplume reading to an EPA standard method. Measurements are taken on
multiple samples on four different sources: bag (SO2 in inflated bag), smoke (spiked
smoke from a smoke generator, two coal-burning power plants (plant1 and plant2). This
data frame contains data only on plant 2. Data set "Visiplume" contains the full data.

Source: Oehlert (2022), originally from McElhoe and Conner (1986).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

McElhoe, H. B. and Conner, W. D. (1986) "Remote measurement of sulfur dioxide emissions
using an ultraviolet light sensitive video system," \emph{Journal of the Air Pollution
Control Association} \strong{36}, 42-47.
}
\keyword{datasets}
