% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Wallbangers}
\alias{Wallbangers}
\title{Rating of Harvey Wallbanger recipes.}
\format{
A dataframe of seven observations on four variables.
\describe{
\item{\code{G}}{The fraction of Galliano.}
\item{\code{V}}{The fraction of vodka.}
\item{\code{O}}{The fraction of orange juice.}
\item{\code{appeal}}{The rating for the drink.}
}
}
\usage{
Wallbangers
}
\description{
An experiment to find the best proportions for orange juice (O),
vodka (V), and Galliano (G) in a mixed drink called a Harvey Wallbanger.  This
is a factorial ratios model, with three levels of the
ratio V/G (1.2, 2.0, and 2.8) and two levels of the ratio O/G (4 and
9).  There is also  a center point at V/G = 2 and O/G = 6.5.  The actual
design included incomplete blocks (so that no evaluator
consumed more than a small number of drinks). However, there were no apparent
evaluator differences, so the average score was used as
response for each mixture, and blocks were ignored.
Evaluators rated the drinks on a 1 to 7 scale.

Source: Oehlert (2022), originally from Sahrmann, Peipel, and Cornell (1987).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Sahrmann, H. F., G. F. Piepel, and J. A. Cornell (1987). "In search of the optimum Harvey
Wallbanger recipe via mixture experiment techniques," \emph{The American Statistician} \strong{41}, 190-194.
}
\keyword{datasets}
