% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WaterBinding}
\alias{WaterBinding}
\title{Gums and water binding in gels}
\format{
A dataframe of 214 observations on seven variables.
\describe{
\item{\code{solute}}{The solute in the gel.}
\item{\code{sol.conc}}{A factor giving the concentration of the solute as "low", "mid", "high".}
\item{\code{sol.conc.z}}{The concentration of the solute.}
\item{\code{gum}}{The gum in the gel.}
\item{\code{gum.conc}}{A factor giving the concentration of the gum as "low", "mid", "high".}
\item{\code{gum.conc.z}}{The concentration of the solute.}
\item{\code{binding}}{The water binding of the experimental run.}
}
}
\usage{
WaterBinding
}
\description{
Gums are used to alter the texture and other properties
of foods, in part by binding water.  An experiment studied the
water-binding of various carrageenan gums in gel systems
under various conditions.  The experiment had factorial
treatment structure with four factors.  Factor 1 was the
type of gum (kappa, mostly kappa with some lambda, and
iota).  Factor 2 was the concentration of the gum in the
gel in g/100g water (level 1 is .1; level 2 is .5; and level 3
is 2 for gums 1 and 2, and 1 for gum 3).
The third factor was type of solute (NaCl,
Na2SO4, sucrose).
The fourth factor was solute concentration (ku/kg water).
For sucrose, the three levels were .05, .1, and .25; for
NaCl and Na2SO4, the levels were .1, .25, and 1.
The response is the water-binding for the gel in mOsm.
This experiment was completely randomized.  There were
two units at each factor-level combination except
solute concentration 3, where
all but one combination had four units.

Source: Oehlert (2022), originally from Rey (1981).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Rey, D. K. (1981) \emph{Characterization of the Effect of Solutes on the Water-Binding
and Gel Strength Properties of Carrageenan.} Ph. D. Thesis, University of Minnesota,
St. Paul, MN.
}
\keyword{datasets}
