% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WaterEjection}
\alias{WaterEjection}
\title{Distance of ejection stream.}
\format{
A dataframe of 30 observations on three variables.
\describe{
\item{\code{depth.z}}{The depth of the water above the hole in cm.}
\item{\code{depth}}{A factor version of depth.}
\item{\code{distance}}{The distance the water is ejected in cm.}
}
}
\usage{
WaterEjection
}
\description{
We puncture a hole in the side of a (empty) 2 liter plastic soda bottle. An
experimental run will consist of filling the bottle with the hole covered and then uncovering
the hole to measure how far the water is ejected (in cm). We fill the bottle 30 times,
randomizing the runs 10 depths of water above the hole (6, 7, 8, ... 15 cm).

Source: Oehlert (2022), data from X. Meng (pers. comm.).
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
