% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WaxDeposit}
\alias{WaxDeposit}
\title{Wax deposited from Malaysian crude oil}
\format{
A data frame of 16 observations on five variables.
\describe{
\item{\code{speed}}{A factor giving the speed of rotation.}
\item{\code{temp}}{A factor giving the cold finger temperature.}
\item{\code{duration}}{A factor giving the duration of the run.}
\item{\code{inhibitor}}{A factor giving the concentration of inhibitor.}
\item{\code{wax}}{The amount of wax deposited.}
}
}
\usage{
WaxDeposit
}
\description{
This experiment considers how four factors affect
the deposition of wax from Malaysian crude oil. The factors are
speed of rotation (0 or 600 rpm), cold finger temperature (5 or 15 C),
duration (2 or 24 h), and inhibitor concentration (200 or 5000 ppm).
The 16 factor/level combinations are run in random order, and the
response is the amount of wax deposited (in g).

Note: the original data set included three center points with responses
1.40, 1.50, 1.45.

Source: Oehlert (2022), originally from Ridzuan et al. (2016) via
Lye (2019).
}
\references{
Lye, Leonard M. (2019) \emph{Applications of DEO in Engineering and Science:
A Collection of 2^6 Case Studies}, Leonard Lye, St. John's, Newfoundland.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Ridzuan, N., F. Adam, and Z. Yaacob (2016) "Screening of factor influencing wax
deposition using full factorial experimental design," \emph{Petroleum Science and Technology}
\strong{34} 84-90.
}
\keyword{datasets}
