% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WeedBiomass}
\alias{WeedBiomass}
\title{Weed biomass}
\format{
A dataframe with 48 observations on seven variables.
\describe{
\item{\code{table}}{A factor indicating the table (1-2).}
\item{\code{tray}}{A factor indicating the tray (1-8).}
\item{\code{wetland}}{A factor indicating the wetland (1-24).}
\item{\code{nitrogen}}{A factor indicating the level of nitrogen.}
\item{\code{weed}}{A factor indicating the weed species addition.}
\item{\code{clip}}{A factor indicating the clipping treatment.}
\item{\code{biomass}}{The percentage of the biomass that is not weeds.}
}
}
\usage{
WeedBiomass
}
\description{
This experiment studies the effect of nitrogen and weed addition on
plant growth in wetlands.  We investigate four levels of
nitrogen, three weed treatments (no additional weeds,
addition of weed species 1, addition of weed species 2),
and two herbivory treatments (clipping and no clipping).
We have eight trays; each tray holds three artificial
wetlands consisting of rectangular wire baskets
containing wetland soil.
The trays are full of water, so the artificial wetlands stay wet.
All of the artificial wetlands receive a standard set of
seeds to start growth.

Four of the trays are placed on a table near the door of the
greenhouse, and the other four trays are placed on a table
in the center of the greenhouse.  On each table, we randomly
assign one of the trays to each of the four nitrogen
treatments.  Within
each tray, we randomly assign the wetlands to the three weed
treatments.  Each wetland is split in half.  One half is
chosen at random and will be clipped after 4 weeks,
with the clippings removed; the other half is not
clipped.   After 8 weeks, we measure the fraction of
biomass in each wetland that is nonweed as our response.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
