% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WeedControl}
\alias{WeedControl}
\title{Weed control}
\format{
A dataframe with 56 observations on two variables.
\describe{
\item{\code{treatment}}{A factor indicating the weed treatment (1-14).}
\item{\code{pct.control}}{The percentage of the plot that is not weeds.}
}
}
\usage{
WeedControl
}
\description{
Fourteen weed control treatments were randomly applied to 56 plots, four
plots per treatment. The plots are later assessed visually for weeds, with the response
being the percent of the plot that is non-weeds.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
