% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{WeightGain}
\alias{WeightGain}
\title{Weight gain of test animals under different diets}
\format{
A dataframe with 9 observations on three variables.
\describe{
\item{\code{source}}{A factor indicating the protein source.}
\item{\code{calories}}{A factor indicating the level of calories.}
\item{\code{gain}}{Average weight gain for the two animals in the treatment group.}
}
}
\usage{
WeightGain
}
\description{
Eighteen test animals were randomlly assigned to nine diets, two animals per diet.
The diets are the combinations of protein source (beef, pork, grain) and calories (low, medium,
high). The error mean square for these data was 8.75.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
