% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Welding}
\alias{Welding}
\title{Strength of welds}
\format{
A dataframe with 16 observations on ten variables.
\describe{
\item{\code{A}}{Treatment factor 1.}
\item{\code{B}}{Treatment factor 2.}
\item{\code{C}}{Treatment factor 3.}
\item{\code{D}}{Treatment factor 4.}
\item{\code{E}}{Treatment factor 5.}
\item{\code{F}}{Treatment factor 6.}
\item{\code{G}}{Treatment factor 7.}
\item{\code{H}}{Treatment factor 8.}
\item{\code{J}}{Treatment factor 9.}
\item{\code{strength}}{The strength of the weld.}
}
}
\usage{
Welding
}
\description{
An experiment to determine factors affecting the strength of welds.
There were nine factors at two levels each to be explored.
The full experiment was much too large, so a 2^(9-5) fractional
factorial with sixteen units was used.  The factors are coded A though J (skipping I);
the generators are -ACE, -ADF, -ACDG, BCDH, ABCDJ.

Source: Oehlert (2022), originally from Taguchi and Wu (1980)
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}

Taguchi, G. and Y. Wu (1980). \emph{Introduction to Off-Line Quality Engineering}. Nagoya, Japan:
Central Japan Quality Control Association.
}
\keyword{datasets}
