% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Yogurt}
\alias{Yogurt}
\title{Consumer scoring of yogurts}
\format{
A dataframe with 52 observations on four variables.
\describe{
\item{\code{C}}{A factor indicating cost.}
\item{\code{S}}{A factor indicating the sensory quality.}
\item{\code{N}}{A factor indicating the nutritional value.}
\item{\code{gain}}{Score for likelihood to buy.}
}
}
\usage{
Yogurt
}
\description{
Consumers who do not regularly eat yogurt are polled
and asked to rate on a 1 to 9 scale the likelihood that they would
buy a certain yogurt product at least once a month;  1 means very
unlikely, 9 means very likely.  The product is hypothetical and
described by three factors:  cost ("C" - low, medium, and high), sensory
quality ("S" - low, medium, and high), and nutritional value ("N" -low and high).
The plan was to poll three consumers for each product type, but it
became clear early in the experiment that people were unlikely to buy
a high-cost, low-nutrition, low-quality product, so only one consumer
was polled for that combination.  Each consumer received one of the
eighteen product descriptions chosen at random.

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
