% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{YogurtCooler}
\alias{YogurtCooler}
\title{Shelf life of yogurts}
\format{
A dataframe with 52 observations on four variables.
\describe{
\item{\code{cooler}}{A factor indicating the cooler in use.}
\item{\code{A}}{A factor indicating the level of sugar.}
\item{\code{B}}{A factor indicating the culture strain.}
\item{\code{C}}{A factor indicating the type of fruit.}
\item{\code{D}}{A factor indicating the pH.}
\item{\code{days}}{Days until off odor is noticed.}
}
}
\usage{
YogurtCooler
}
\description{
Scientists wish to understand how
the amount of sugar (two levels), culture strain (two levels),
type of fruit (blueberry or strawberry), and pH (two levels)
influence shelf life of refrigerated yogurt.
In a preliminary experiment, they produce one batch
of each of the sixteen kinds of yogurt.  The yogurt is then
placed in two coolers, eight batches in each cooler.
The response is the number of days till
an off odor is detected from the batch

Source: Oehlert (2022), synthetic data.
}
\references{
Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
