% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ZincRetention}
\alias{ZincRetention}
\title{Zinc retention in rats}
\format{
A dataframe with 16 observations on four variables.
\describe{
\item{\code{m.zinc}}{A factor indicating the level of zinc in the final meal.}
\item{\code{d.zinc}}{A factor indicating the level of zinc in the usual diet.}
\item{\code{m.protein}}{A factor indicating the level of protein in the usual diet.}
\item{\code{retention}}{Zinc retained in the rat.}
}
}
\usage{
ZincRetention
}
\description{
The treatments were the combinations of two levels
of zinc in the usual diet, two levels of zinc in the final meal
(before the rats are sacrificed), and four levels of protein in the final meal.

Source: Oehlert (2022), originally from Hunt and Larson (1990).
}
\references{
Hunt, J. R. and B. J. Larson (1990). "Meal protein and zinc levels interact to
influence zinc retention by the rat," \emph{Nutrition Research} \strong{10}, 697-705.

Oehlert, Gary W. (2022) \emph{A First Course in Design and Analysis of Experiments,
Second Edition}, \url{http://users.stat.umn.edu/~gary/Book.html}
}
\keyword{datasets}
